Apply blocked weighted bootstrap to indicator datasets

A wrapper for bootBW function in bbw package.

boot_estimate(
  indicator = c("ifaDF", "iycfDF", "mnpDF", "vitDF", "screenDF", "anthroDF"),
  county,
  w,
  replicates = 399,
  core.columns = c("spid", "cid", "did", "eid", "motherID", "m2")
)

Arguments

indicator
A character vector of indicator data.frame names

county
An integer indicating which county to interpolate; 1 for Greater Monrovia; 2 for Grand Bassa

w
Population dataset

replicates
Number of bootstrap replicates

core.columns
A vector of variable names included in indicator data.frames

Value

A data.frame of indicator estimates with lower and upper confidence limits

Examples

boot_estimate(indicator = "iycfDF", county = 1, w = psuDataGM, replicates = 9)

#> $iycfDF
#>    varLabel    estimate         lcl        ucl
#> 1      icf1 0.895348837 0.871197572 0.91146474
#> 2      icf2 0.869772999 0.840393904 0.89719052
#> 3     icf2a 0.168067227 0.135238095 0.29033101
#> 4     icf2b 0.604166667 0.462020906 0.73809015
#> 5     icf2c 0.000000000 0.000000000 0.00000000
#> 6     icf2d 0.073170732 0.032792208 0.15061267
#> 7     icf2e 0.064220183 0.001680672 0.11777778
#> 8     icf2f 0.031250000 0.000000000 0.07840083
#> 9     icf2g 0.017857143 0.000000000 0.08071157
#> 10    icf2h 0.010416667 0.000000000 0.03490715
#> 11    icf2i 0.000000000 0.000000000 0.04667799
#> 12    icf2j 0.000000000 0.000000000 0.00000000
#> 13    icf2k 0.000000000 0.000000000 0.00000000
#> 14    icf2l 0.000000000 0.000000000 0.00000000
#> 15    icf2m 0.000000000 0.000000000 0.00000000
#> 16    icf2n 0.008928571 0.000000000 0.04942704
#> 17     icf3 5.363103953 5.089308587 5.67848437
#>