Function to recode age variable and convert to age in days

Function to recode age variable and convert to age in days

recode_age(df)

Arguments

df
A coverage data.frame collected for the Liberia Coverage Survey

Value

A vector of recoded ages based on any re-testing done

Examples

recode_age(df = coverageData.r2)

#>    [1]  700.0625  365.2500  608.7500  426.1250 1461.0000  700.0625  213.0625
#>    [8] 1552.3125 1582.7500 1156.6250 1430.5625  456.5625  760.9375 1095.7500
#>   [15] 1521.8750 1095.7500  304.3750  304.3750  182.6250  426.1250  426.1250
#>   [22]  974.0000 1491.4375  852.2500  213.0625 1461.0000 1095.7500 1095.7500
#>   [29]  760.9375  547.8750  334.8125  182.6250 1126.1875 1461.0000 1247.9375
#>   [36] 1247.9375  487.0000  182.6250 1004.4375  669.6250  547.8750  395.6875
#>   [43] 1126.1875  456.5625  669.6250 1613.1875  943.5625 1461.0000  456.5625
#>   [50]  608.7500  882.6875  882.6875 1187.0625 1491.4375  456.5625  669.6250
#>   [57]  182.6250  487.0000 1004.4375 1734.9375  304.3750 1765.3750  182.6250
#>   [64] 1461.0000 1187.0625 1521.8750  426.1250  517.4375  304.3750 1065.3125
#>   [71] 1217.5000 1278.3750  882.6875  760.9375 1095.7500  730.5000  791.3750
#>   [78] 1034.8750 1369.6875 1034.8750  334.8125 1461.0000 1491.4375  456.5625
#>   [85] 1734.9375  639.1875  821.8125 1126.1875 1308.8125  821.8125  334.8125
#>   [92]  365.2500  213.0625 1247.9375 1034.8750  547.8750 1430.5625 1278.3750
#>   [99]  821.8125  273.9375  547.8750  608.7500  791.3750  456.5625 1278.3750
#>  [106]  487.0000 1734.9375  821.8125  608.7500 1034.8750  730.5000  608.7500
#>  [113] 1308.8125  852.2500  791.3750  395.6875 1187.0625  426.1250 1491.4375
#>  [120]  700.0625 1765.3750  669.6250  456.5625 1034.8750  456.5625 1004.4375
#>  [127]        NA  578.3125 1126.1875 1704.5000  578.3125  700.0625  730.5000
#>  [134]  821.8125  578.3125  517.4375 1095.7500  426.1250 1004.4375 1247.9375
#>  [141]  395.6875  639.1875  760.9375  669.6250  426.1250  974.0000 1491.4375
#>  [148]  700.0625 1369.6875 1156.6250  791.3750 1613.1875 1461.0000 1095.7500
#>  [155] 1765.3750 1217.5000 1126.1875  852.2500 1187.0625  273.9375  517.4375
#>  [162] 1643.6250  456.5625  913.1250  334.8125 1765.3750  730.5000 1643.6250
#>  [169]  426.1250 1095.7500  730.5000  304.3750  608.7500 1765.3750  456.5625
#>  [176]  182.6250  669.6250  456.5625 1217.5000  213.0625 1247.9375 1369.6875
#>  [183] 1461.0000 1521.8750 1491.4375  730.5000  182.6250 1552.3125  639.1875
#>  [190]  760.9375 1613.1875  487.0000  821.8125  456.5625  913.1250 1552.3125
#>  [197]  669.6250 1156.6250  821.8125  547.8750  882.6875  395.6875 1065.3125
#>  [204]  334.8125 1308.8125 1247.9375 1126.1875  243.5000 1095.7500 1461.0000
#>  [211] 1674.0625  334.8125 1430.5625 1430.5625  213.0625  943.5625 1247.9375
#>  [218]  273.9375 1491.4375  456.5625 1156.6250  395.6875 1156.6250 1156.6250
#>  [225]  608.7500 1552.3125  821.8125  669.6250  547.8750 1339.2500 1004.4375
#>  [232]        NA  243.5000  791.3750 1065.3125  426.1250  821.8125 1156.6250
#>  [239] 1461.0000  487.0000  243.5000 1187.0625 1187.0625  182.6250 1613.1875
#>  [246] 1095.7500  182.6250 1126.1875 1308.8125  669.6250  669.6250 1034.8750
#>  [253] 1126.1875  213.0625 1461.0000  304.3750  821.8125 1795.8125  365.2500
#>  [260] 1034.8750  426.1250  456.5625 1400.1250 1552.3125  547.8750 1765.3750
#>  [267]  243.5000  456.5625  426.1250  760.9375  213.0625 1369.6875 1521.8750
#>  [274]  791.3750  639.1875  517.4375 1095.7500 1704.5000 1734.9375 1095.7500
#>  [281] 1369.6875 1552.3125  730.5000 1491.4375  487.0000 1156.6250 1187.0625
#>  [288] 1795.8125  243.5000 1552.3125 1156.6250 1400.1250  243.5000 1339.2500
#>  [295]  273.9375  791.3750  243.5000 1339.2500  395.6875 1461.0000  730.5000
#>  [302]  273.9375  852.2500  273.9375 1095.7500 1247.9375  913.1250  852.2500
#>  [309] 1582.7500  456.5625  334.8125 1734.9375 1521.8750 1552.3125  243.5000
#>  [316]  608.7500 1369.6875  547.8750  669.6250 1582.7500 1461.0000 1247.9375
#>  [323] 1643.6250  334.8125  517.4375  456.5625  334.8125  882.6875  974.0000
#>  [330]  974.0000  578.3125 1582.7500 1369.6875 1156.6250  608.7500 1217.5000
#>  [337]  487.0000 1491.4375 1613.1875  304.3750 1126.1875 1034.8750 1308.8125
#>  [344]  334.8125  487.0000  273.9375  639.1875  426.1250  882.6875 1004.4375
#>  [351]  395.6875 1217.5000  243.5000  243.5000  243.5000 1095.7500 1491.4375
#>  [358]  669.6250  882.6875  821.8125 1582.7500 1552.3125  760.9375 1613.1875
#>  [365] 1187.0625 1613.1875 1795.8125  791.3750 1369.6875 1521.8750  669.6250
#>  [372]  791.3750 1126.1875 1552.3125 1339.2500 1278.3750 1552.3125 1156.6250
#>  [379]  608.7500 1734.9375  943.5625 1582.7500        NA  821.8125 1582.7500
#>  [386]  730.5000  273.9375 1126.1875 1247.9375  456.5625  395.6875  821.8125
#>  [393] 1217.5000  365.2500 1339.2500 1034.8750 1674.0625 1217.5000  669.6250
#>  [400] 1095.7500 1461.0000 1674.0625  395.6875  456.5625 1521.8750  243.5000
#>  [407] 1339.2500 1095.7500  700.0625 1187.0625 1613.1875  852.2500  456.5625
#>  [414] 1400.1250  426.1250 1521.8750  821.8125  487.0000 1430.5625 1095.7500
#>  [421] 1339.2500  578.3125 1065.3125  821.8125 1521.8750  852.2500 1095.7500
#>  [428]  852.2500  487.0000  791.3750  852.2500  334.8125  852.2500  426.1250
#>  [435] 1552.3125 1217.5000 1369.6875  547.8750 1430.5625  395.6875  913.1250
#>  [442] 1065.3125 1095.7500  669.6250 1795.8125  882.6875 1156.6250 1278.3750
#>  [449]  791.3750  365.2500 1704.5000  395.6875 1126.1875  243.5000 1004.4375
#>  [456]  639.1875 1461.0000 1643.6250  913.1250  608.7500  821.8125 1217.5000
#>  [463] 1704.5000 1187.0625 1704.5000  304.3750  243.5000  730.5000  700.0625
#>  [470] 1795.8125 1065.3125 1308.8125  608.7500  395.6875  608.7500 1339.2500
#>  [477]  913.1250 1643.6250 1126.1875 1004.4375 1734.9375  517.4375 1126.1875
#>  [484]  852.2500  791.3750  974.0000  517.4375  426.1250  760.9375  913.1250
#>  [491] 1582.7500 1552.3125 1095.7500 1095.7500 1582.7500  730.5000  639.1875
#>  [498] 1247.9375  639.1875 1582.7500 1034.8750  487.0000 1734.9375  821.8125
#>  [505] 1126.1875  547.8750  760.9375  456.5625  182.6250  395.6875 1187.0625
#>  [512]  304.3750 1156.6250 1430.5625 1643.6250  821.8125  730.5000 1034.8750
#>  [519]  882.6875 1461.0000  517.4375  182.6250 1430.5625  213.0625  456.5625
#>  [526]  273.9375 1339.2500  821.8125  974.0000 1430.5625  517.4375 1187.0625
#>  [533]  913.1250 1217.5000 1582.7500  608.7500 1369.6875 1095.7500 1034.8750
#>  [540]  304.3750 1704.5000 1491.4375 1552.3125 1278.3750  182.6250  913.1250
#>  [547] 1187.0625  821.8125  243.5000 1521.8750        NA  913.1250 1095.7500
#>  [554]  365.2500 1674.0625 1095.7500  730.5000 1126.1875  213.0625  913.1250
#>  [561]  882.6875 1308.8125 1156.6250  213.0625 1734.9375  821.8125 1004.4375
#>  [568] 1430.5625        NA  365.2500  334.8125 1339.2500 1369.6875  395.6875
#>  [575] 1217.5000 1034.8750  304.3750 1095.7500 1461.0000  760.9375  456.5625
#>  [582] 1339.2500 1095.7500 1187.0625  243.5000 1278.3750 1187.0625  700.0625
#>  [589]  669.6250  669.6250 1187.0625 1461.0000 1491.4375 1156.6250 1126.1875
#>  [596]  760.9375  639.1875  760.9375  365.2500  608.7500  852.2500 1339.2500
#>  [603] 1095.7500 1247.9375  365.2500 1156.6250  639.1875 1156.6250  426.1250
#>  [610]  182.6250  913.1250 1217.5000  456.5625 1095.7500 1552.3125  669.6250
#>  [617]  517.4375 1369.6875  760.9375 1795.8125 1126.1875 1765.3750 1065.3125
#>  [624] 1521.8750 1004.4375 1095.7500 1034.8750 1582.7500  608.7500  943.5625
#>  [631]  669.6250 1034.8750  821.8125 1217.5000  578.3125        NA  882.6875
#>  [638] 1156.6250 1582.7500  974.0000 1461.0000  669.6250 1308.8125 1369.6875
#>  [645]  730.5000  730.5000  213.0625 1034.8750  882.6875 1065.3125  517.4375
#>  [652] 1187.0625 1521.8750  517.4375 1004.4375 1095.7500  760.9375  821.8125
#>  [659] 1674.0625 1643.6250  913.1250  760.9375 1521.8750  456.5625  487.0000
#>  [666]  882.6875 1278.3750 1674.0625  760.9375 1034.8750  487.0000  213.0625
#>  [673] 1065.3125  273.9375  760.9375  304.3750 1187.0625  730.5000 1156.6250
#>  [680] 1247.9375 1247.9375  273.9375 1247.9375  304.3750 1461.0000  821.8125
#>  [687] 1765.3750  821.8125 1004.4375        NA 1461.0000  700.0625 1095.7500
#>  [694]  760.9375 1187.0625 1126.1875 1369.6875 1430.5625  730.5000  700.0625
#>  [701]  821.8125  730.5000 1552.3125  700.0625  730.5000 1704.5000  821.8125
#>  [708] 1461.0000  426.1250 1582.7500 1217.5000  365.2500  578.3125  395.6875
#>  [715] 1400.1250 1217.5000 1369.6875  791.3750  487.0000 1369.6875 1156.6250
#>  [722] 1339.2500  487.0000  700.0625  182.6250  243.5000 1004.4375  487.0000
#>  [729] 1613.1875 1156.6250  882.6875  669.6250  913.1250 1247.9375 1461.0000
#>  [736]  182.6250 1643.6250 1643.6250 1521.8750 1369.6875  334.8125 1400.1250
#>  [743] 1795.8125  791.3750  487.0000 1308.8125 1217.5000 1217.5000 1369.6875
#>  [750]  669.6250  669.6250 1430.5625 1126.1875 1339.2500  700.0625 1156.6250
#>  [757]  882.6875 1400.1250  365.2500 1461.0000 1095.7500  334.8125 1461.0000
#>  [764] 1187.0625 1278.3750 1126.1875  760.9375 1156.6250 1704.5000  547.8750
#>  [771] 1126.1875  365.2500 1004.4375  913.1250  821.8125  730.5000  517.4375
#>  [778] 1613.1875  913.1250  821.8125  487.0000 1582.7500 1126.1875  395.6875
#>  [785]  943.5625  456.5625 1430.5625  334.8125  182.6250  517.4375  639.1875
#>  [792] 1278.3750 1461.0000 1004.4375 1400.1250 1339.2500 1308.8125 1187.0625
#>  [799] 1430.5625  700.0625 1126.1875  608.7500  365.2500  974.0000 1247.9375
#>  [806]  182.6250  487.0000  547.8750  426.1250 1582.7500 1004.4375  426.1250
#>  [813] 1278.3750  517.4375 1491.4375  882.6875 1613.1875 1643.6250  547.8750
#>  [820] 1674.0625  273.9375 1552.3125  760.9375 1582.7500 1369.6875  578.3125
#>  [827] 1613.1875 1156.6250 1461.0000 1430.5625 1552.3125 1400.1250  365.2500
#>  [834]  182.6250  304.3750 1095.7500  639.1875 1156.6250  304.3750 1369.6875
#>  [841] 1126.1875  700.0625  456.5625  943.5625  365.2500 1674.0625  852.2500
#>  [848] 1643.6250  182.6250  334.8125 1491.4375  517.4375 1461.0000 1187.0625
#>  [855] 1187.0625 1217.5000  791.3750 1461.0000 1400.1250 1156.6250 1065.3125
#>  [862] 1187.0625 1734.9375  547.8750  334.8125  608.7500  913.1250 1126.1875
#>  [869] 1095.7500  760.9375  700.0625  730.5000  608.7500  882.6875 1795.8125
#>  [876] 1126.1875  304.3750 1369.6875 1065.3125 1582.7500  730.5000 1765.3750
#>  [883]  669.6250 1034.8750  487.0000  791.3750  426.1250 1369.6875 1461.0000
#>  [890]  182.6250  913.1250  821.8125  700.0625  243.5000  669.6250  517.4375
#>  [897]  913.1250 1491.4375  426.1250 1674.0625 1582.7500 1156.6250 1339.2500
#>  [904]  365.2500  608.7500 1613.1875  791.3750 1278.3750 1582.7500 1217.5000
#>  [911] 1339.2500  547.8750 1491.4375 1734.9375  426.1250  426.1250 1582.7500
#>  [918]  730.5000  669.6250  760.9375  852.2500  395.6875  852.2500  426.1250
#>  [925] 1065.3125  791.3750  304.3750 1521.8750 1552.3125 1095.7500  821.8125
#>  [932]  426.1250  334.8125  334.8125 1095.7500  821.8125  943.5625 1674.0625
#>  [939] 1156.6250  760.9375  213.0625  517.4375  760.9375  760.9375  395.6875
#>  [946]  365.2500  517.4375 1521.8750  608.7500 1795.8125  760.9375  334.8125
#>  [953]  760.9375  913.1250 1430.5625 1004.4375  943.5625  578.3125  426.1250
#>  [960] 1187.0625  730.5000 1461.0000 1521.8750  578.3125  243.5000  700.0625
#>  [967]  639.1875 1643.6250  608.7500  608.7500  852.2500  395.6875  730.5000
#>  [974]  365.2500 1187.0625  243.5000 1339.2500  273.9375  821.8125 1430.5625
#>  [981] 1461.0000 1339.2500 1643.6250 1613.1875  791.3750 1400.1250  700.0625
#>  [988]  456.5625  213.0625 1643.6250 1369.6875  517.4375 1217.5000 1613.1875
#>  [995]  182.6250 1156.6250 1613.1875 1643.6250  943.5625 1369.6875 1308.8125
#> [1002]  791.3750  791.3750 1704.5000  821.8125  974.0000  547.8750 1278.3750
#> [1009]  395.6875  821.8125  487.0000  821.8125  669.6250  243.5000 1217.5000
#> [1016] 1187.0625  882.6875 1065.3125  639.1875 1095.7500  182.6250  791.3750
#> [1023] 1430.5625  456.5625 1430.5625  395.6875  821.8125 1217.5000 1461.0000
#> [1030]  974.0000 1187.0625 1156.6250  365.2500  578.3125  456.5625 1004.4375
#> [1037]  669.6250 1400.1250  487.0000 1400.1250 1521.8750 1187.0625 1430.5625
#> [1044]  182.6250 1521.8750 1339.2500  669.6250  760.9375 1278.3750 1126.1875
#> [1051]  517.4375        NA 1065.3125  760.9375  669.6250 1613.1875 1582.7500
#> [1058]  517.4375 1278.3750 1156.6250 1613.1875 1461.0000 1704.5000  334.8125
#> [1065] 1765.3750 1461.0000  639.1875  791.3750 1582.7500 1308.8125  213.0625
#> [1072] 1521.8750 1156.6250  669.6250 1521.8750  700.0625 1613.1875 1613.1875
#> [1079] 1156.6250  639.1875 1004.4375  791.3750  334.8125 1765.3750 1217.5000
#> [1086]  821.8125  913.1250  395.6875 1095.7500  791.3750 1582.7500  669.6250
#> [1093]  182.6250  213.0625  791.3750  913.1250  943.5625 1065.3125  273.9375
#> [1100]  943.5625  730.5000 1187.0625  852.2500  456.5625 1156.6250 1278.3750
#> [1107]  395.6875  943.5625 1521.8750 1339.2500 1704.5000  578.3125 1674.0625
#> [1114] 1308.8125  487.0000  456.5625 1521.8750  821.8125  608.7500 1187.0625
#> [1121]  213.0625  700.0625  760.9375 1521.8750  213.0625 1369.6875  852.2500
#> [1128] 1491.4375  426.1250 1582.7500  608.7500 1461.0000 1156.6250 1491.4375
#> [1135]  821.8125 1217.5000  182.6250  365.2500 1004.4375  791.3750  974.0000
#> [1142] 1187.0625  334.8125  791.3750  852.2500 1004.4375  882.6875  547.8750
#> [1149]  578.3125  578.3125 1461.0000  821.8125 1095.7500 1187.0625  852.2500
#> [1156] 1065.3125  608.7500 1491.4375  730.5000  730.5000  882.6875  882.6875
#> [1163] 1461.0000  487.0000  974.0000 1217.5000 1552.3125  974.0000 1247.9375
#> [1170]  456.5625  304.3750  608.7500 1582.7500  395.6875 1704.5000 1734.9375
#> [1177]  700.0625  304.3750  487.0000  791.3750 1582.7500 1095.7500 1582.7500
#> [1184]  608.7500 1491.4375  943.5625  395.6875  395.6875  426.1250 1765.3750
#> [1191]  700.0625 1308.8125  487.0000  913.1250  578.3125 1765.3750  547.8750
#> [1198] 1734.9375  334.8125  304.3750 1765.3750 1095.7500 1643.6250 1278.3750
#> [1205]  852.2500  882.6875  517.4375 1339.2500  395.6875 1400.1250 1734.9375
#> [1212] 1095.7500  791.3750 1613.1875 1613.1875  182.6250  639.1875  669.6250
#> [1219] 1126.1875 1095.7500 1187.0625  304.3750 1339.2500 1521.8750 1187.0625
#> [1226] 1461.0000  334.8125  395.6875 1004.4375 1034.8750  182.6250  821.8125
#> [1233]  243.5000 1582.7500  669.6250  456.5625 1400.1250 1491.4375  882.6875
#> [1240] 1217.5000  395.6875  243.5000  517.4375 1521.8750  852.2500 1461.0000
#> [1247]  852.2500  639.1875  821.8125  334.8125 1795.8125  608.7500  304.3750
#> [1254]  821.8125  395.6875 1156.6250  730.5000 1552.3125 1400.1250  456.5625
#> [1261]  882.6875 1400.1250 1461.0000 1461.0000 1674.0625 1674.0625 1643.6250
#> [1268]  882.6875  578.3125 1339.2500 1034.8750  974.0000  243.5000 1247.9375
#> [1275] 1461.0000  334.8125 1613.1875  760.9375  730.5000 1156.6250        NA
#> [1282]  791.3750  760.9375  395.6875  517.4375 1400.1250 1217.5000  821.8125
#> [1289]  852.2500  821.8125  426.1250  182.6250 1400.1250 1369.6875 1247.9375
#> [1296] 1247.9375 1734.9375  213.0625        NA  304.3750 1643.6250  334.8125
#> [1303]  700.0625  882.6875 1126.1875 1552.3125  852.2500 1552.3125  456.5625
#> [1310] 1369.6875 1795.8125  760.9375  760.9375  304.3750 1704.5000 1461.0000
#> [1317] 1521.8750 1339.2500 1582.7500  273.9375  608.7500  760.9375 1095.7500
#> [1324]  852.2500 1582.7500 1278.3750  791.3750  821.8125 1034.8750 1095.7500
#> [1331] 1278.3750 1004.4375  273.9375  791.3750 1674.0625  547.8750 1217.5000
#> [1338]  700.0625 1187.0625 1187.0625  182.6250  791.3750 1004.4375  334.8125
#> [1345]  517.4375  608.7500 1004.4375 1034.8750 1217.5000  578.3125  791.3750
#> [1352]  913.1250 1065.3125  974.0000  882.6875  669.6250  517.4375 1339.2500
#> [1359]  426.1250 1095.7500 1095.7500  882.6875  821.8125 1278.3750 1521.8750
#> [1366]  304.3750  304.3750 1095.7500  578.3125  517.4375 1765.3750 1187.0625
#> [1373]  243.5000  334.8125  791.3750 1369.6875  243.5000  547.8750  456.5625
#> [1380] 1582.7500 1187.0625  821.8125 1156.6250 1156.6250  487.0000  487.0000
#> [1387] 1643.6250 1004.4375 1156.6250 1521.8750 1734.9375  821.8125 1521.8750
#> [1394]  821.8125  639.1875 1369.6875 1734.9375  791.3750 1582.7500  365.2500
#> [1401] 1095.7500  730.5000 1521.8750 1065.3125 1491.4375  456.5625 1278.3750
#> [1408]  213.0625 1065.3125 1095.7500 1187.0625 1765.3750 1674.0625 1065.3125
#> [1415]  760.9375 1187.0625 1765.3750  426.1250 1308.8125 1247.9375  243.5000
#> [1422]  608.7500  639.1875 1095.7500  426.1250 1461.0000 1704.5000  334.8125
#> [1429]  243.5000  608.7500  730.5000 1095.7500 1339.2500 1613.1875 1643.6250
#> [1436] 1491.4375  578.3125  943.5625 1521.8750 1795.8125 1400.1250 1491.4375
#> [1443] 1430.5625  852.2500 1552.3125 1004.4375  730.5000  334.8125  821.8125
#> [1450] 1126.1875  517.4375  913.1250 1613.1875  547.8750 1156.6250 1156.6250
#> [1457]  456.5625  913.1250  273.9375 1491.4375 1400.1250 1521.8750  273.9375
#> [1464]  426.1250  913.1250  760.9375  213.0625 1521.8750 1643.6250 1430.5625
#> [1471] 1156.6250 1521.8750 1126.1875 1187.0625  243.5000  243.5000  974.0000
#> [1478]  608.7500  608.7500 1339.2500 1461.0000  639.1875 1004.4375 1674.0625
#> [1485]  791.3750  182.6250 1004.4375 1734.9375 1187.0625 1034.8750  182.6250
#> [1492] 1095.7500 1461.0000  730.5000  334.8125  395.6875 1521.8750  426.1250
#> [1499]  517.4375 1308.8125  182.6250  730.5000 1582.7500  760.9375 1491.4375
#> [1506] 1187.0625  730.5000 1217.5000  700.0625 1552.3125 1156.6250  304.3750
#> [1513] 1156.6250 1491.4375 1126.1875  365.2500  182.6250  456.5625 1187.0625
#> [1520] 1156.6250  365.2500  882.6875 1187.0625 1430.5625  791.3750  426.1250
#> [1527]  213.0625 1521.8750 1643.6250  395.6875 1430.5625 1247.9375  730.5000
#> [1534] 1795.8125 1400.1250 1156.6250 1156.6250  669.6250 1674.0625  456.5625
#> [1541] 1461.0000 1217.5000 1582.7500 1613.1875  639.1875  547.8750 1095.7500
#> [1548]  395.6875  426.1250 1065.3125  182.6250 1704.5000  821.8125  517.4375
#> [1555] 1491.4375  882.6875 1217.5000 1521.8750  913.1250 1704.5000  487.0000
#> [1562] 1765.3750 1034.8750  304.3750 1491.4375  547.8750 1765.3750  426.1250
#> [1569] 1582.7500  487.0000 1674.0625  426.1250  517.4375  730.5000  365.2500
#> [1576] 1095.7500  700.0625 1278.3750  395.6875 1339.2500 1126.1875 1734.9375
#> [1583] 1552.3125 1613.1875  426.1250  304.3750  700.0625  547.8750 1674.0625
#> [1590] 1430.5625  913.1250 1217.5000 1126.1875 1095.7500  913.1250 1461.0000
#> [1597]  669.6250 1461.0000  395.6875 1795.8125  639.1875  608.7500  456.5625
#> [1604]  578.3125  395.6875 1613.1875  365.2500 1065.3125  456.5625  974.0000
#> [1611] 1308.8125 1430.5625  974.0000 1461.0000  791.3750 1004.4375 1582.7500
#> [1618] 1217.5000  395.6875 1491.4375 1430.5625 1521.8750  213.0625 1400.1250
#> [1625]  821.8125  608.7500  730.5000 1643.6250  547.8750  730.5000 1187.0625
#> [1632] 1004.4375  395.6875 1278.3750  213.0625 1217.5000  730.5000  456.5625
#> [1639] 1734.9375 1430.5625 1095.7500  943.5625 1126.1875 1156.6250  821.8125
#> [1646] 1308.8125 1613.1875  304.3750  578.3125  334.8125  913.1250 1369.6875
#> [1653] 1095.7500 1369.6875  487.0000  487.0000  669.6250 1247.9375  578.3125
#> [1660] 1461.0000  882.6875 1217.5000  700.0625 1126.1875 1095.7500  791.3750
#> [1667] 1187.0625 1278.3750 1552.3125 1187.0625 1065.3125 1552.3125 1217.5000
#> [1674]  608.7500  882.6875  760.9375 1339.2500 1126.1875  791.3750  669.6250
#> [1681] 1582.7500 1034.8750 1308.8125  730.5000 1582.7500 1126.1875  456.5625
#> [1688] 1156.6250  821.8125  821.8125 1552.3125  821.8125 1369.6875 1217.5000
#> [1695]  487.0000 1308.8125  273.9375  852.2500 1004.4375  487.0000  334.8125
#> [1702]  882.6875  243.5000  852.2500 1400.1250 1613.1875 1126.1875  426.1250
#> [1709] 1461.0000  852.2500  791.3750  974.0000 1095.7500 1400.1250  426.1250
#> [1716]  730.5000  426.1250 1339.2500 1278.3750  395.6875 1674.0625 1704.5000
#> [1723] 1582.7500 1369.6875 1674.0625 1765.3750  456.5625 1704.5000 1065.3125
#> [1730] 1065.3125  456.5625  913.1250 1643.6250  517.4375  760.9375 1613.1875
#> [1737]  578.3125 1461.0000 1491.4375  334.8125  182.6250  395.6875  456.5625
#> [1744]  760.9375  882.6875  760.9375 1491.4375 1126.1875  974.0000 1278.3750
#> [1751] 1217.5000 1278.3750  426.1250  852.2500 1369.6875  304.3750 1156.6250
#> [1758] 1704.5000  791.3750  334.8125  943.5625 1034.8750  456.5625  791.3750
#> [1765] 1339.2500  395.6875  669.6250  882.6875 1004.4375  487.0000 1552.3125
#> [1772]  547.8750  669.6250  608.7500 1430.5625  -91.3125  487.0000 1034.8750
#> [1779]  700.0625 1004.4375  639.1875 1369.6875  791.3750  760.9375  578.3125
#> [1786] 1247.9375  639.1875  182.6250 1339.2500 1369.6875  517.4375 1521.8750
#> [1793]  365.2500 1430.5625 1521.8750  913.1250  334.8125 1491.4375 1308.8125
#> [1800] 1034.8750 1126.1875 1461.0000 1095.7500  974.0000  213.0625  547.8750
#> [1807] 1339.2500  791.3750  700.0625 1156.6250 1582.7500  821.8125 1187.0625
#> [1814]  487.0000  791.3750  213.0625  517.4375  334.8125  730.5000 1156.6250
#> [1821]  730.5000 1491.4375 1369.6875  456.5625  669.6250  852.2500 1156.6250
#> [1828] 1247.9375  821.8125 1491.4375 1613.1875  426.1250  487.0000  882.6875
#> [1835] 1491.4375 1095.7500  913.1250 1521.8750  213.0625  974.0000  426.1250
#> [1842] 1339.2500 1217.5000 1095.7500 1095.7500  365.2500 1308.8125  608.7500
#> [1849] 1461.0000  882.6875  243.5000 1065.3125 1643.6250 1004.4375 1613.1875
#> [1856] 1247.9375 1034.8750  821.8125 1521.8750  760.9375  334.8125  426.1250
#> [1863] 1095.7500 1613.1875 1521.8750  395.6875  395.6875 1187.0625 1126.1875
#> [1870]  852.2500  913.1250  213.0625 1004.4375  273.9375  456.5625  243.5000
#> [1877]  304.3750 1156.6250 1187.0625 1187.0625 1004.4375  273.9375  639.1875
#> [1884]  821.8125 1004.4375  943.5625 1126.1875  943.5625  821.8125 1704.5000
#> [1891] 1156.6250  365.2500 1765.3750  882.6875  882.6875  882.6875  426.1250
#> [1898] 1156.6250  547.8750 1339.2500  882.6875 1308.8125  456.5625  243.5000
#> [1905]  669.6250  974.0000  943.5625  943.5625 1187.0625  304.3750  852.2500
#> [1912]  943.5625 1734.9375  882.6875 1034.8750 1461.0000  943.5625  365.2500
#> [1919] 1613.1875 1400.1250 1247.9375  760.9375 1430.5625 1095.7500  487.0000
#> [1926] 1187.0625  365.2500 1156.6250 1704.5000  456.5625  821.8125  791.3750
#> [1933]  730.5000 1461.0000 1308.8125 1247.9375  608.7500 1704.5000 1765.3750
#> [1940]  243.5000 1491.4375 1065.3125 1552.3125 1126.1875 1491.4375 1369.6875
#> [1947]  669.6250 1034.8750 1004.4375  243.5000  943.5625  304.3750  760.9375
#> [1954]  547.8750 1247.9375 1552.3125 1004.4375  304.3750  852.2500 1278.3750
#> [1961]  760.9375 1369.6875 1400.1250  517.4375 1400.1250 1734.9375  639.1875
#> [1968] 1339.2500  487.0000 1156.6250  578.3125 1156.6250  943.5625 1552.3125
#> [1975]  243.5000 1613.1875  273.9375 1400.1250  852.2500 1156.6250 1461.0000
#> [1982]  243.5000  608.7500 1065.3125  730.5000 1613.1875 1765.3750 1704.5000
#> [1989] 1156.6250 1461.0000  395.6875  791.3750 1095.7500  608.7500  517.4375
#> [1996] 1369.6875  304.3750 1156.6250 1126.1875  700.0625 1430.5625  456.5625
#> [2003]  882.6875 1126.1875 1095.7500  730.5000  273.9375 1126.1875  700.0625
#> [2010]  273.9375  821.8125  639.1875 1095.7500  517.4375 1674.0625  547.8750
#> [2017]  760.9375 1247.9375 1339.2500  578.3125 1674.0625  639.1875 1095.7500
#> [2024]  700.0625  365.2500  487.0000 1461.0000 1095.7500  882.6875 1430.5625
#> [2031]  974.0000  974.0000 1308.8125  456.5625 1552.3125  700.0625 1339.2500
#> [2038]  913.1250  852.2500  852.2500  182.6250 1552.3125  852.2500 1582.7500
#> [2045]  639.1875 1034.8750 1065.3125  760.9375 1461.0000 1187.0625  578.3125
#> [2052]  791.3750  426.1250 1217.5000 1765.3750 1521.8750 1461.0000 1247.9375
#> [2059]  426.1250 1674.0625 1308.8125  547.8750  730.5000  304.3750  487.0000
#> [2066] 1369.6875  821.8125 1491.4375  821.8125 1095.7500 1552.3125  882.6875
#> [2073] 1491.4375  213.0625 1156.6250  669.6250  243.5000 1704.5000 1369.6875
#> [2080] 1156.6250 1613.1875  760.9375 1095.7500  426.1250 1643.6250  730.5000
#> [2087]  882.6875  213.0625  730.5000 1217.5000 1095.7500 1126.1875  334.8125
#> [2094]  213.0625  547.8750  304.3750  730.5000 1339.2500 1521.8750 1187.0625
#> [2101]  821.8125  821.8125  760.9375  608.7500  669.6250 1521.8750  913.1250
#> [2108] 1217.5000 1643.6250  882.6875  334.8125  487.0000  426.1250 1521.8750
#> [2115] 1126.1875 1704.5000 1704.5000  608.7500 1461.0000  517.4375 1004.4375
#> [2122]  243.5000  791.3750  365.2500  365.2500  821.8125  974.0000  852.2500
#> [2129] 1552.3125 1552.3125  974.0000 1613.1875  760.9375 1430.5625  943.5625
#> [2136]  182.6250 1095.7500 1430.5625 1247.9375 1369.6875 1217.5000 1521.8750
#> [2143] 1156.6250  760.9375  821.8125 1400.1250 1613.1875 1339.2500  608.7500
#> [2150] 1156.6250 1400.1250  273.9375  791.3750  243.5000  821.8125  365.2500
#> [2157] 1126.1875  487.0000  273.9375  395.6875  608.7500 1674.0625  273.9375
#> [2164] 1795.8125 1187.0625 1217.5000  426.1250  730.5000 1126.1875  730.5000
#> [2171]  821.8125 1156.6250 1004.4375 1095.7500  974.0000  517.4375  334.8125
#> [2178] 1461.0000  395.6875  578.3125 1278.3750  730.5000 1461.0000  273.9375
#> [2185]  943.5625 1613.1875 1461.0000 1156.6250 1156.6250 1400.1250  273.9375
#> [2192]  578.3125 1643.6250 1704.5000  487.0000  608.7500  791.3750  760.9375
#> [2199] 1156.6250  304.3750  243.5000 1461.0000  700.0625 1004.4375 1400.1250
#> [2206]  487.0000 1674.0625  730.5000 1004.4375  456.5625 1034.8750 1034.8750
#> [2213]  882.6875 1491.4375 1795.8125  426.1250  426.1250  821.8125  487.0000
#> [2220]  578.3125 1217.5000  700.0625  852.2500  852.2500  608.7500 2039.3125
#> [2227] 1582.7500 1065.3125  395.6875 1095.7500  243.5000  426.1250 1552.3125
#> [2234] 1217.5000  760.9375 1674.0625 1126.1875 1187.0625 1704.5000 1795.8125
#> [2241] 1217.5000 1339.2500 1095.7500 1491.4375 1369.6875 1217.5000  821.8125
#> [2248]  395.6875 1400.1250 1217.5000 1704.5000  821.8125  182.6250  273.9375
#> [2255] 1734.9375 1765.3750  182.6250 1613.1875  365.2500  547.8750  426.1250
#> [2262]  182.6250  791.3750 1034.8750 1521.8750 1491.4375 1217.5000 1400.1250
#> [2269]  821.8125  243.5000  334.8125 1430.5625 1674.0625  791.3750 1521.8750
#> [2276]  456.5625 1217.5000  273.9375 1156.6250  578.3125 1004.4375  426.1250
#> [2283] 1126.1875  304.3750  700.0625 1217.5000  426.1250 1308.8125  304.3750
#> [2290] 1126.1875 1369.6875 1521.8750  487.0000  852.2500 1491.4375  821.8125
#> [2297]  426.1250  334.8125  304.3750 1247.9375  334.8125  517.4375  395.6875
#> [2304]  426.1250 1004.4375  456.5625 1308.8125 1187.0625 1521.8750 1491.4375
#> [2311]  517.4375 1004.4375  517.4375  487.0000  273.9375 1430.5625  487.0000
#> [2318] 1643.6250 1339.2500  273.9375  974.0000  974.0000 1339.2500 1095.7500
#> [2325]  821.8125  487.0000 1734.9375  273.9375 1004.4375  487.0000 1552.3125
#> [2332]  426.1250 1461.0000  334.8125  943.5625 1461.0000 1552.3125 1674.0625
#> [2339] 1095.7500  760.9375  760.9375 1156.6250 1095.7500 1674.0625 1034.8750
#> [2346]  791.3750  517.4375  243.5000 1461.0000 1095.7500  608.7500 1461.0000
#> [2353]  943.5625 1582.7500  608.7500  608.7500  852.2500 1582.7500 1369.6875
#> [2360] 1582.7500  213.0625  760.9375 1095.7500 1156.6250  882.6875  700.0625
#> [2367] 1004.4375  943.5625 1582.7500 1582.7500 1521.8750  273.9375 1734.9375
#> [2374]  365.2500 1126.1875 1004.4375 1339.2500 1521.8750 1187.0625 1065.3125
#> [2381] 1065.3125 1430.5625  608.7500  700.0625  608.7500  730.5000 1095.7500
#> [2388] 1156.6250  517.4375  700.0625 1308.8125  487.0000 1095.7500  821.8125
#> [2395] 1187.0625  243.5000 1521.8750 1095.7500  821.8125  182.6250  821.8125
#> [2402]  791.3750  273.9375  882.6875 1339.2500  760.9375 1339.2500 1156.6250
#> [2409]  608.7500 1126.1875 1552.3125  791.3750 1369.6875  821.8125 1582.7500
#> [2416] 1552.3125 1247.9375  821.8125 1613.1875  487.0000 1217.5000  273.9375
#> [2423]  304.3750 1247.9375 1369.6875  700.0625  426.1250 1461.0000 1187.0625
#> [2430] 1734.9375  395.6875 1187.0625  456.5625 1430.5625 1734.9375 1217.5000
#> [2437]  182.6250 1795.8125  456.5625  365.2500  608.7500  882.6875 1095.7500
#> [2444] 1613.1875 1461.0000  517.4375 1765.3750  365.2500 1582.7500  791.3750
#> [2451] 1491.4375 1034.8750 1521.8750  760.9375  974.0000  273.9375 1034.8750
#> [2458] 1369.6875 1065.3125  487.0000 1430.5625 1369.6875  213.0625 1400.1250
#> [2465] 1126.1875 1095.7500 1461.0000  700.0625  426.1250 1034.8750 1582.7500
#> [2472]  456.5625 1491.4375 1004.4375  487.0000  669.6250 1521.8750  487.0000
#> [2479]  456.5625  213.0625  517.4375  395.6875 1308.8125 1065.3125 1461.0000
#> [2486]  213.0625  182.6250  456.5625  365.2500 1217.5000  882.6875  578.3125
#> [2493]  517.4375 1095.7500 1095.7500  304.3750 1034.8750  243.5000 1643.6250
#> [2500] 1156.6250  974.0000 1552.3125 1491.4375        NA  456.5625  487.0000
#> [2507]  273.9375 1217.5000 1126.1875  273.9375 1704.5000  365.2500 1095.7500
#> [2514] 1004.4375 1126.1875  639.1875  547.8750 1552.3125  547.8750  456.5625
#> [2521] 1156.6250 1004.4375  700.0625  821.8125 1095.7500  334.8125  304.3750
#> [2528]  395.6875  213.0625  882.6875  182.6250 1765.3750 1126.1875  974.0000
#> [2535] 1613.1875  334.8125  304.3750 1552.3125  760.9375  487.0000  760.9375
#> [2542] 1430.5625  395.6875  456.5625  273.9375  273.9375 1339.2500 1187.0625
#> [2549]  760.9375  882.6875  487.0000 1217.5000 1339.2500  547.8750 1491.4375
#> [2556]  608.7500  791.3750  791.3750 1674.0625 1400.1250  273.9375 1095.7500
#> [2563] 1734.9375 1461.0000 1217.5000 1065.3125  852.2500  730.5000 1582.7500
#> [2570] 1461.0000  456.5625 1734.9375  334.8125 1156.6250 1369.6875  395.6875
#> [2577]  487.0000  426.1250 1430.5625  608.7500  578.3125 1065.3125  304.3750
#> [2584] 1095.7500  882.6875 1004.4375  852.2500 1095.7500 1187.0625  273.9375
#> [2591] 1552.3125 1217.5000  578.3125  913.1250 1247.9375 1521.8750  639.1875
#> [2598]  456.5625 1004.4375  700.0625  365.2500 1582.7500  700.0625 1521.8750
#> [2605] 1034.8750 1156.6250 1065.3125 1065.3125  913.1250 1430.5625 1491.4375
#> [2612] 1156.6250  213.0625  669.6250 1339.2500 1156.6250  760.9375 1065.3125
#> [2619] 1278.3750 1095.7500  365.2500 1582.7500 1795.8125 1217.5000  365.2500
#> [2626] 1734.9375  426.1250  791.3750  913.1250 1491.4375  913.1250 1643.6250
#> [2633]  456.5625 1187.0625  547.8750 1095.7500  243.5000 1095.7500 1095.7500
#> [2640]  852.2500  791.3750  243.5000  273.9375 1187.0625  669.6250  243.5000
#> [2647] 1369.6875 1369.6875  395.6875 1491.4375  700.0625 1308.8125 1126.1875
#> [2654] 1034.8750 1126.1875  913.1250 1217.5000  487.0000  547.8750  578.3125
#> [2661] 1461.0000 1704.5000  456.5625  456.5625 1004.4375 1004.4375  608.7500
#> [2668]  182.6250  821.8125  273.9375  882.6875  608.7500 1095.7500 1491.4375
#> [2675] 1552.3125  243.5000  213.0625 1095.7500 1552.3125  365.2500 1461.0000
#> [2682] 1156.6250  974.0000 1339.2500 1430.5625  487.0000 1369.6875  852.2500
#> [2689] 1461.0000 1095.7500  669.6250 1400.1250  456.5625 1095.7500        NA
#> [2696]  578.3125 1247.9375 1126.1875  487.0000  791.3750  456.5625  608.7500
#> [2703] 1095.7500  365.2500  760.9375  365.2500 1552.3125 1552.3125 1065.3125
#> [2710]  395.6875 1552.3125 1430.5625  426.1250  882.6875 1369.6875  578.3125
#> [2717]  456.5625  639.1875 1369.6875 1247.9375  760.9375 1552.3125  213.0625
#> [2724]  517.4375 1674.0625 1369.6875  487.0000 1521.8750 1491.4375 1004.4375
#> [2731] 1674.0625  730.5000 1004.4375  791.3750 1095.7500 1461.0000  669.6250
#> [2738] 1004.4375  791.3750  974.0000  760.9375  791.3750  213.0625 1308.8125
#> [2745]  365.2500 1339.2500 1765.3750 1065.3125  791.3750  700.0625  730.5000
#> [2752] 1491.4375  213.0625  974.0000 1674.0625 1491.4375  304.3750  669.6250
#> [2759]  669.6250 1217.5000  913.1250 1095.7500 1126.1875  669.6250 1065.3125
#> [2766]  273.9375  426.1250  852.2500 1430.5625 1187.0625 1004.4375 1765.3750
#> [2773]  669.6250  304.3750  426.1250  426.1250  304.3750  760.9375 1065.3125
#> [2780]  456.5625 1491.4375 1095.7500 1643.6250  700.0625  456.5625 1795.8125
#> [2787] 1126.1875 1734.9375  487.0000 1521.8750 1126.1875        NA  639.1875
#> [2794] 1674.0625  943.5625 1126.1875  882.6875 1217.5000  243.5000  487.0000
#> [2801]  487.0000  334.8125  456.5625 1156.6250  730.5000  456.5625  304.3750
#> [2808] 1247.9375  487.0000 1613.1875  395.6875 1034.8750 1308.8125 1187.0625
#> [2815]  365.2500  547.8750  304.3750        NA        NA        NA        NA
#> [2822] 1065.3125        NA        NA        NA        NA        NA        NA
#> [2829]  700.0625 1126.1875  669.6250 1674.0625  365.2500  456.5625 1156.6250
#> [2836] 1430.5625 1400.1250  700.0625 1795.8125  639.1875  852.2500 1034.8750
#> [2843] 1156.6250  974.0000 1582.7500 1430.5625  974.0000  273.9375 1461.0000
#> [2850]  365.2500 1247.9375 1674.0625 1095.7500  669.6250 1521.8750 1247.9375
#> [2857]  273.9375 1521.8750 1126.1875 1004.4375  791.3750        NA  334.8125
#> [2864] 1339.2500  700.0625  791.3750 1613.1875 1308.8125  913.1250 1491.4375
#> [2871]  273.9375  852.2500 1126.1875  760.9375 1461.0000 1491.4375  791.3750
#> [2878]  243.5000  456.5625 1369.6875  882.6875  821.8125  639.1875 1217.5000
#> [2885] 1217.5000  213.0625  517.4375  974.0000 1582.7500  639.1875  639.1875
#> [2892] 1582.7500  547.8750  760.9375  669.6250  669.6250  730.5000  487.0000
#> [2899]  243.5000 1430.5625        NA 1613.1875 1217.5000  669.6250 1613.1875
#> [2906] 1004.4375  882.6875  882.6875  487.0000 1034.8750  213.0625  852.2500
#> [2913]  821.8125  852.2500 1430.5625  669.6250  395.6875  760.9375 1521.8750
#> [2920]  304.3750  578.3125  334.8125 1217.5000 1095.7500  456.5625  487.0000
#> [2927] 1187.0625 1369.6875  456.5625 1095.7500 1795.8125  334.8125  517.4375
#> [2934]  913.1250  213.0625 1034.8750 1521.8750  395.6875  395.6875 1582.7500
#> [2941] 1156.6250  882.6875 1765.3750 1156.6250  669.6250  791.3750 1187.0625
#> [2948] 1095.7500  974.0000  669.6250 1278.3750 1461.0000  639.1875  395.6875
#> [2955]  456.5625  760.9375 1643.6250 1461.0000  882.6875  882.6875  791.3750
#> [2962]  273.9375 1156.6250 1491.4375  243.5000  669.6250 1613.1875  639.1875
#> [2969]  852.2500 1552.3125 1643.6250  821.8125 1126.1875 1521.8750  243.5000
#> [2976]  760.9375 1461.0000 1004.4375  852.2500  456.5625 1461.0000  730.5000
#> [2983] 1004.4375 1674.0625 1217.5000  426.1250 1004.4375  273.9375  243.5000
#> [2990] 1674.0625  426.1250  608.7500  334.8125 1582.7500  760.9375 1400.1250
#> [2997] 1795.8125 1491.4375 1126.1875 1582.7500 1491.4375  882.6875 1369.6875
#> [3004] 1491.4375  304.3750  760.9375  700.0625  608.7500  243.5000 1400.1250
#> [3011] 1126.1875  456.5625 1278.3750 1430.5625 1187.0625  365.2500 1065.3125
#> [3018]  456.5625 1217.5000  243.5000 1582.7500 1187.0625  365.2500  304.3750
#> [3025]  243.5000 1521.8750  578.3125 1704.5000  639.1875 1461.0000 1004.4375
#> [3032]  974.0000 1400.1250 1582.7500  974.0000 1278.3750 1734.9375 1613.1875
#> [3039] 1521.8750 1704.5000 1521.8750 1582.7500 1400.1250 1795.8125 1552.3125
#> [3046] 1308.8125 1461.0000 1065.3125  730.5000 1461.0000 1430.5625  456.5625
#> [3053]  669.6250  243.5000 1400.1250  487.0000 1247.9375 1795.8125 1400.1250
#> [3060] 1004.4375        NA  669.6250 1126.1875 1582.7500 1156.6250 1582.7500
#> [3067]  395.6875  943.5625  578.3125 1187.0625  608.7500  547.8750 1095.7500
#> [3074] 1065.3125  791.3750 1278.3750  760.9375  700.0625  700.0625 1704.5000
#> [3081]  669.6250  304.3750 1795.8125  760.9375  426.1250        NA 1187.0625
#> [3088] 1430.5625  791.3750  882.6875 1552.3125 1004.4375 1247.9375  821.8125
#> [3095] 1430.5625  182.6250 1095.7500 1308.8125  243.5000 1095.7500 1552.3125
#> [3102]  913.1250 1491.4375 1095.7500  913.1250  608.7500  882.6875 1552.3125
#> [3109] 1095.7500  882.6875 1247.9375 1095.7500  730.5000  273.9375  517.4375
#> [3116]  760.9375 1004.4375 1461.0000 1095.7500  304.3750  517.4375  487.0000
#> [3123]  730.5000 1278.3750 1613.1875        NA  334.8125 1247.9375 1126.1875
#> [3130] 1126.1875 1552.3125  365.2500 1156.6250  395.6875  243.5000 1065.3125
#> [3137]  487.0000 1582.7500  243.5000  243.5000 1430.5625  974.0000        NA
#> [3144] 1430.5625 1400.1250  852.2500  395.6875 1552.3125 1552.3125 1582.7500
#> [3151] 1126.1875 1491.4375  395.6875  821.8125 1674.0625 1613.1875  456.5625
#> [3158]  852.2500 1034.8750 1400.1250  974.0000  974.0000 1674.0625 1704.5000
#> [3165]  882.6875 1400.1250  700.0625 1156.6250 1126.1875 1339.2500  334.8125
#> [3172]  852.2500 1461.0000 1126.1875 1765.3750 1187.0625  943.5625  760.9375
#> [3179]  760.9375 1369.6875 1552.3125  760.9375 1582.7500 1734.9375  882.6875
#> [3186]  791.3750  669.6250  547.8750 1552.3125 1461.0000  395.6875  913.1250
#> [3193]  791.3750  213.0625 1461.0000 1095.7500 1034.8750 1004.4375 1400.1250
#> [3200] 1308.8125 1613.1875  365.2500  365.2500 1004.4375 1521.8750 1795.8125
#> [3207] 1582.7500  213.0625 1582.7500 1034.8750  608.7500  578.3125  273.9375
#> [3214]  913.1250  365.2500  365.2500 1217.5000 1704.5000  365.2500 1613.1875
#> [3221] 1065.3125 1095.7500  365.2500 1034.8750 1126.1875  517.4375  487.0000
#> [3228] 1369.6875  426.1250 1126.1875  213.0625 1491.4375  943.5625 1217.5000
#> [3235]  578.3125  791.3750 1308.8125  365.2500 1187.0625 1734.9375  730.5000
#> [3242] 1491.4375  760.9375 1795.8125 1156.6250 1521.8750 1674.0625  791.3750
#> [3249]  852.2500 1491.4375  943.5625 1278.3750  882.6875  730.5000 1126.1875
#> [3256]  426.1250  730.5000  243.5000  608.7500  730.5000  852.2500 1247.9375
#> [3263] 1065.3125 1795.8125 1065.3125  456.5625 1187.0625  791.3750 1339.2500
#> [3270]  334.8125  395.6875 1308.8125  913.1250  334.8125  487.0000  639.1875
#> [3277]  730.5000  426.1250  974.0000 1491.4375 1187.0625 1004.4375 1065.3125
#> [3284]  395.6875 1582.7500  821.8125  730.5000 1065.3125 1521.8750  365.2500
#> [3291]  791.3750 1491.4375 1795.8125 1004.4375  487.0000  882.6875 1034.8750
#> [3298] 1521.8750  852.2500 1308.8125 1765.3750  882.6875  426.1250  760.9375
#> [3305] 1552.3125 1034.8750  791.3750  426.1250  243.5000  730.5000 1156.6250
#> [3312]  395.6875 1491.4375  426.1250 1247.9375  182.6250 1582.7500 1065.3125
#> [3319] 1552.3125 1156.6250  821.8125 1613.1875  913.1250  365.2500 1491.4375
#> [3326]  700.0625  700.0625 1521.8750  669.6250  700.0625 1095.7500 1734.9375
#> [3333] 1582.7500  213.0625  456.5625 1795.8125  974.0000 1065.3125 1643.6250
#> [3340]  852.2500  213.0625  547.8750 1156.6250  700.0625  456.5625  821.8125
#> [3347]  821.8125 2039.3125 1400.1250 1308.8125 1247.9375 1126.1875  395.6875
#> [3354] 1126.1875 1582.7500  243.5000  547.8750 1126.1875  487.0000 1126.1875
#> [3361] 1065.3125  395.6875  882.6875 1065.3125  821.8125 1704.5000  304.3750
#> [3368]  608.7500  974.0000  913.1250  395.6875 1400.1250  273.9375 1369.6875
#> [3375]  426.1250 1004.4375 1217.5000  517.4375 1187.0625  700.0625  669.6250
#> [3382] 1521.8750 1156.6250  913.1250