Apply blocked weighted bootstrap to indicator datasets

A wrapper for bootBW function in bbw package.

boot_estimate(indicator = c("ifaDF", "iycfDF", "mnpDF", "vitDF",
  "screenDF", "anthroDF"), w, replicates = 399,
  core.columns = c("spid", "cid", "did", "eid", "motherID", "m2"))

Arguments

indicator
A character vector of indicator data.frame names

w
Population dataset

replicates
Number of bootstrap replicates

core.columns
A vector of variable names included in indicator data.frames

Value

A data.frame of indicator estimates with lower and upper confidence limits

Examples

boot_estimate(indicator = "iycfDF", w = psuDataGM, replicates = 9)

#> $iycfDF
#>    varLabel   estimate        lcl        ucl
#> 1      icf1 0.87672956 0.83953463 0.91500323
#> 2      icf2 0.86797386 0.81236352 0.89322703
#> 3     icf2a 0.18627451 0.13294385 0.29989238
#> 4     icf2b 0.56521739 0.40189882 0.73376509
#> 5     icf2c 0.00000000 0.00000000 0.00000000
#> 6     icf2d 0.07608696 0.02847234 0.11547587
#> 7     icf2e 0.04504505 0.00000000 0.09435939
#> 8     icf2f 0.04901961 0.00000000 0.07445812
#> 9     icf2g 0.02531646 0.00000000 0.05008085
#> 10     icf3 5.45029240 5.30829951 5.73911483
#>