Function to recode age variable and convert to age in days

Function to recode age variable and convert to age in days

recode_age(df)

Arguments

df
A coverage data.frame collected for the Liberia Coverage Survey

Value

A vector of recoded ages based on any re-testing done

Examples

recode_age(df = coverageData.r2)

#>    [1]  456.5625  882.6875  669.6250 1187.0625 1187.0625 1582.7500  852.2500
#>    [8] 1278.3750 1582.7500  760.9375  913.1250  852.2500  578.3125  760.9375
#>   [15]  882.6875  882.6875 1461.0000 1217.5000 1582.7500  273.9375  395.6875
#>   [22] 1765.3750  304.3750 1461.0000 1491.4375 1126.1875  243.5000  426.1250
#>   [29] 1187.0625 1278.3750 1156.6250  334.8125 1704.5000  547.8750 1095.7500
#>   [36] 1247.9375  243.5000  821.8125  821.8125 1704.5000  882.6875  821.8125
#>   [43] 1004.4375  578.3125  608.7500 1674.0625  426.1250  821.8125 1369.6875
#>   [50] 1126.1875  821.8125 1400.1250  974.0000 1491.4375  852.2500 1217.5000
#>   [57] 1582.7500  487.0000  608.7500 1156.6250  791.3750  913.1250  547.8750
#>   [64] 1126.1875  426.1250  334.8125 1552.3125  273.9375  182.6250 1491.4375
#>   [71] 1491.4375  760.9375 1552.3125  639.1875 1400.1250  791.3750 1034.8750
#>   [78] 1339.2500  304.3750 1674.0625 1126.1875 1704.5000  334.8125 1400.1250
#>   [85] 1521.8750 1065.3125  882.6875  395.6875  791.3750  334.8125  456.5625
#>   [92]  456.5625  821.8125  608.7500 1308.8125  791.3750 1369.6875 1065.3125
#>   [99]  395.6875 1613.1875  852.2500  913.1250 1430.5625 1552.3125  791.3750
#>  [106]  700.0625 1643.6250  517.4375  547.8750 1065.3125 1247.9375 1247.9375
#>  [113] 1521.8750  334.8125 1004.4375 1217.5000 1552.3125  821.8125 1217.5000
#>  [120] 1552.3125  791.3750 1126.1875  639.1875  791.3750  426.1250  456.5625
#>  [127]  182.6250  730.5000 1339.2500 1582.7500 1552.3125  456.5625  639.1875
#>  [134]  304.3750 1521.8750 1126.1875  760.9375  213.0625 1126.1875  365.2500
#>  [141] 1461.0000 1126.1875 1369.6875  639.1875 1369.6875  882.6875  821.8125
#>  [148]  852.2500 1004.4375  456.5625 1187.0625 1339.2500  395.6875  730.5000
#>  [155] 1552.3125 1187.0625 1369.6875 1156.6250 1339.2500  395.6875  304.3750
#>  [162]  456.5625 1461.0000  760.9375 1461.0000  882.6875  943.5625  608.7500
#>  [169]  395.6875 1521.8750  882.6875  213.0625  608.7500  334.8125 1491.4375
#>  [176] 1187.0625  943.5625  882.6875 1430.5625 1643.6250  821.8125  608.7500
#>  [183] 1034.8750 1734.9375 1765.3750 1247.9375  791.3750  760.9375 1582.7500
#>  [190]  456.5625 1065.3125  700.0625 1156.6250 1430.5625 1278.3750  852.2500
#>  [197]  821.8125 1491.4375  974.0000 1734.9375  547.8750  791.3750 1491.4375
#>  [204] 1613.1875  974.0000 1643.6250  456.5625  304.3750  669.6250  273.9375
#>  [211] 1643.6250  760.9375 1734.9375  730.5000  334.8125 1065.3125 1065.3125
#>  [218] 1095.7500  639.1875  882.6875  791.3750  517.4375  426.1250 1430.5625
#>  [225] 1491.4375  578.3125 1217.5000  304.3750 1247.9375        NA 1004.4375
#>  [232] 1369.6875  365.2500  243.5000 1704.5000  334.8125 1065.3125 1217.5000
#>  [239] 1004.4375 1461.0000 1552.3125  243.5000  182.6250  882.6875 1704.5000
#>  [246]  517.4375 1065.3125  243.5000 1582.7500  456.5625 1552.3125 1430.5625
#>  [253]  426.1250  365.2500 1582.7500 1613.1875  456.5625  304.3750 1247.9375
#>  [260]  608.7500  243.5000 1521.8750 1704.5000 1126.1875  760.9375 1795.8125
#>  [267] 1643.6250  639.1875  395.6875  487.0000  365.2500  730.5000  334.8125
#>  [274]  213.0625  334.8125  243.5000 1095.7500  273.9375  395.6875  426.1250
#>  [281]  365.2500 1247.9375 1065.3125 1795.8125 1156.6250 1217.5000  882.6875
#>  [288]  760.9375  791.3750  273.9375  487.0000  487.0000  639.1875 1369.6875
#>  [295] 1339.2500 1521.8750 1247.9375  760.9375  487.0000  852.2500 1430.5625
#>  [302] 1308.8125 1521.8750  243.5000 1156.6250  730.5000  395.6875 1704.5000
#>  [309] 1491.4375  487.0000  487.0000  791.3750 1734.9375 1369.6875 1034.8750
#>  [316] 1613.1875  456.5625 1126.1875 1126.1875 1095.7500 1095.7500  943.5625
#>  [323] 1521.8750 1369.6875  456.5625 1156.6250 1704.5000 1156.6250  669.6250
#>  [330] 1247.9375  456.5625  334.8125 1643.6250 1491.4375 1552.3125 1491.4375
#>  [337] 1369.6875  639.1875  821.8125  243.5000 1734.9375  456.5625  243.5000
#>  [344] 1095.7500 1247.9375  913.1250 1369.6875 1521.8750 1400.1250 1156.6250
#>  [351] 1369.6875  334.8125 1339.2500 1126.1875  943.5625 1461.0000  730.5000
#>  [358]  365.2500  426.1250  243.5000 1156.6250 1582.7500  213.0625 1552.3125
#>  [365]  182.6250  608.7500 1704.5000 1765.3750  395.6875  487.0000 1430.5625
#>  [372]  608.7500  882.6875  974.0000  760.9375  426.1250  456.5625  913.1250
#>  [379] 1582.7500 1004.4375 1521.8750 1400.1250  821.8125 1400.1250  365.2500
#>  [386]  426.1250 1613.1875 1278.3750 1400.1250 1126.1875 1643.6250 1521.8750
#>  [393] 1339.2500 1461.0000 1552.3125        NA 1613.1875 1156.6250 1582.7500
#>  [400] 1004.4375  517.4375 1369.6875 1278.3750  730.5000  243.5000  213.0625
#>  [407]  852.2500  426.1250  608.7500  273.9375 1369.6875  913.1250  913.1250
#>  [414]  791.3750  730.5000  547.8750  669.6250 1674.0625  669.6250  517.4375
#>  [421]  913.1250  426.1250 1491.4375 1156.6250  669.6250  426.1250  852.2500
#>  [428]  426.1250  943.5625  456.5625 1400.1250  395.6875 1643.6250 1521.8750
#>  [435]  547.8750  456.5625  273.9375 1095.7500 1491.4375  578.3125 1308.8125
#>  [442]  395.6875 1795.8125  426.1250 1217.5000 1704.5000  213.0625  669.6250
#>  [449] 1491.4375  213.0625  487.0000  334.8125  700.0625 1156.6250 1126.1875
#>  [456] 1339.2500  791.3750 1734.9375 2039.3125  852.2500 1521.8750 1734.9375
#>  [463]  913.1250 1034.8750  882.6875  304.3750 1339.2500  608.7500 1704.5000
#>  [470]  426.1250  334.8125  821.8125  913.1250  182.6250 1674.0625  456.5625
#>  [477] 1765.3750  487.0000 1430.5625 1643.6250  669.6250  487.0000  517.4375
#>  [484] 1430.5625  791.3750 1430.5625  730.5000 1004.4375 1704.5000 1004.4375
#>  [491]  608.7500 1734.9375  517.4375 1095.7500  913.1250 1400.1250 1034.8750
#>  [498]  243.5000 1065.3125 1582.7500  730.5000 1004.4375 1004.4375  304.3750
#>  [505]  913.1250 1217.5000  882.6875 1613.1875 1065.3125 1065.3125  213.0625
#>  [512] 1187.0625 1734.9375  608.7500        NA  730.5000  882.6875  882.6875
#>  [519]  882.6875  821.8125  213.0625  821.8125  700.0625  882.6875  365.2500
#>  [526]  456.5625 1461.0000 1187.0625 1004.4375 1734.9375  395.6875  760.9375
#>  [533]  273.9375 1247.9375  700.0625  182.6250  578.3125  273.9375  456.5625
#>  [540] 1187.0625 1704.5000  487.0000 1521.8750 1095.7500  395.6875  974.0000
#>  [547]  943.5625  456.5625  943.5625  791.3750 1521.8750 1400.1250  487.0000
#>  [554] 1217.5000 1461.0000  852.2500  608.7500 1734.9375  395.6875  426.1250
#>  [561] 1582.7500  213.0625 1126.1875  365.2500  608.7500 1278.3750  913.1250
#>  [568]  456.5625 1704.5000 1400.1250 1430.5625  730.5000  456.5625  426.1250
#>  [575] 1734.9375 1674.0625  639.1875  943.5625  760.9375  487.0000 1187.0625
#>  [582] 1521.8750  456.5625 1339.2500 1491.4375 1004.4375 1247.9375  456.5625
#>  [589] 1308.8125  608.7500 1308.8125 1187.0625 1095.7500  730.5000  517.4375
#>  [596]  882.6875  517.4375  334.8125 1126.1875 1521.8750  700.0625 1126.1875
#>  [603] 1613.1875  821.8125 1582.7500 1521.8750  669.6250 1308.8125 1430.5625
#>  [610]  304.3750 1461.0000  547.8750  669.6250 1521.8750 1552.3125  395.6875
#>  [617]  182.6250 1369.6875  487.0000  882.6875  395.6875  700.0625 1004.4375
#>  [624]  517.4375  304.3750  304.3750 1156.6250 1308.8125  791.3750  608.7500
#>  [631] 1126.1875  882.6875  700.0625  365.2500 1613.1875 1004.4375 1339.2500
#>  [638] 1278.3750 1095.7500  578.3125  456.5625 1156.6250  426.1250 1278.3750
#>  [645] 1034.8750 1674.0625 1217.5000 1308.8125 1521.8750 1095.7500 1704.5000
#>  [652] 1369.6875 1461.0000 1430.5625  760.9375 1065.3125  273.9375 1034.8750
#>  [659]  517.4375  578.3125  547.8750 1034.8750 1491.4375  517.4375  943.5625
#>  [666]  365.2500  243.5000  213.0625  639.1875 1430.5625 1430.5625  365.2500
#>  [673]  791.3750  243.5000  639.1875 1643.6250 1491.4375  882.6875  639.1875
#>  [680]  821.8125 1095.7500 1461.0000  760.9375 1187.0625 1217.5000  852.2500
#>  [687] 1674.0625  547.8750  913.1250  304.3750  639.1875  913.1250  395.6875
#>  [694]  456.5625 1582.7500 1461.0000  395.6875  669.6250  608.7500 1126.1875
#>  [701] 1156.6250  974.0000  730.5000 1734.9375 1247.9375 1034.8750  730.5000
#>  [708]  608.7500  395.6875 1095.7500 1126.1875  882.6875 1278.3750  182.6250
#>  [715]  791.3750  487.0000 1095.7500  821.8125 1156.6250  547.8750  547.8750
#>  [722]  882.6875  882.6875  974.0000  852.2500 1430.5625  974.0000  365.2500
#>  [729] 1308.8125 1308.8125  852.2500 1126.1875  334.8125 1308.8125 1187.0625
#>  [736]  730.5000 1765.3750  730.5000 1004.4375  821.8125 1278.3750 1521.8750
#>  [743]  304.3750  791.3750 1217.5000  974.0000 1521.8750        NA 1430.5625
#>  [750] 1613.1875  882.6875  791.3750  243.5000 1095.7500  669.6250  517.4375
#>  [757] 1065.3125  578.3125  669.6250  334.8125 1339.2500 1339.2500 1461.0000
#>  [764] 1369.6875  365.2500 1552.3125  791.3750 1430.5625 1156.6250 1187.0625
#>  [771]  182.6250 1461.0000  730.5000 1217.5000  517.4375  487.0000  547.8750
#>  [778] 1126.1875  365.2500  456.5625  578.3125  182.6250  821.8125 1461.0000
#>  [785]  669.6250  456.5625  273.9375  882.6875  608.7500  821.8125  882.6875
#>  [792]        NA 1552.3125 1247.9375  487.0000 1704.5000 1095.7500  304.3750
#>  [799]  213.0625 1217.5000  365.2500  943.5625 1521.8750  334.8125  608.7500
#>  [806] 1156.6250 1491.4375 1461.0000 1095.7500  365.2500  760.9375  365.2500
#>  [813]  243.5000 1217.5000 1582.7500  669.6250  334.8125  487.0000 1521.8750
#>  [820]  760.9375  334.8125 1461.0000 1004.4375 1278.3750 1461.0000 1521.8750
#>  [827]  273.9375  182.6250  760.9375  273.9375 1430.5625  852.2500  213.0625
#>  [834]  821.8125  852.2500 1491.4375 1308.8125  913.1250  304.3750 1643.6250
#>  [841] 1491.4375  395.6875  852.2500 1613.1875 1734.9375  304.3750  213.0625
#>  [848] 1217.5000 1461.0000  395.6875  395.6875  426.1250  700.0625  426.1250
#>  [855] 1278.3750        NA  669.6250  304.3750 1613.1875  669.6250  791.3750
#>  [862] 1734.9375  943.5625 1156.6250 1643.6250  974.0000  182.6250 1704.5000
#>  [869] 1126.1875  456.5625 1156.6250  487.0000 1247.9375 1187.0625  487.0000
#>  [876]  243.5000 1643.6250  547.8750 1034.8750  943.5625 1034.8750 1613.1875
#>  [883] 1217.5000  395.6875  974.0000 1278.3750 1369.6875  882.6875  821.8125
#>  [890]  578.3125 1582.7500  487.0000  730.5000 1126.1875  273.9375  395.6875
#>  [897]  304.3750  456.5625 1795.8125  547.8750  669.6250 1004.4375 1278.3750
#>  [904] 1065.3125  760.9375  730.5000  669.6250 1095.7500 1339.2500  700.0625
#>  [911]  304.3750 1126.1875  760.9375 1004.4375  487.0000 1339.2500  456.5625
#>  [918]  760.9375 1552.3125 1400.1250 1126.1875 1126.1875 1065.3125  608.7500
#>  [925]  730.5000  426.1250  974.0000 1278.3750 1369.6875  517.4375  243.5000
#>  [932] 1095.7500 1004.4375  791.3750 1400.1250 1034.8750 1613.1875 1491.4375
#>  [939] 1004.4375  913.1250  791.3750 1795.8125  517.4375 1795.8125  517.4375
#>  [946] 1582.7500 1095.7500  791.3750 1065.3125  639.1875 1795.8125 1369.6875
#>  [953]  852.2500  852.2500 1643.6250 1552.3125 1400.1250  182.6250 1156.6250
#>  [960] 1217.5000 1461.0000  273.9375 1521.8750        NA  487.0000 1339.2500
#>  [967]  365.2500 1126.1875  852.2500 1582.7500  578.3125 1461.0000  791.3750
#>  [974] 1308.8125  974.0000 1308.8125 1095.7500 1461.0000 1734.9375 1674.0625
#>  [981]  365.2500  669.6250 1034.8750  517.4375  730.5000 1582.7500 1217.5000
#>  [988] 1461.0000 1156.6250 1734.9375 1491.4375 1187.0625 1065.3125 1339.2500
#>  [995] 1187.0625 1491.4375 1582.7500 1430.5625  882.6875 1582.7500 1369.6875
#> [1002] 1217.5000  669.6250  182.6250  730.5000  730.5000  760.9375  882.6875
#> [1009]  426.1250  426.1250 1430.5625  365.2500  669.6250  578.3125  456.5625
#> [1016]  395.6875 1095.7500 1461.0000  426.1250 1704.5000  821.8125  365.2500
#> [1023]  304.3750  243.5000  669.6250 1552.3125 1278.3750 1156.6250 1247.9375
#> [1030] 1430.5625  913.1250 1308.8125 1674.0625 1217.5000  852.2500  700.0625
#> [1037] 1521.8750  578.3125 1004.4375 1156.6250  426.1250 1521.8750 1491.4375
#> [1044] 1095.7500  243.5000  852.2500  974.0000 1400.1250 1430.5625  273.9375
#> [1051] 1126.1875 1217.5000 1278.3750  304.3750  182.6250  700.0625 1643.6250
#> [1058] 1461.0000 1461.0000 1126.1875  456.5625 1156.6250  456.5625 1308.8125
#> [1065]  852.2500 1065.3125  304.3750  456.5625 1491.4375  730.5000  243.5000
#> [1072] 1187.0625  730.5000  213.0625 1095.7500  304.3750 1247.9375  456.5625
#> [1079] 1430.5625 1461.0000 1369.6875 1339.2500 1156.6250 1156.6250 1095.7500
#> [1086]  821.8125 1095.7500  882.6875 1126.1875 1369.6875 1095.7500 1034.8750
#> [1093] 1247.9375 1461.0000  821.8125 1461.0000 1369.6875 1491.4375  639.1875
#> [1100]  334.8125  456.5625 1704.5000  821.8125  456.5625 1278.3750 1795.8125
#> [1107]  578.3125 1461.0000 1674.0625  395.6875  760.9375 1187.0625 1095.7500
#> [1114] 1613.1875 1430.5625  639.1875 1521.8750  304.3750  334.8125 1217.5000
#> [1121]  395.6875  547.8750  821.8125  395.6875  760.9375  760.9375  608.7500
#> [1128] 1034.8750  517.4375  547.8750 1339.2500  760.9375  943.5625 1704.5000
#> [1135]  456.5625  517.4375  334.8125 1126.1875 1187.0625  243.5000  365.2500
#> [1142] 1217.5000  852.2500  913.1250 1613.1875  669.6250 1156.6250 1765.3750
#> [1149] 1552.3125 1582.7500  760.9375 1765.3750  943.5625 1491.4375  669.6250
#> [1156]  426.1250 1156.6250  608.7500 1339.2500 1552.3125  669.6250 1795.8125
#> [1163]  700.0625  213.0625 1187.0625 1065.3125  395.6875  273.9375 1430.5625
#> [1170]  974.0000  608.7500 1156.6250  334.8125  426.1250 1430.5625  487.0000
#> [1177]  456.5625        NA 1156.6250 1126.1875  882.6875 1126.1875  913.1250
#> [1184]  547.8750  700.0625  700.0625 1095.7500        NA 1582.7500  913.1250
#> [1191]  608.7500  821.8125 1795.8125 1765.3750  487.0000 1643.6250  882.6875
#> [1198]  426.1250  243.5000 1095.7500  791.3750 1430.5625 1674.0625 1582.7500
#> [1205]  974.0000 1247.9375 1582.7500  273.9375  304.3750 1734.9375  882.6875
#> [1212] 1674.0625  578.3125 1613.1875  365.2500  852.2500  334.8125 1582.7500
#> [1219]  426.1250  791.3750  913.1250  791.3750  852.2500 1217.5000  334.8125
#> [1226] 1156.6250 1034.8750 1461.0000  365.2500  913.1250  700.0625  243.5000
#> [1233]  395.6875  273.9375  639.1875  730.5000 1187.0625  395.6875 1187.0625
#> [1240]  852.2500  304.3750 1156.6250 1247.9375 1491.4375 1187.0625  882.6875
#> [1247]  578.3125  608.7500 1613.1875 1034.8750 1400.1250  791.3750  365.2500
#> [1254] 1004.4375 1217.5000 1582.7500  852.2500 1491.4375 1308.8125  456.5625
#> [1261]  273.9375  456.5625  669.6250  791.3750  273.9375 1369.6875 1765.3750
#> [1268]  304.3750  273.9375 1187.0625 1369.6875 1187.0625  273.9375  608.7500
#> [1275]  487.0000 1369.6875  487.0000  821.8125  760.9375 1582.7500  882.6875
#> [1282]  791.3750 1795.8125 1095.7500  547.8750  608.7500  426.1250  913.1250
#> [1289]  882.6875  821.8125  365.2500 1095.7500  760.9375  304.3750 1156.6250
#> [1296]  669.6250  456.5625 1613.1875  852.2500  639.1875 1339.2500  426.1250
#> [1303] 1156.6250 1308.8125        NA 1308.8125 1582.7500  700.0625  974.0000
#> [1310]  791.3750 1491.4375 1004.4375  821.8125  456.5625  365.2500 1126.1875
#> [1317]  243.5000  730.5000 1004.4375  426.1250  791.3750  882.6875 1643.6250
#> [1324] 1065.3125 1704.5000  852.2500 1156.6250 1034.8750 1339.2500  304.3750
#> [1331] 1613.1875  974.0000  487.0000 1765.3750  578.3125  213.0625 1278.3750
#> [1338] 1369.6875  426.1250 1582.7500  852.2500 1339.2500 1582.7500 1704.5000
#> [1345] 1491.4375  639.1875  243.5000 1247.9375  578.3125  517.4375 1034.8750
#> [1352]  760.9375  639.1875 1613.1875 1156.6250  852.2500 1247.9375 1734.9375
#> [1359] 1065.3125  852.2500  182.6250  182.6250  426.1250  882.6875  669.6250
#> [1366]  730.5000  243.5000  517.4375  700.0625  304.3750 1400.1250 1491.4375
#> [1373]  517.4375  547.8750 1521.8750  791.3750  730.5000 1613.1875  243.5000
#> [1380] 1613.1875 1339.2500  760.9375 1095.7500  182.6250 1795.8125  882.6875
#> [1387]  273.9375 1217.5000  730.5000 1187.0625 1765.3750 1552.3125 1126.1875
#> [1394] 1004.4375 1521.8750  182.6250 1004.4375 1400.1250  943.5625  456.5625
#> [1401] 1126.1875  456.5625 1613.1875  730.5000 1552.3125 1217.5000 1582.7500
#> [1408]  760.9375  821.8125 1156.6250 1095.7500  791.3750  943.5625 1491.4375
#> [1415]  852.2500  426.1250  365.2500 1613.1875  913.1250  365.2500 1004.4375
#> [1422] 1552.3125  669.6250 1461.0000 1095.7500  395.6875 1400.1250  882.6875
#> [1429] 1095.7500 1126.1875  395.6875 1765.3750 1582.7500  730.5000 1491.4375
#> [1436]  913.1250  974.0000  821.8125  456.5625 1095.7500  791.3750 1643.6250
#> [1443]  608.7500 1582.7500  639.1875 1765.3750  304.3750 1582.7500 1430.5625
#> [1450] 1613.1875 1461.0000 1278.3750 1095.7500  882.6875  913.1250  974.0000
#> [1457] 1187.0625 1308.8125 1156.6250  456.5625 1095.7500  913.1250  182.6250
#> [1464] 1643.6250  821.8125 1461.0000 1095.7500  547.8750 1491.4375  304.3750
#> [1471] 1795.8125  304.3750 1643.6250 1521.8750 1521.8750 1156.6250  700.0625
#> [1478] 1156.6250 1095.7500 1400.1250  273.9375  882.6875 1004.4375 1156.6250
#> [1485] 1126.1875  821.8125  487.0000  669.6250  456.5625  426.1250  487.0000
#> [1492]  578.3125 1643.6250  487.0000  334.8125 1034.8750 1095.7500  578.3125
#> [1499]  791.3750  456.5625 1734.9375 1095.7500  913.1250 1521.8750 1643.6250
#> [1506]  943.5625 1004.4375  426.1250  913.1250  852.2500 1034.8750 1643.6250
#> [1513] 1339.2500  791.3750  821.8125 1521.8750  913.1250  913.1250 1339.2500
#> [1520]  182.6250  913.1250  213.0625  791.3750 1217.5000  852.2500 1187.0625
#> [1527]  760.9375 1430.5625  943.5625  456.5625  304.3750 1674.0625 1217.5000
#> [1534]  273.9375 1126.1875 1369.6875 1308.8125  273.9375 1491.4375 1552.3125
#> [1541] 1339.2500  304.3750  547.8750 1034.8750  821.8125 1491.4375  943.5625
#> [1548] 1126.1875  213.0625  730.5000 1095.7500  760.9375 1187.0625 1400.1250
#> [1555] 1156.6250 1461.0000  882.6875  243.5000 1461.0000  334.8125  943.5625
#> [1562] 1491.4375  304.3750  182.6250  700.0625  517.4375  456.5625 1339.2500
#> [1569] 1674.0625 1491.4375 1339.2500  456.5625 1004.4375 1004.4375 1247.9375
#> [1576] 1400.1250 1126.1875  700.0625  243.5000 1095.7500  608.7500  487.0000
#> [1583] 1217.5000  821.8125 1156.6250  395.6875  487.0000  821.8125  974.0000
#> [1590]  517.4375 1156.6250  487.0000  547.8750 1674.0625  426.1250  791.3750
#> [1597]  821.8125 1095.7500  273.9375  426.1250  821.8125 1552.3125 1095.7500
#> [1604] 1552.3125 1217.5000 1278.3750 1613.1875  395.6875 1126.1875  760.9375
#> [1611] 1126.1875  487.0000 1187.0625  365.2500 1521.8750 1552.3125  487.0000
#> [1618] 1643.6250  669.6250  426.1250  487.0000 1795.8125 1308.8125  426.1250
#> [1625]  304.3750  608.7500 1034.8750  273.9375  882.6875  213.0625  608.7500
#> [1632] 1552.3125  608.7500  669.6250 1552.3125 1734.9375  791.3750  517.4375
#> [1639]  791.3750  304.3750 1126.1875 1126.1875 1187.0625 1521.8750  578.3125
#> [1646] 1247.9375  182.6250  426.1250 1095.7500 1400.1250  578.3125 1582.7500
#> [1653] 1704.5000  791.3750  487.0000  791.3750  608.7500 1369.6875 1369.6875
#> [1660]  243.5000 1339.2500 1126.1875  213.0625  821.8125  974.0000  547.8750
#> [1667]  669.6250 1369.6875  334.8125 1552.3125  365.2500 1582.7500  760.9375
#> [1674] 1278.3750 1461.0000  730.5000  487.0000 1187.0625  791.3750  791.3750
#> [1681] 1095.7500  456.5625 1126.1875  669.6250  760.9375 1156.6250  487.0000
#> [1688] 1095.7500 1339.2500 1095.7500 1095.7500 1704.5000 1552.3125  821.8125
#> [1695] 1704.5000  791.3750 1674.0625  821.8125  730.5000  273.9375  700.0625
#> [1702] 1187.0625 1491.4375 1004.4375 1795.8125  821.8125  395.6875  943.5625
#> [1709] 1339.2500 1004.4375 1095.7500  426.1250 1065.3125 1400.1250 1552.3125
#> [1716] 1521.8750  213.0625  974.0000  791.3750  791.3750  669.6250  395.6875
#> [1723] 1521.8750 1308.8125  243.5000 1247.9375        NA 1247.9375  395.6875
#> [1730]  791.3750 1187.0625 1461.0000  547.8750  578.3125 1187.0625  487.0000
#> [1737]  426.1250 1004.4375  517.4375 1095.7500 1369.6875 1004.4375  517.4375
#> [1744]  669.6250  791.3750 1765.3750  578.3125  974.0000  365.2500 1004.4375
#> [1751] 1034.8750  608.7500  821.8125 1278.3750  700.0625 1187.0625  821.8125
#> [1758]  669.6250 1582.7500  913.1250  821.8125 1339.2500 1369.6875 1430.5625
#> [1765]  730.5000  730.5000  821.8125 1156.6250  182.6250 1430.5625  456.5625
#> [1772]  334.8125 1187.0625 1126.1875  578.3125 1308.8125 1308.8125 1065.3125
#> [1779]  821.8125  974.0000  365.2500 1582.7500  213.0625 1126.1875 1521.8750
#> [1786]  273.9375  426.1250 1004.4375 1400.1250 1065.3125  395.6875 1582.7500
#> [1793]  700.0625 1187.0625 1217.5000 1095.7500  608.7500 1187.0625 1217.5000
#> [1800] 1156.6250 1369.6875 1461.0000 1704.5000 1004.4375 1034.8750  730.5000
#> [1807]  700.0625  608.7500 1400.1250  395.6875 1095.7500 1308.8125  821.8125
#> [1814] 1521.8750  395.6875  669.6250  760.9375 1400.1250  547.8750  578.3125
#> [1821]  456.5625 1613.1875 1491.4375 1765.3750  517.4375  517.4375 1339.2500
#> [1828] 1095.7500  639.1875 1126.1875  760.9375 1734.9375 1613.1875  304.3750
#> [1835]  821.8125  639.1875  700.0625  395.6875        NA  395.6875  791.3750
#> [1842] 1491.4375  487.0000 1004.4375  608.7500 1095.7500 1765.3750  426.1250
#> [1849]  487.0000  213.0625 1095.7500 1156.6250 1126.1875 1247.9375 1095.7500
#> [1856] 1704.5000  213.0625  669.6250  426.1250 1308.8125 1004.4375  334.8125
#> [1863] 1095.7500 1065.3125  334.8125 1400.1250  974.0000 1156.6250 1491.4375
#> [1870] 1126.1875  974.0000 1734.9375  730.5000 1308.8125  669.6250 1156.6250
#> [1877] 1734.9375 1552.3125 1156.6250 1461.0000  456.5625 1339.2500  821.8125
#> [1884]  639.1875  974.0000  426.1250 1095.7500 1034.8750  487.0000  456.5625
#> [1891] 1674.0625  395.6875 1004.4375 1613.1875 1095.7500  334.8125 1765.3750
#> [1898] 1552.3125 1461.0000 1521.8750 1795.8125 1339.2500  974.0000 1034.8750
#> [1905]  182.6250  700.0625 1795.8125  456.5625 1461.0000 1126.1875  791.3750
#> [1912] 1491.4375 1521.8750 1065.3125 1704.5000  365.2500  974.0000 1521.8750
#> [1919]  791.3750  517.4375 1674.0625 1004.4375 1521.8750 1369.6875  700.0625
#> [1926] 1247.9375  578.3125  821.8125 1461.0000 1065.3125  760.9375 1004.4375
#> [1933]  669.6250 1004.4375  821.8125  882.6875  821.8125  243.5000  243.5000
#> [1940] 1004.4375  456.5625 1400.1250  882.6875 1004.4375 1126.1875 1308.8125
#> [1947]  182.6250 1247.9375 1552.3125  547.8750  882.6875 1491.4375 1095.7500
#> [1954]  913.1250 1491.4375 1217.5000 1734.9375 1095.7500 1613.1875  547.8750
#> [1961] 1187.0625  821.8125 1552.3125 1704.5000 1034.8750 1126.1875  791.3750
#> [1968] 1095.7500 1126.1875  760.9375 1674.0625  791.3750 1613.1875 1156.6250
#> [1975] 1461.0000 1095.7500  700.0625 1034.8750  273.9375 1461.0000  517.4375
#> [1982]  913.1250 1095.7500 1461.0000 1187.0625 1521.8750  760.9375  334.8125
#> [1989]  243.5000  547.8750 1400.1250  700.0625  852.2500  821.8125 1461.0000
#> [1996] 1765.3750 1065.3125  243.5000  882.6875 1095.7500  395.6875  487.0000
#> [2003]  182.6250 1065.3125 1400.1250  213.0625 1095.7500  273.9375 1704.5000
#> [2010]  487.0000 1308.8125  243.5000  334.8125  760.9375  730.5000  730.5000
#> [2017] 1095.7500 1461.0000  182.6250  791.3750  821.8125  608.7500 1095.7500
#> [2024] 1095.7500 1034.8750  487.0000 1156.6250 1369.6875 1126.1875  517.4375
#> [2031]  639.1875  821.8125  639.1875  487.0000  974.0000  700.0625  547.8750
#> [2038] 1461.0000  182.6250 1065.3125 1582.7500  182.6250  365.2500  882.6875
#> [2045] 1552.3125  578.3125  456.5625 1491.4375  821.8125 1613.1875 1126.1875
#> [2052] 1308.8125  213.0625  913.1250  365.2500  791.3750  730.5000 1765.3750
#> [2059]  791.3750  608.7500  547.8750 1613.1875 1339.2500 1156.6250  669.6250
#> [2066]  821.8125  913.1250 1643.6250 1491.4375 1187.0625 1430.5625  213.0625
#> [2073] 1247.9375  213.0625  487.0000 1491.4375  395.6875 1217.5000 1430.5625
#> [2080] 1491.4375  913.1250 1674.0625 1369.6875  608.7500 1552.3125 1065.3125
#> [2087] 1004.4375  182.6250  365.2500 1308.8125 1369.6875 1400.1250  639.1875
#> [2094] 1613.1875 1065.3125  426.1250  243.5000 1004.4375 1126.1875  456.5625
#> [2101]  791.3750  243.5000  517.4375 1552.3125 1674.0625 1765.3750  821.8125
#> [2108] 1339.2500  547.8750 1065.3125 1552.3125  487.0000  882.6875  334.8125
#> [2115] 1034.8750  669.6250  882.6875 1247.9375 1430.5625  456.5625  791.3750
#> [2122] 1552.3125  882.6875 1491.4375 1156.6250  852.2500 1643.6250  639.1875
#> [2129]  365.2500  487.0000  760.9375 1217.5000  182.6250  182.6250  760.9375
#> [2136]  365.2500  700.0625  943.5625  487.0000  395.6875  791.3750  821.8125
#> [2143] 1247.9375 1674.0625 1369.6875 1461.0000        NA  243.5000  365.2500
#> [2150]  395.6875  273.9375 1613.1875 1156.6250 1095.7500 1126.1875 1461.0000
#> [2157]  913.1250  821.8125 1582.7500 1674.0625 1095.7500  760.9375 1187.0625
#> [2164] 1065.3125  547.8750  334.8125  182.6250  760.9375 1187.0625 1582.7500
#> [2171]  821.8125  608.7500 1065.3125 1491.4375  487.0000  426.1250 1643.6250
#> [2178]  669.6250  852.2500  456.5625        NA  578.3125 1491.4375  304.3750
#> [2185] 1369.6875  334.8125 1552.3125  517.4375  517.4375  852.2500  243.5000
#> [2192]  578.3125  456.5625 1430.5625  852.2500  456.5625  882.6875  852.2500
#> [2199] 1582.7500 1156.6250  456.5625  426.1250 1430.5625 1795.8125 1674.0625
#> [2206]  852.2500  821.8125 1521.8750 1582.7500  456.5625 1126.1875 1034.8750
#> [2213] 1461.0000  852.2500 1734.9375 1339.2500  852.2500  821.8125 1461.0000
#> [2220]  639.1875  730.5000  669.6250 1491.4375  882.6875  669.6250  182.6250
#> [2227]  760.9375 1704.5000  669.6250 1369.6875 1187.0625 1095.7500 1217.5000
#> [2234]  517.4375  852.2500 1461.0000  730.5000 1187.0625  365.2500 1095.7500
#> [2241] 1734.9375        NA 1582.7500 1095.7500 1187.0625 1461.0000 1491.4375
#> [2248] 1095.7500 1734.9375 1004.4375 1217.5000  852.2500 1461.0000  334.8125
#> [2255]  365.2500  243.5000  547.8750  791.3750 1187.0625  334.8125  700.0625
#> [2262] 1582.7500 1339.2500  913.1250 1582.7500 1339.2500  395.6875 1126.1875
#> [2269] 1034.8750  730.5000  700.0625  974.0000  760.9375  608.7500 1491.4375
#> [2276] 1461.0000 1461.0000  304.3750 1552.3125  730.5000 1187.0625  669.6250
#> [2283] 1674.0625  760.9375  517.4375  487.0000  547.8750 1582.7500  760.9375
#> [2290] 1065.3125 1400.1250  578.3125  639.1875 1765.3750 1400.1250  547.8750
#> [2297]  791.3750  365.2500  730.5000 1095.7500 1795.8125 1034.8750  395.6875
#> [2304]  213.0625 1491.4375  760.9375 1187.0625 1491.4375  578.3125 1278.3750
#> [2311] 1339.2500 1339.2500 1491.4375 1156.6250  730.5000 1095.7500  243.5000
#> [2318]  243.5000 1065.3125 1156.6250 1187.0625  456.5625  730.5000 1156.6250
#> [2325] 1643.6250 1217.5000  974.0000  304.3750  365.2500  852.2500  273.9375
#> [2332]  974.0000 1643.6250 1369.6875  760.9375 1461.0000  669.6250 1613.1875
#> [2339] 1217.5000  487.0000 1613.1875  608.7500  669.6250 1461.0000  426.1250
#> [2346]  334.8125 1765.3750  243.5000 1065.3125 1461.0000 1004.4375 1004.4375
#> [2353]  943.5625 1613.1875 1095.7500 1187.0625  334.8125  821.8125  700.0625
#> [2360] 1461.0000  517.4375 1613.1875 1095.7500  487.0000 1004.4375  365.2500
#> [2367]  852.2500 1521.8750  547.8750 1126.1875  974.0000 1400.1250  182.6250
#> [2374] 1461.0000  760.9375 1339.2500  700.0625  365.2500  334.8125 1095.7500
#> [2381]  578.3125 1095.7500 1095.7500 1004.4375  639.1875  213.0625  700.0625
#> [2388]  395.6875 1095.7500  913.1250  334.8125  456.5625 1461.0000  852.2500
#> [2395]  517.4375  943.5625  182.6250 1156.6250 1430.5625 1339.2500 1034.8750
#> [2402]  821.8125 1461.0000  821.8125  273.9375 1339.2500 1521.8750 1795.8125
#> [2409] 1187.0625  791.3750  760.9375  730.5000  213.0625 1095.7500 1095.7500
#> [2416]  365.2500 1339.2500 1187.0625  365.2500  487.0000  365.2500  760.9375
#> [2423] 1674.0625  213.0625 1369.6875  547.8750  365.2500  882.6875 1004.4375
#> [2430]  304.3750 1095.7500 1065.3125  395.6875  974.0000  547.8750  700.0625
#> [2437] 1095.7500 1582.7500 1674.0625 1582.7500  578.3125  852.2500 1582.7500
#> [2444] 1613.1875  334.8125  882.6875 1065.3125  669.6250  852.2500  213.0625
#> [2451]  760.9375 1491.4375 1765.3750 1034.8750  487.0000 1795.8125 1004.4375
#> [2458]  791.3750  426.1250 1795.8125  852.2500 1004.4375 1278.3750  213.0625
#> [2465]  426.1250 1552.3125  669.6250  517.4375  456.5625  273.9375 1339.2500
#> [2472] 1461.0000 1491.4375 1734.9375  365.2500  730.5000 1339.2500 1156.6250
#> [2479] 1521.8750  395.6875  760.9375 1704.5000 1247.9375  273.9375 1095.7500
#> [2486]  608.7500 1674.0625  243.5000 1217.5000  669.6250  395.6875  547.8750
#> [2493]  821.8125  669.6250 1521.8750 1430.5625 1552.3125 1521.8750 1065.3125
#> [2500]  213.0625 1126.1875 1247.9375  426.1250  669.6250 1643.6250  760.9375
#> [2507] 1704.5000 1369.6875 1187.0625  213.0625  608.7500  700.0625  578.3125
#> [2514] 1247.9375 1339.2500  365.2500  182.6250  700.0625 1004.4375 1521.8750
#> [2521]  456.5625  426.1250  821.8125 1613.1875 1126.1875  426.1250 1004.4375
#> [2528] 1613.1875  608.7500 1065.3125 1217.5000  700.0625  700.0625 1674.0625
#> [2535]  456.5625 1582.7500 1065.3125  791.3750 1430.5625 1004.4375  852.2500
#> [2542]  182.6250  426.1250 1278.3750  304.3750 1004.4375 1430.5625 1765.3750
#> [2549] 1674.0625 1156.6250 1034.8750        NA  426.1250 1126.1875 1126.1875
#> [2556] 1552.3125 1369.6875  273.9375 1156.6250 1430.5625 1765.3750 1491.4375
#> [2563] 1552.3125 1095.7500 1187.0625 1217.5000  456.5625  608.7500 1643.6250
#> [2570] 1674.0625 1643.6250  852.2500 1126.1875  882.6875 1795.8125 1521.8750
#> [2577] 1156.6250  639.1875  700.0625 1674.0625 1095.7500 1126.1875  608.7500
#> [2584]  304.3750 1795.8125  821.8125 1217.5000  487.0000  456.5625 1765.3750
#> [2591]  547.8750  791.3750 1217.5000 1126.1875 1765.3750 1156.6250  517.4375
#> [2598] 1491.4375 1461.0000 1065.3125  273.9375  578.3125 1217.5000 1430.5625
#> [2605] 1734.9375 1034.8750  365.2500 1552.3125 1521.8750 1613.1875 1734.9375
#> [2612]  487.0000  426.1250  669.6250  639.1875 1278.3750 1461.0000 1369.6875
#> [2619]  700.0625 1156.6250  243.5000  426.1250  395.6875 1004.4375 1582.7500
#> [2626] 1369.6875 1430.5625 1521.8750 1400.1250  700.0625 1461.0000 1217.5000
#> [2633] 1187.0625 1582.7500 1034.8750  700.0625 1247.9375  760.9375 1552.3125
#> [2640]  426.1250  578.3125 1491.4375 1156.6250  821.8125 1795.8125 1704.5000
#> [2647] 1613.1875  669.6250 1765.3750 1065.3125 1156.6250  943.5625 1674.0625
#> [2654]  243.5000  974.0000 1278.3750  669.6250 1217.5000  943.5625  243.5000
#> [2661] 1156.6250  213.0625 1582.7500  243.5000  700.0625  821.8125 1034.8750
#> [2668] 1156.6250  821.8125 1156.6250  304.3750 1156.6250 1430.5625 1795.8125
#> [2675] 1674.0625 1126.1875 1308.8125  974.0000 1461.0000 1430.5625 1400.1250
#> [2682]  304.3750  487.0000 1461.0000  882.6875  943.5625  760.9375 1095.7500
#> [2689]  608.7500 1187.0625 1034.8750 1461.0000  760.9375  213.0625 1521.8750
#> [2696]  882.6875  426.1250  487.0000  730.5000 1582.7500  273.9375  213.0625
#> [2703] 1065.3125 1095.7500  243.5000  334.8125 1674.0625 1552.3125 1065.3125
#> [2710]        NA 1156.6250 1491.4375  334.8125 1521.8750  760.9375 1765.3750
#> [2717]  395.6875  243.5000 1156.6250 1095.7500 1430.5625  730.5000  730.5000
#> [2724]  243.5000  760.9375 1369.6875  243.5000 1795.8125  730.5000  913.1250
#> [2731]  182.6250 1187.0625  700.0625  243.5000  700.0625 1095.7500  700.0625
#> [2738]  273.9375  334.8125  547.8750  821.8125  821.8125  730.5000 1400.1250
#> [2745]  974.0000  243.5000  730.5000 1461.0000 1034.8750 1369.6875 1065.3125
#> [2752]  487.0000 1247.9375  669.6250  700.0625 1643.6250 1734.9375  791.3750
#> [2759] 1521.8750 1247.9375  821.8125 1095.7500  608.7500  730.5000 1521.8750
#> [2766]  730.5000 1765.3750 1521.8750 1521.8750  426.1250  487.0000  639.1875
#> [2773] 1734.9375  426.1250  791.3750  608.7500  669.6250  913.1250 1521.8750
#> [2780] 1278.3750 1247.9375 1643.6250  639.1875        NA 1217.5000  608.7500
#> [2787]  730.5000  456.5625  913.1250  395.6875  882.6875  852.2500 1582.7500
#> [2794]  639.1875  700.0625 1521.8750  760.9375 1552.3125 1065.3125  882.6875
#> [2801]  700.0625  334.8125 1308.8125 1613.1875 1156.6250  273.9375 1491.4375
#> [2808] 1156.6250 1521.8750 1613.1875 1461.0000  791.3750  304.3750 1734.9375
#> [2815]  304.3750 1156.6250  974.0000 1187.0625 1613.1875 1278.3750  304.3750
#> [2822]  669.6250 1187.0625 1521.8750  395.6875  487.0000 1582.7500 1582.7500
#> [2829] 1156.6250  395.6875 1430.5625 1126.1875 1369.6875  273.9375 1126.1875
#> [2836] 1430.5625  273.9375  365.2500 1400.1250 1095.7500 1400.1250  760.9375
#> [2843]  243.5000  700.0625 1582.7500 1034.8750 1278.3750  243.5000 1156.6250
#> [2850] 1004.4375 1582.7500  821.8125  365.2500 1613.1875  608.7500 1430.5625
#> [2857] 1156.6250 1187.0625  456.5625  -91.3125 1217.5000 1400.1250 1308.8125
#> [2864]  578.3125 1156.6250 1156.6250  487.0000 1400.1250  639.1875  243.5000
#> [2871] 1582.7500 1156.6250  426.1250 1582.7500  821.8125 1369.6875  487.0000
#> [2878] 1217.5000  578.3125  395.6875  669.6250 1430.5625  974.0000  273.9375
#> [2885]  182.6250  943.5625        NA  852.2500        NA  365.2500  608.7500
#> [2892]  639.1875 1461.0000  669.6250 1187.0625 1004.4375        NA  669.6250
#> [2899]        NA 1126.1875 1126.1875 1765.3750 1095.7500  243.5000        NA
#> [2906]  365.2500        NA  213.0625 1339.2500  456.5625 1065.3125 1400.1250
#> [2913]  487.0000        NA        NA        NA 1187.0625 1613.1875 1217.5000
#> [2920]  426.1250  395.6875 1521.8750 1278.3750 1369.6875        NA 1217.5000
#> [2927] 1734.9375  517.4375 1521.8750 1278.3750 1400.1250 1217.5000 1339.2500
#> [2934] 1187.0625  304.3750  334.8125  487.0000 1582.7500 1491.4375  821.8125
#> [2941]  304.3750 1430.5625  821.8125 1582.7500  487.0000 1369.6875 1552.3125
#> [2948] 1247.9375 1734.9375  943.5625 1430.5625 1734.9375 1400.1250  213.0625
#> [2955] 1095.7500 1430.5625  974.0000  426.1250  608.7500 1430.5625 1034.8750
#> [2962]  182.6250  913.1250  821.8125  882.6875  852.2500  821.8125 1095.7500
#> [2969]  304.3750 1308.8125  882.6875  487.0000  243.5000 1095.7500 1552.3125
#> [2976]  760.9375  882.6875 1521.8750  456.5625  395.6875 1491.4375 1430.5625
#> [2983] 1674.0625  456.5625 1095.7500 1217.5000 1187.0625  791.3750  760.9375
#> [2990] 1400.1250  913.1250  182.6250  395.6875 1704.5000  821.8125 1461.0000
#> [2997]  304.3750  730.5000 1217.5000  456.5625  730.5000 1461.0000 1461.0000
#> [3004] 1095.7500 1065.3125  547.8750  791.3750 1582.7500 1582.7500 1217.5000
#> [3011] 1613.1875  517.4375  456.5625  517.4375 1095.7500 1065.3125  730.5000
#> [3018] 1095.7500 1521.8750 1461.0000  974.0000 1156.6250 1582.7500 1156.6250
#> [3025]  913.1250 1095.7500  974.0000  943.5625  213.0625  487.0000  517.4375
#> [3032] 1126.1875  365.2500  395.6875  760.9375 1126.1875 1126.1875 1552.3125
#> [3039]  426.1250 1217.5000 1187.0625 1521.8750  882.6875  426.1250 1552.3125
#> [3046] 1552.3125  730.5000 1095.7500  243.5000 1400.1250 1156.6250 1187.0625
#> [3053] 1247.9375  669.6250 1339.2500 1217.5000 1187.0625 1308.8125 1552.3125
#> [3060]  487.0000  243.5000  456.5625  456.5625 1369.6875  456.5625 1217.5000
#> [3067]  730.5000  700.0625  243.5000  182.6250 1187.0625  882.6875 1217.5000
#> [3074]  700.0625  943.5625 1491.4375 1278.3750 1400.1250 1430.5625 1582.7500
#> [3081] 1004.4375 1552.3125  852.2500 1247.9375 1004.4375 1613.1875  213.0625
#> [3088] 1187.0625 1004.4375 1187.0625 1065.3125 1156.6250 1582.7500  760.9375
#> [3095]  395.6875  608.7500  273.9375  821.8125 1369.6875 1126.1875  760.9375
#> [3102] 1004.4375  456.5625  974.0000 1126.1875  639.1875 1156.6250  821.8125
#> [3109] 1308.8125  273.9375 1126.1875  517.4375 1674.0625  700.0625 1065.3125
#> [3116]  334.8125  791.3750 1674.0625 1491.4375  882.6875 1552.3125  913.1250
#> [3123] 1187.0625  243.5000 1795.8125 1613.1875  760.9375  821.8125 1217.5000
#> [3130]  669.6250  760.9375 1095.7500 1095.7500 1400.1250  913.1250 1552.3125
#> [3137] 1034.8750 1795.8125 1491.4375 1582.7500 1674.0625 1461.0000  517.4375
#> [3144]  760.9375  669.6250 1156.6250  760.9375 1461.0000 1461.0000 1217.5000
#> [3151] 1217.5000  608.7500 1187.0625 1095.7500 1126.1875  304.3750  273.9375
#> [3158] 1278.3750  791.3750 1552.3125  760.9375  395.6875  760.9375 1065.3125
#> [3165]  273.9375  608.7500  517.4375  821.8125 1795.8125 1582.7500 1247.9375
#> [3172] 1004.4375  608.7500  974.0000  547.8750 1461.0000 1491.4375  760.9375
#> [3179] 1004.4375 1126.1875  730.5000 1552.3125 1461.0000  517.4375 1187.0625
#> [3186]  791.3750  213.0625  608.7500  517.4375  334.8125  882.6875 1521.8750
#> [3193] 1582.7500  669.6250 1217.5000 1217.5000 1004.4375  456.5625 1674.0625
#> [3200]  456.5625  852.2500 1217.5000 1156.6250 1552.3125  821.8125 1187.0625
#> [3207]  821.8125 1034.8750  791.3750 1034.8750  791.3750  943.5625  426.1250
#> [3214]  334.8125 1369.6875 1400.1250  243.5000  365.2500 1521.8750 1369.6875
#> [3221]  913.1250  334.8125  578.3125  608.7500 1643.6250  760.9375  517.4375
#> [3228]  426.1250 1156.6250 1247.9375 1217.5000 1643.6250 1247.9375 1004.4375
#> [3235] 1247.9375 1278.3750 1674.0625  273.9375 1369.6875  700.0625 1034.8750
#> [3242]  882.6875  213.0625 1308.8125 1004.4375  974.0000  395.6875  791.3750
#> [3249] 1674.0625  487.0000 1734.9375  365.2500 1187.0625 1247.9375 1461.0000
#> [3256]  426.1250  760.9375  700.0625 2039.3125  426.1250  974.0000 1369.6875
#> [3263]  791.3750  334.8125  395.6875 1034.8750 1521.8750 1461.0000  273.9375
#> [3270]  273.9375 1795.8125 1095.7500 1156.6250 1765.3750  760.9375 1613.1875
#> [3277]  913.1250  791.3750        NA  304.3750 1156.6250  395.6875  182.6250
#> [3284] 1095.7500  334.8125  487.0000  273.9375 1065.3125 1217.5000  487.0000
#> [3291] 1552.3125 1126.1875  639.1875  395.6875 1643.6250 1339.2500 1095.7500
#> [3298]  882.6875  517.4375  182.6250  791.3750  334.8125 1156.6250 1247.9375
#> [3305]  426.1250  547.8750 1217.5000 1461.0000 1491.4375 1095.7500  273.9375
#> [3312]  608.7500  213.0625  517.4375 1491.4375  334.8125  791.3750 1461.0000
#> [3319] 1187.0625  456.5625 1095.7500  852.2500 1461.0000  730.5000  213.0625
#> [3326] 1095.7500  852.2500  821.8125 1004.4375  669.6250  426.1250  730.5000
#> [3333]  730.5000  730.5000  182.6250  882.6875 1095.7500  760.9375  730.5000
#> [3340]  243.5000 1034.8750 1247.9375  608.7500 1034.8750  456.5625  700.0625
#> [3347] 1308.8125  730.5000  943.5625 1521.8750 1552.3125  578.3125 1065.3125
#> [3354]  273.9375 1217.5000  760.9375 1613.1875  639.1875  760.9375  882.6875
#> [3361] 1521.8750 1065.3125 1491.4375  608.7500  487.0000 1217.5000 1278.3750
#> [3368] 1369.6875  700.0625  426.1250  639.1875  821.8125 1126.1875  304.3750
#> [3375]  943.5625 1461.0000  700.0625  791.3750 1369.6875 1156.6250 1247.9375
#> [3382] 1308.8125 1126.1875  182.6250 1034.8750  608.7500 1187.0625  334.8125
#> [3389] 1339.2500  517.4375 1704.5000  547.8750  487.0000 1765.3750  821.8125
#> [3396] 1187.0625 1613.1875 1643.6250 1430.5625  456.5625  213.0625  821.8125
#> [3403] 1369.6875  608.7500 1461.0000 1095.7500  760.9375 1034.8750  517.4375
#> [3410]  974.0000  395.6875  334.8125  913.1250  304.3750 1461.0000  426.1250
#> [3417]  852.2500  608.7500  365.2500 1461.0000 1004.4375  395.6875 1704.5000
#> [3424]  578.3125  730.5000