Function to recode age variable and convert to age in days

Function to recode age variable and convert to age in days

recode_age(df)

Arguments

df
A coverage data.frame collected for the Liberia Coverage Survey

Value

A vector of recoded ages based on any re-testing done

Examples

recode_age(df = coverageData.r2)

#>    [1]  304.3750 1217.5000  882.6875 1582.7500  760.9375  760.9375 1582.7500
#>    [8]  669.6250  913.1250 1461.0000  456.5625  852.2500 1461.0000 1278.3750
#>   [15]  882.6875  395.6875 1582.7500 1187.0625  882.6875  852.2500  578.3125
#>   [22] 1187.0625 1765.3750  273.9375 1491.4375  243.5000 1704.5000  821.8125
#>   [29] 1004.4375  243.5000 1126.1875  426.1250 1156.6250 1247.9375 1187.0625
#>   [36]  882.6875 1278.3750  608.7500 1095.7500 1674.0625 1369.6875 1704.5000
#>   [43]  821.8125  821.8125 1126.1875  547.8750  578.3125  821.8125  426.1250
#>   [50]  334.8125  821.8125 1674.0625  791.3750  334.8125  487.0000  304.3750
#>   [57]  974.0000  426.1250 1339.2500  608.7500  273.9375 1582.7500 1156.6250
#>   [64] 1704.5000 1400.1250 1552.3125 1491.4375  760.9375  791.3750  334.8125
#>   [71]  913.1250 1491.4375 1552.3125 1217.5000 1126.1875  852.2500  547.8750
#>   [78] 1400.1250  182.6250  639.1875 1126.1875 1034.8750 1491.4375 1643.6250
#>   [85] 1552.3125  639.1875  791.3750 1065.3125  334.8125 1430.5625  547.8750
#>   [92]  913.1250 1521.8750 1613.1875  395.6875 1308.8125  791.3750 1004.4375
#>   [99]  791.3750  852.2500  882.6875 1400.1250 1065.3125  517.4375 1247.9375
#>  [106] 1521.8750 1217.5000  821.8125  821.8125  334.8125 1552.3125  456.5625
#>  [113]  700.0625 1126.1875  791.3750 1552.3125  791.3750 1247.9375  608.7500
#>  [120]  395.6875 1369.6875  456.5625 1217.5000 1065.3125  639.1875 1126.1875
#>  [127]  426.1250 1461.0000 1582.7500  182.6250 1552.3125  365.2500 1126.1875
#>  [134]  760.9375  456.5625 1521.8750  730.5000  456.5625  304.3750  213.0625
#>  [141] 1369.6875 1126.1875 1339.2500  943.5625  334.8125 1461.0000 1461.0000
#>  [148]  821.8125 1430.5625 1521.8750  395.6875 1156.6250  730.5000 1339.2500
#>  [155] 1004.4375  882.6875 1339.2500 1369.6875  882.6875  608.7500  882.6875
#>  [162]  760.9375  943.5625  304.3750  213.0625  395.6875 1187.0625  852.2500
#>  [169] 1552.3125  639.1875 1187.0625 1369.6875 1491.4375  608.7500  456.5625
#>  [176]  456.5625 1187.0625  882.6875  395.6875  760.9375  791.3750  974.0000
#>  [183] 1278.3750 1034.8750 1430.5625 1491.4375  821.8125 1582.7500  456.5625
#>  [190] 1247.9375 1065.3125  700.0625  821.8125 1643.6250 1156.6250 1765.3750
#>  [197]  852.2500 1734.9375  608.7500 1734.9375  517.4375 1734.9375  273.9375
#>  [204] 1491.4375 1643.6250 1643.6250  974.0000  730.5000 1095.7500  547.8750
#>  [211]  334.8125  304.3750 1491.4375  426.1250 1065.3125  456.5625  669.6250
#>  [218]  760.9375 1613.1875 1065.3125  791.3750 1430.5625  791.3750  639.1875
#>  [225]  882.6875 1369.6875 1552.3125 1704.5000  456.5625 1704.5000  334.8125
#>  [232] 1004.4375  578.3125  365.2500 1217.5000  395.6875  243.5000 1582.7500
#>  [239] 1552.3125  730.5000  304.3750 1613.1875  882.6875  334.8125  334.8125
#>  [246]        NA  426.1250  487.0000 1461.0000  213.0625 1126.1875  365.2500
#>  [253]  639.1875  182.6250  243.5000 1430.5625  760.9375 1217.5000  608.7500
#>  [260] 1065.3125 1247.9375 1643.6250  243.5000  304.3750 1065.3125  243.5000
#>  [267]  273.9375 1704.5000 1095.7500  517.4375  243.5000 1004.4375 1521.8750
#>  [274]  365.2500 1795.8125 1582.7500  456.5625 1247.9375 1795.8125 1247.9375
#>  [281]  882.6875  395.6875  365.2500 1156.6250 1065.3125  426.1250  760.9375
#>  [288] 1217.5000  273.9375  791.3750  760.9375  487.0000 1430.5625  487.0000
#>  [295]  639.1875 1247.9375 1521.8750 1369.6875 1308.8125  487.0000  852.2500
#>  [302] 1339.2500 1521.8750  456.5625 1126.1875 1369.6875  791.3750  487.0000
#>  [309] 1734.9375 1369.6875 1613.1875  943.5625  395.6875  730.5000 1704.5000
#>  [316] 1521.8750 1126.1875  456.5625  487.0000 1491.4375 1095.7500  243.5000
#>  [323] 1156.6250 1704.5000 1034.8750 1156.6250 1095.7500 1734.9375 1521.8750
#>  [330]  913.1250  608.7500 1156.6250 1369.6875  943.5625  730.5000  182.6250
#>  [337]  821.8125 1765.3750  608.7500 1491.4375 1339.2500 1369.6875  760.9375
#>  [344]  334.8125  426.1250 1582.7500 1369.6875 1126.1875  243.5000 1461.0000
#>  [351]  487.0000 1400.1250  334.8125  426.1250  456.5625 1491.4375  243.5000
#>  [358] 1430.5625  365.2500 1704.5000 1095.7500 1643.6250  882.6875 1552.3125
#>  [365]  213.0625  243.5000  913.1250  456.5625  395.6875 1247.9375 1552.3125
#>  [372]  456.5625 1247.9375  639.1875 1156.6250  974.0000  669.6250 1156.6250
#>  [379] 1521.8750 1521.8750  517.4375  730.5000 1400.1250 1613.1875        NA
#>  [386] 1278.3750 1004.4375  821.8125 1582.7500 1461.0000  243.5000 1552.3125
#>  [393] 1156.6250 1400.1250 1643.6250 1400.1250  273.9375 1004.4375 1126.1875
#>  [400]  426.1250 1582.7500  852.2500 1278.3750  213.0625  426.1250 1369.6875
#>  [407] 1339.2500 1613.1875  365.2500  608.7500 1521.8750  913.1250  456.5625
#>  [414]  517.4375 1156.6250 1400.1250  456.5625  426.1250  578.3125  913.1250
#>  [421]  669.6250  791.3750 1491.4375  426.1250  547.8750  669.6250 1491.4375
#>  [428]  426.1250  730.5000  273.9375  395.6875 1369.6875  547.8750  669.6250
#>  [435] 1095.7500 1643.6250  943.5625 1674.0625  852.2500  913.1250  426.1250
#>  [442] 1734.9375 1734.9375  395.6875 1217.5000 1339.2500 1126.1875  487.0000
#>  [449] 1156.6250  426.1250  213.0625  334.8125  669.6250  791.3750 1704.5000
#>  [456] 1339.2500 1521.8750 1034.8750 1491.4375 1308.8125  608.7500  334.8125
#>  [463] 1795.8125  821.8125 2039.3125  700.0625 1704.5000  913.1250  882.6875
#>  [470]  852.2500  213.0625  304.3750 1004.4375  913.1250  517.4375 1582.7500
#>  [477] 1765.3750 1095.7500  243.5000  669.6250 1430.5625  730.5000 1643.6250
#>  [484]  182.6250  304.3750 1065.3125  730.5000  517.4375 1734.9375  913.1250
#>  [491] 1704.5000  882.6875  882.6875 1400.1250 1004.4375        NA  882.6875
#>  [498] 1187.0625 1674.0625  608.7500  487.0000 1430.5625 1004.4375  913.1250
#>  [505] 1065.3125 1217.5000 1065.3125 1613.1875 1004.4375  456.5625  730.5000
#>  [512] 1734.9375 1034.8750  608.7500  791.3750 1430.5625  487.0000  213.0625
#>  [519]  395.6875 1521.8750 1247.9375  456.5625  273.9375 1187.0625  182.6250
#>  [526] 1004.4375  821.8125  578.3125  365.2500  821.8125 1704.5000  882.6875
#>  [533]  456.5625  273.9375  760.9375  700.0625  487.0000 1187.0625 1734.9375
#>  [540]  395.6875 1461.0000  213.0625  700.0625 1095.7500  882.6875  852.2500
#>  [547] 1217.5000 1521.8750  487.0000 1400.1250 1126.1875 1278.3750  426.1250
#>  [554]  456.5625  974.0000 1582.7500  639.1875 1674.0625  426.1250  791.3750
#>  [561] 1400.1250  943.5625 1734.9375  608.7500  760.9375  456.5625  456.5625
#>  [568]  608.7500 1461.0000  943.5625  913.1250  365.2500  730.5000  943.5625
#>  [575] 1704.5000 1734.9375  213.0625  395.6875 1430.5625 1552.3125  456.5625
#>  [582]  608.7500  487.0000  334.8125 1126.1875  821.8125 1308.8125  517.4375
#>  [589]  669.6250  669.6250 1095.7500  547.8750 1582.7500 1126.1875 1369.6875
#>  [596] 1308.8125 1521.8750 1521.8750  182.6250  456.5625 1004.4375  395.6875
#>  [603] 1491.4375  730.5000 1613.1875  882.6875 1247.9375 1521.8750  487.0000
#>  [610] 1461.0000 1187.0625  700.0625 1430.5625 1308.8125 1339.2500 1521.8750
#>  [617]  882.6875  304.3750  517.4375  395.6875 1187.0625  700.0625 1674.0625
#>  [624] 1704.5000  547.8750  304.3750 1065.3125 1430.5625 1369.6875  608.7500
#>  [631]  578.3125 1217.5000  517.4375  365.2500  578.3125 1004.4375  456.5625
#>  [638] 1521.8750 1034.8750  882.6875  273.9375  304.3750 1339.2500 1004.4375
#>  [645] 1613.1875 1461.0000  426.1250 1278.3750  760.9375 1156.6250 1308.8125
#>  [652]  700.0625 1034.8750  791.3750 1278.3750 1095.7500 1126.1875 1156.6250
#>  [659] 1308.8125 1095.7500  517.4375  365.2500 1491.4375  243.5000  913.1250
#>  [666]  639.1875  639.1875  760.9375  547.8750  304.3750 1582.7500  852.2500
#>  [673] 1095.7500 1187.0625 1156.6250 1430.5625  243.5000  943.5625 1491.4375
#>  [680]  213.0625  365.2500 1674.0625 1461.0000  882.6875 1217.5000  608.7500
#>  [687]  517.4375  395.6875  821.8125 1643.6250  639.1875 1461.0000  974.0000
#>  [694]  456.5625  791.3750  395.6875 1430.5625  639.1875 1034.8750  669.6250
#>  [701]  913.1250 1126.1875 1247.9375  487.0000 1095.7500  395.6875  730.5000
#>  [708]  334.8125 1308.8125  852.2500  730.5000 1278.3750  882.6875 1126.1875
#>  [715]  547.8750  882.6875 1034.8750 1095.7500 1187.0625  547.8750  608.7500
#>  [722] 1765.3750 1308.8125  365.2500  730.5000  882.6875  821.8125  852.2500
#>  [729]  730.5000 1734.9375 1308.8125 1126.1875  821.8125  974.0000  791.3750
#>  [736]  182.6250  974.0000 1430.5625 1156.6250 1004.4375  882.6875 1613.1875
#>  [743]  974.0000 1369.6875 1430.5625  182.6250  517.4375  334.8125  791.3750
#>  [750] 1552.3125  365.2500 1521.8750 1278.3750  791.3750 1430.5625  578.3125
#>  [757] 1339.2500  243.5000 1339.2500 1461.0000 1187.0625 1156.6250  669.6250
#>  [764]  304.3750 1065.3125  791.3750 1461.0000  669.6250 1217.5000 1521.8750
#>  [771]        NA 1095.7500 1247.9375 1582.7500 1278.3750  365.2500 1491.4375
#>  [778] 1126.1875  365.2500  487.0000  365.2500 1095.7500 1095.7500  456.5625
#>  [785] 1461.0000  273.9375  213.0625        NA 1521.8750  365.2500 1461.0000
#>  [792]  487.0000 1004.4375  760.9375  882.6875  456.5625  730.5000 1552.3125
#>  [799]  608.7500 1461.0000  669.6250  760.9375 1217.5000 1521.8750  334.8125
#>  [806]  182.6250  304.3750 1156.6250  943.5625 1461.0000 1217.5000  547.8750
#>  [813]  669.6250  334.8125  821.8125 1217.5000  243.5000  882.6875  487.0000
#>  [820]  608.7500  334.8125  821.8125 1704.5000  517.4375  578.3125  395.6875
#>  [827]  273.9375  760.9375  913.1250 1308.8125  395.6875  213.0625 1491.4375
#>  [834]  273.9375  669.6250 1643.6250 1278.3750  304.3750  852.2500        NA
#>  [841] 1613.1875  821.8125  213.0625  395.6875  852.2500  852.2500 1430.5625
#>  [848]  304.3750 1461.0000 1521.8750 1613.1875 1734.9375  304.3750  426.1250
#>  [855] 1491.4375  700.0625  182.6250  426.1250 1217.5000 1034.8750  456.5625
#>  [862] 1643.6250  882.6875  578.3125 1156.6250  395.6875  547.8750 1247.9375
#>  [869]  456.5625 1217.5000  730.5000  304.3750  669.6250  943.5625  273.9375
#>  [876]  943.5625 1278.3750  974.0000 1582.7500  487.0000 1187.0625 1734.9375
#>  [883] 1156.6250  395.6875 1704.5000 1126.1875 1369.6875  182.6250  821.8125
#>  [890]  487.0000 1795.8125  243.5000 1643.6250 1034.8750 1613.1875  791.3750
#>  [897] 1126.1875  487.0000  974.0000  608.7500  547.8750 1126.1875  456.5625
#>  [904] 1339.2500  487.0000 1126.1875  730.5000  760.9375  700.0625 1004.4375
#>  [911]  426.1250 1339.2500 1126.1875 1065.3125  669.6250 1065.3125 1278.3750
#>  [918]  760.9375 1004.4375  760.9375  730.5000 1095.7500  669.6250 1400.1250
#>  [925]  304.3750 1552.3125  791.3750 1004.4375 1795.8125  791.3750  639.1875
#>  [932] 1369.6875  517.4375  517.4375 1004.4375 1278.3750 1065.3125 1613.1875
#>  [939] 1582.7500  517.4375 1491.4375  974.0000  243.5000 1095.7500 1095.7500
#>  [946] 1795.8125 1400.1250 1034.8750  913.1250  791.3750  791.3750 1582.7500
#>  [953]  730.5000 1308.8125 1400.1250  487.0000 1217.5000 1187.0625  365.2500
#>  [960] 1491.4375  974.0000 1126.1875 1430.5625 1461.0000 1187.0625  365.2500
#>  [967]  852.2500 1034.8750  182.6250 1339.2500 1339.2500 1461.0000 1795.8125
#>  [974] 1582.7500  852.2500 1582.7500 1674.0625 1521.8750 1217.5000 1582.7500
#>  [981] 1461.0000        NA  578.3125 1491.4375 1369.6875 1734.9375 1643.6250
#>  [988] 1461.0000  669.6250 1095.7500 1734.9375  517.4375 1156.6250  273.9375
#>  [995] 1156.6250 1308.8125 1065.3125 1552.3125  852.2500  882.6875  365.2500
#> [1002]  730.5000  730.5000 1430.5625 1369.6875  456.5625  882.6875  304.3750
#> [1009] 1704.5000  821.8125  243.5000  395.6875 1095.7500  426.1250  669.6250
#> [1016]  669.6250  578.3125  426.1250 1217.5000  760.9375  365.2500 1461.0000
#> [1023]  426.1250  182.6250 1308.8125 1156.6250 1278.3750  913.1250  243.5000
#> [1030] 1491.4375  852.2500  852.2500 1156.6250 1430.5625  578.3125 1004.4375
#> [1037]  669.6250 1552.3125 1674.0625 1521.8750 1247.9375  426.1250 1521.8750
#> [1044] 1217.5000  700.0625 1095.7500  974.0000 1126.1875  456.5625  304.3750
#> [1051] 1461.0000 1278.3750  700.0625 1126.1875  852.2500  273.9375 1430.5625
#> [1058]  456.5625 1400.1250 1461.0000 1491.4375 1065.3125 1217.5000  456.5625
#> [1065] 1156.6250 1308.8125  182.6250  304.3750  730.5000 1643.6250  730.5000
#> [1072]  334.8125 1095.7500 1095.7500 1187.0625  395.6875 1369.6875 1704.5000
#> [1079] 1674.0625  304.3750 1095.7500 1339.2500  760.9375  639.1875 1034.8750
#> [1086] 1247.9375 1461.0000 1278.3750 1369.6875 1461.0000  213.0625  821.8125
#> [1093]  243.5000 1187.0625 1369.6875 1126.1875 1613.1875 1095.7500  456.5625
#> [1100] 1491.4375 1247.9375  821.8125  882.6875  821.8125 1430.5625 1461.0000
#> [1107] 1156.6250 1156.6250 1461.0000  456.5625  578.3125  456.5625 1095.7500
#> [1114] 1795.8125  943.5625  852.2500  456.5625  913.1250  760.9375 1217.5000
#> [1121]  760.9375  334.8125  395.6875  395.6875 1430.5625 1217.5000  760.9375
#> [1128]  547.8750  547.8750  243.5000 1704.5000 1187.0625  334.8125  608.7500
#> [1135] 1521.8750 1126.1875 1339.2500  517.4375 1034.8750  821.8125  639.1875
#> [1142]  365.2500  304.3750  517.4375  395.6875 1491.4375 1765.3750 1613.1875
#> [1149] 1795.8125  669.6250  760.9375 1552.3125 1552.3125  669.6250  669.6250
#> [1156]  273.9375  974.0000 1430.5625 1582.7500 1765.3750  700.0625  426.1250
#> [1163]  943.5625 1156.6250  213.0625 1065.3125  608.7500 1339.2500 1187.0625
#> [1170] 1156.6250 1582.7500 1795.8125  608.7500 1095.7500 1156.6250  334.8125
#> [1177]  487.0000        NA  700.0625  700.0625 1765.3750 1430.5625 1643.6250
#> [1184] 1126.1875  243.5000 1156.6250 1095.7500  882.6875 1582.7500  426.1250
#> [1191]  426.1250  913.1250 1674.0625 1430.5625        NA  882.6875  913.1250
#> [1198] 1126.1875  821.8125  974.0000  547.8750  608.7500  791.3750  487.0000
#> [1205]  456.5625 1247.9375  456.5625 1734.9375  578.3125  395.6875  913.1250
#> [1212]  852.2500 1765.3750  882.6875 1004.4375 1308.8125  669.6250 1613.1875
#> [1219] 1582.7500  273.9375 1187.0625 1247.9375  304.3750 1400.1250 1613.1875
#> [1226] 1187.0625  913.1250  365.2500  791.3750 1187.0625  334.8125 1369.6875
#> [1233]  273.9375  456.5625  882.6875  273.9375  304.3750  578.3125 1582.7500
#> [1240]  243.5000 1217.5000  426.1250  273.9375  700.0625  608.7500  639.1875
#> [1247]  365.2500  852.2500  791.3750 1187.0625  304.3750 1491.4375  852.2500
#> [1254] 1217.5000 1674.0625 1034.8750 1156.6250 1461.0000  273.9375 1582.7500
#> [1261]  365.2500  791.3750  334.8125 1187.0625 1369.6875 1156.6250 1034.8750
#> [1268]  852.2500  395.6875 1491.4375  730.5000  791.3750  821.8125  426.1250
#> [1275]  760.9375  304.3750  791.3750  426.1250 1156.6250  882.6875  913.1250
#> [1282] 1369.6875  365.2500  608.7500 1613.1875 1339.2500 1095.7500  547.8750
#> [1289]  760.9375 1156.6250  669.6250  882.6875 1491.4375  487.0000 1795.8125
#> [1296] 1308.8125 1582.7500  456.5625  365.2500  821.8125  273.9375 1582.7500
#> [1303]  852.2500        NA  821.8125  974.0000  487.0000  700.0625  791.3750
#> [1310] 1308.8125 1004.4375  456.5625 1095.7500  639.1875  608.7500  730.5000
#> [1317] 1582.7500 1491.4375  426.1250  487.0000 1582.7500  791.3750  882.6875
#> [1324]  243.5000 1034.8750 1704.5000 1704.5000  852.2500  426.1250 1156.6250
#> [1331] 1613.1875 1126.1875 1369.6875  852.2500  243.5000 1247.9375  517.4375
#> [1338] 1278.3750  974.0000 1339.2500 1643.6250  639.1875  578.3125 1065.3125
#> [1345]  213.0625 1765.3750 1004.4375 1339.2500  304.3750  578.3125  730.5000
#> [1352]  791.3750 1734.9375  639.1875  517.4375 1521.8750  304.3750 1065.3125
#> [1359]  730.5000 1034.8750  852.2500 1400.1250  852.2500  426.1250  547.8750
#> [1366] 1613.1875  882.6875  760.9375  182.6250 1613.1875  243.5000 1247.9375
#> [1373]  182.6250  700.0625 1156.6250 1613.1875  517.4375 1491.4375  669.6250
#> [1380]  243.5000  273.9375 1491.4375  365.2500 1339.2500  760.9375 1095.7500
#> [1387] 1095.7500 1521.8750  760.9375  456.5625  730.5000 1613.1875 1004.4375
#> [1394]  913.1250  182.6250 1004.4375  852.2500  730.5000  821.8125  182.6250
#> [1401] 1552.3125 1156.6250  943.5625 1004.4375 1613.1875  882.6875  456.5625
#> [1408] 1126.1875 1400.1250 1217.5000 1765.3750 1126.1875  791.3750  426.1250
#> [1415]  365.2500 1795.8125 1552.3125 1582.7500  943.5625 1217.5000 1187.0625
#> [1422] 1765.3750  456.5625 1582.7500 1156.6250 1582.7500 1095.7500 1582.7500
#> [1429]  974.0000  639.1875 1095.7500 1187.0625 1461.0000  395.6875 1126.1875
#> [1436]  913.1250 1095.7500 1613.1875  821.8125 1095.7500  791.3750 1308.8125
#> [1443] 1552.3125 1643.6250  182.6250  304.3750  608.7500  669.6250  395.6875
#> [1450]  913.1250  456.5625 1400.1250 1430.5625 1461.0000 1491.4375 1278.3750
#> [1457]  730.5000  913.1250  882.6875  882.6875 1095.7500 1765.3750  974.0000
#> [1464] 1521.8750 1095.7500 1521.8750  700.0625  547.8750 1795.8125 1643.6250
#> [1471]  304.3750  669.6250 1156.6250  821.8125 1095.7500 1461.0000 1491.4375
#> [1478] 1034.8750  487.0000  456.5625 1643.6250 1734.9375  791.3750 1095.7500
#> [1485]  487.0000  334.8125  304.3750  578.3125  456.5625  487.0000 1004.4375
#> [1492]  913.1250  578.3125 1095.7500 1126.1875 1156.6250  273.9375  426.1250
#> [1499]  882.6875 1400.1250 1156.6250  821.8125 1643.6250 1126.1875  456.5625
#> [1506]  913.1250  821.8125 1339.2500 1034.8750 1521.8750  273.9375 1369.6875
#> [1513]  760.9375  943.5625  852.2500 1552.3125 1491.4375  182.6250  213.0625
#> [1520] 1521.8750 1339.2500 1217.5000  426.1250 1187.0625  913.1250  304.3750
#> [1527] 1643.6250  913.1250 1004.4375  852.2500  273.9375 1308.8125 1674.0625
#> [1534]  943.5625  791.3750 1430.5625 1217.5000  791.3750  913.1250 1643.6250
#> [1541] 1339.2500 1491.4375  456.5625  700.0625 1491.4375  517.4375 1461.0000
#> [1548] 1400.1250 1004.4375 1095.7500 1491.4375 1674.0625 1126.1875  304.3750
#> [1555] 1339.2500 1187.0625  821.8125  760.9375  456.5625 1004.4375  943.5625
#> [1562]  547.8750  334.8125  243.5000  730.5000 1461.0000  182.6250 1156.6250
#> [1569]  304.3750  943.5625 1034.8750 1339.2500  882.6875  213.0625  487.0000
#> [1576] 1156.6250  760.9375  791.3750  913.1250 1461.0000  791.3750  456.5625
#> [1583] 1430.5625  852.2500 1278.3750  487.0000 1461.0000 1095.7500  791.3750
#> [1590] 1521.8750  821.8125  547.8750 1400.1250  243.5000  273.9375 1095.7500
#> [1597] 1704.5000  426.1250 1461.0000  182.6250  365.2500  273.9375  456.5625
#> [1604] 1156.6250  913.1250  487.0000  487.0000  913.1250 1430.5625  487.0000
#> [1611] 1278.3750 1247.9375 1582.7500 1613.1875 1217.5000 1308.8125 1065.3125
#> [1618] 1582.7500 1461.0000  791.3750  456.5625  943.5625 1095.7500 1613.1875
#> [1625]  913.1250 1034.8750  791.3750 1734.9375  730.5000 1521.8750 1552.3125
#> [1632]  791.3750 1582.7500 1491.4375  913.1250 1582.7500 1034.8750  882.6875
#> [1639]  243.5000  821.8125 1065.3125  517.4375 1400.1250  852.2500  700.0625
#> [1646]  182.6250  395.6875  913.1250 1521.8750  700.0625 1674.0625 1247.9375
#> [1653] 1734.9375 1461.0000 1156.6250 1643.6250 1400.1250 1521.8750  334.8125
#> [1660]  852.2500 1217.5000 1187.0625 1582.7500  182.6250  852.2500  821.8125
#> [1667]  243.5000 1247.9375  365.2500 1674.0625 1095.7500  273.9375 1491.4375
#> [1674] 1674.0625 1308.8125 1674.0625  334.8125  456.5625 1369.6875 1217.5000
#> [1681]  639.1875 1095.7500 1004.4375 1613.1875 1187.0625  395.6875 1369.6875
#> [1688]  243.5000  365.2500  760.9375 1491.4375 1461.0000 1521.8750  700.0625
#> [1695] 1247.9375 1430.5625 1552.3125  791.3750  517.4375 1491.4375  700.0625
#> [1702]  304.3750 1582.7500 1217.5000 1643.6250 1400.1250 1095.7500  821.8125
#> [1709] 1004.4375  456.5625 1034.8750 1217.5000 1187.0625  760.9375 1187.0625
#> [1716] 1065.3125  304.3750 1795.8125  974.0000 1187.0625 1552.3125  791.3750
#> [1723] 1552.3125 1461.0000  974.0000 1278.3750 1521.8750 1217.5000  943.5625
#> [1730]  943.5625  547.8750  821.8125 1369.6875  852.2500 1400.1250  426.1250
#> [1737]  821.8125  273.9375 1217.5000  487.0000  821.8125  517.4375  608.7500
#> [1744] 1674.0625  760.9375  395.6875 1095.7500 1217.5000 1552.3125 1613.1875
#> [1751]  487.0000 1126.1875  547.8750 1400.1250 1156.6250  700.0625  821.8125
#> [1758] 1095.7500  487.0000  395.6875  821.8125  487.0000 1126.1875 1247.9375
#> [1765] 1552.3125 1156.6250  243.5000 1187.0625 1126.1875  426.1250 1095.7500
#> [1772]  791.3750 1278.3750  974.0000  608.7500 1126.1875  791.3750  517.4375
#> [1779]  426.1250 1247.9375  487.0000  487.0000 1034.8750  304.3750 1521.8750
#> [1786] 1400.1250  578.3125 1552.3125  882.6875  426.1250 1187.0625  365.2500
#> [1793] 1552.3125  608.7500  669.6250  974.0000 1095.7500 1582.7500  426.1250
#> [1800]  243.5000  273.9375 1734.9375 1521.8750 1126.1875 1552.3125 1643.6250
#> [1807]  213.0625  791.3750 1704.5000  608.7500 1339.2500 1126.1875 1795.8125
#> [1814]  182.6250  213.0625  669.6250  821.8125  578.3125 1369.6875  608.7500
#> [1821]  791.3750  304.3750 1369.6875 1308.8125  487.0000  791.3750  669.6250
#> [1828]  547.8750 1339.2500  334.8125 1095.7500  365.2500 1156.6250  791.3750
#> [1835]  791.3750 1582.7500 1187.0625  487.0000  791.3750 1704.5000 1095.7500
#> [1842]  821.8125 1552.3125  821.8125  669.6250  700.0625 1095.7500  760.9375
#> [1849] 1278.3750 1187.0625 1795.8125  730.5000  456.5625 1552.3125  760.9375
#> [1856] 1461.0000  730.5000 1369.6875  273.9375 1674.0625 1004.4375 1491.4375
#> [1863]  821.8125 1704.5000  487.0000 1126.1875 1095.7500 1095.7500 1400.1250
#> [1870] 1004.4375 1187.0625 1308.8125  547.8750  395.6875  426.1250 1521.8750
#> [1877]  974.0000  213.0625  395.6875  243.5000  578.3125 1521.8750 1552.3125
#> [1884]  791.3750 1187.0625 1369.6875  791.3750 1095.7500 1339.2500  395.6875
#> [1891]  943.5625        NA  487.0000 1004.4375  791.3750  426.1250 1247.9375
#> [1898] 1065.3125 1461.0000 1247.9375  669.6250  517.4375  273.9375 1308.8125
#> [1905]  182.6250  456.5625 1582.7500  821.8125  821.8125 1187.0625  821.8125
#> [1912]  365.2500  517.4375  608.7500 1400.1250  821.8125  730.5000 1308.8125
#> [1919]  730.5000  669.6250 1004.4375  426.1250 1187.0625  821.8125 1034.8750
#> [1926] 1765.3750  974.0000  791.3750 1339.2500 1430.5625 1126.1875 1430.5625
#> [1933]  365.2500  974.0000  913.1250 1004.4375  669.6250  578.3125 1065.3125
#> [1940] 1065.3125 1278.3750 1156.6250  213.0625  334.8125 1582.7500  700.0625
#> [1947] 1521.8750  578.3125 1369.6875 1126.1875 1004.4375 1004.4375  395.6875
#> [1954] 1095.7500  608.7500 1491.4375        NA 1187.0625  608.7500  700.0625
#> [1961]  487.0000 1400.1250 1765.3750  517.4375  517.4375  547.8750  578.3125
#> [1968]  700.0625 1461.0000 1095.7500  639.1875 1613.1875  821.8125  395.6875
#> [1975] 1095.7500 1308.8125 1156.6250  700.0625 1734.9375  791.3750 1369.6875
#> [1982]  669.6250 1187.0625  395.6875  304.3750 1217.5000 1004.4375 1521.8750
#> [1989]  639.1875 1339.2500  395.6875  760.9375 1491.4375 1582.7500  821.8125
#> [1996] 1400.1250 1704.5000 1095.7500 1217.5000 1034.8750 1126.1875  730.5000
#> [2003]  760.9375  395.6875 1613.1875  456.5625  608.7500  426.1250  669.6250
#> [2010] 1004.4375 1308.8125 1156.6250 1308.8125 1095.7500 1126.1875  213.0625
#> [2017] 1400.1250 1704.5000  669.6250  974.0000 1156.6250 1491.4375 1156.6250
#> [2024] 1461.0000 1734.9375 1126.1875  334.8125 1734.9375 1765.3750  974.0000
#> [2031] 1247.9375  334.8125 1095.7500  730.5000 1156.6250  487.0000  426.1250
#> [2038] 1552.3125  213.0625 1065.3125 1095.7500 1765.3750  365.2500  974.0000
#> [2045] 1552.3125 1065.3125  182.6250  426.1250  639.1875 1491.4375  791.3750
#> [2052] 1704.5000 1795.8125  334.8125 1674.0625 1339.2500 1095.7500 1004.4375
#> [2059] 1613.1875 1461.0000 1034.8750  456.5625  395.6875  974.0000  487.0000
#> [2066]  456.5625 1795.8125  974.0000 1461.0000 1521.8750  700.0625 1126.1875
#> [2073] 1034.8750  821.8125 1521.8750  456.5625 1095.7500 1339.2500  669.6250
#> [2080]  760.9375  547.8750  882.6875  243.5000 1004.4375  821.8125 1400.1250
#> [2087]  882.6875  821.8125  578.3125 1521.8750  791.3750 1308.8125 1126.1875
#> [2094] 1247.9375 1461.0000  456.5625 1004.4375  821.8125  700.0625 1065.3125
#> [2101] 1521.8750 1247.9375 1004.4375 1369.6875  882.6875 1674.0625  517.4375
#> [2108]  243.5000 1004.4375 1004.4375 1552.3125  182.6250 1187.0625  730.5000
#> [2115] 1034.8750 1734.9375 1065.3125 1461.0000 1674.0625  821.8125  760.9375
#> [2122]  700.0625 1156.6250 1461.0000 1400.1250  487.0000  730.5000 1552.3125
#> [2129] 1126.1875  700.0625  852.2500 1095.7500 1704.5000  791.3750 1217.5000
#> [2136]  243.5000  395.6875 1095.7500 1461.0000  243.5000  334.8125 1095.7500
#> [2143] 1095.7500 1095.7500  273.9375  182.6250  821.8125  821.8125  243.5000
#> [2150] 1126.1875  213.0625  760.9375 1461.0000  273.9375  760.9375  882.6875
#> [2157]  487.0000 1095.7500 1461.0000 1613.1875  913.1250 1704.5000  547.8750
#> [2164] 1491.4375  913.1250 1095.7500  517.4375  547.8750 1308.8125  791.3750
#> [2171] 1491.4375 1065.3125 1187.0625  182.6250 1095.7500 1521.8750 1034.8750
#> [2178] 1613.1875 1400.1250  791.3750  334.8125 1765.3750  487.0000 1582.7500
#> [2185] 1552.3125 1308.8125 1065.3125 1126.1875  700.0625 1613.1875 1095.7500
#> [2192]  182.6250  913.1250  213.0625  882.6875 1126.1875  487.0000  974.0000
#> [2199] 1461.0000  365.2500 1765.3750  608.7500  639.1875  456.5625  791.3750
#> [2206]  365.2500 1369.6875  821.8125  578.3125  517.4375  182.6250  639.1875
#> [2213]  821.8125 1095.7500 1491.4375  547.8750 1034.8750  730.5000 1156.6250
#> [2220]  608.7500 1491.4375 1065.3125 1065.3125  243.5000 1430.5625  243.5000
#> [2227] 1156.6250 1765.3750 1004.4375  213.0625  913.1250  517.4375  639.1875
#> [2234] 1339.2500 1369.6875 1369.6875  182.6250 1674.0625 1491.4375  456.5625
#> [2241]  791.3750 1430.5625 1552.3125  395.6875 1065.3125 1247.9375 1613.1875
#> [2248]  608.7500  547.8750 1339.2500  334.8125  487.0000 1126.1875 1552.3125
#> [2255] 1308.8125 1491.4375 1643.6250  365.2500 1400.1250 1613.1875  547.8750
#> [2262]  213.0625  791.3750  426.1250 1674.0625  487.0000 1004.4375 1552.3125
#> [2269]  882.6875  913.1250 1187.0625  669.6250 1217.5000  821.8125  821.8125
#> [2276] 1491.4375  395.6875 1156.6250 1491.4375 1217.5000  882.6875  852.2500
#> [2283]  821.8125  426.1250  365.2500  456.5625 1552.3125 1582.7500  487.0000
#> [2290]  852.2500  456.5625  852.2500  456.5625 1674.0625  517.4375 1461.0000
#> [2297] 1674.0625 1552.3125  760.9375 1674.0625  456.5625 1521.8750  882.6875
#> [2304] 1430.5625 1339.2500  365.2500 1795.8125  578.3125  243.5000  882.6875
#> [2311] 1065.3125 1187.0625 1430.5625  852.2500 1095.7500  182.6250  334.8125
#> [2318] 1613.1875  456.5625 1461.0000 1643.6250  182.6250 1369.6875  852.2500
#> [2325]  760.9375  943.5625        NA 1187.0625        NA  760.9375 1126.1875
#> [2332]  821.8125 1034.8750 1582.7500  852.2500  852.2500 1034.8750 1734.9375
#> [2339] 1156.6250  760.9375  669.6250  913.1250  426.1250  852.2500  395.6875
#> [2346] 1643.6250  578.3125  182.6250 1095.7500 1369.6875  487.0000  334.8125
#> [2353] 1156.6250  791.3750  547.8750  821.8125  669.6250 1247.9375 1461.0000
#> [2360]  304.3750  243.5000  608.7500  639.1875 1582.7500  273.9375 1491.4375
#> [2367] 1126.1875 1582.7500 1247.9375  821.8125  456.5625  517.4375 1430.5625
#> [2374]  791.3750 1065.3125  487.0000  365.2500  700.0625 1491.4375  730.5000
#> [2381] 1704.5000  669.6250 1369.6875  669.6250  760.9375  669.6250 1461.0000
#> [2388]  182.6250  639.1875  882.6875  821.8125 1461.0000 1491.4375 1187.0625
#> [2395] 1217.5000 1187.0625 1461.0000 1217.5000        NA  517.4375 1734.9375
#> [2402]  730.5000 1095.7500 1582.7500 1095.7500 1095.7500 1095.7500  365.2500
#> [2409] 1734.9375  852.2500 1187.0625 1004.4375 1461.0000  852.2500  213.0625
#> [2416] 1156.6250  639.1875 1156.6250 1065.3125  913.1250 1156.6250  243.5000
#> [2423]  334.8125 1187.0625  730.5000  700.0625 1461.0000  730.5000 1278.3750
#> [2430] 1552.3125 1095.7500  760.9375  517.4375  760.9375 1491.4375 1095.7500
#> [2437]  760.9375  578.3125 1339.2500  243.5000 1339.2500 1065.3125  974.0000
#> [2444] 1187.0625  334.8125 1674.0625 1339.2500  578.3125  395.6875 1582.7500
#> [2451]  365.2500 1491.4375  669.6250 1034.8750  730.5000  547.8750 1034.8750
#> [2458]  395.6875  487.0000 1126.1875 1187.0625 1765.3750 1582.7500  791.3750
#> [2465] 1491.4375 1582.7500  456.5625  547.8750  547.8750 1187.0625  304.3750
#> [2472]  730.5000 1400.1250  730.5000  791.3750  365.2500 1491.4375 1400.1250
#> [2479]  760.9375 1339.2500  608.7500  243.5000  700.0625 1461.0000 1795.8125
#> [2486]  273.9375 1369.6875 1461.0000 1004.4375  365.2500  669.6250 1095.7500
#> [2493] 1095.7500  760.9375  974.0000  365.2500  365.2500 1004.4375  700.0625
#> [2500]  547.8750  821.8125 1095.7500  273.9375  304.3750  487.0000  182.6250
#> [2507] 1461.0000 1065.3125 1613.1875  365.2500 1400.1250 1461.0000  456.5625
#> [2514]  974.0000  426.1250 1095.7500 1461.0000  213.0625  517.4375 1613.1875
#> [2521]  304.3750  760.9375  821.8125  608.7500  487.0000  700.0625  578.3125
#> [2528] 1065.3125  974.0000 1095.7500  334.8125  852.2500 1217.5000 1613.1875
#> [2535]  395.6875  365.2500 1004.4375  365.2500 1461.0000  334.8125  760.9375
#> [2542] 1187.0625  913.1250 1765.3750 1339.2500  243.5000  639.1875 1461.0000
#> [2549] 1004.4375 1339.2500 1095.7500  791.3750  700.0625  943.5625  974.0000
#> [2556] 1004.4375 1156.6250  852.2500 1187.0625  852.2500 1126.1875  334.8125
#> [2563]  365.2500 1217.5000  760.9375  334.8125 1461.0000 1643.6250 1613.1875
#> [2570] 1095.7500  669.6250 1339.2500 1521.8750  821.8125  182.6250  213.0625
#> [2577]  547.8750  882.6875  943.5625  517.4375  395.6875 1034.8750 1095.7500
#> [2584] 1674.0625 1339.2500  730.5000 1430.5625 1521.8750 1795.8125 1095.7500
#> [2591]  487.0000 1643.6250 1369.6875  213.0625 1187.0625  547.8750  700.0625
#> [2598]  669.6250  273.9375 1734.9375  882.6875  426.1250  578.3125 1339.2500
#> [2605]  334.8125  213.0625 1795.8125  852.2500  760.9375 1004.4375  852.2500
#> [2612]  791.3750 1491.4375 1004.4375 1278.3750  852.2500  730.5000 1156.6250
#> [2619]  426.1250 1491.4375 1582.7500 1582.7500  213.0625 1065.3125  669.6250
#> [2626] 1461.0000 1795.8125 1552.3125  365.2500 1582.7500 1095.7500 1034.8750
#> [2633] 1674.0625  456.5625 1765.3750 1613.1875  517.4375 1339.2500  487.0000
#> [2640]  821.8125  243.5000  608.7500 1247.9375  395.6875 1217.5000 1247.9375
#> [2647] 1674.0625 1521.8750 1643.6250  213.0625 1430.5625 1521.8750  669.6250
#> [2654] 1126.1875  578.3125 1095.7500  608.7500 1247.9375  669.6250 1187.0625
#> [2661]  700.0625 1704.5000 1369.6875  426.1250  760.9375  395.6875 1521.8750
#> [2668] 1065.3125  213.0625 1552.3125  760.9375  669.6250  547.8750 1704.5000
#> [2675]  273.9375  304.3750 1521.8750        NA 1034.8750  182.6250 1126.1875
#> [2682] 1126.1875 1613.1875 1552.3125 1004.4375 1156.6250 1552.3125 1430.5625
#> [2689] 1674.0625 1491.4375  182.6250 1369.6875 1217.5000  426.1250  791.3750
#> [2696] 1095.7500 1765.3750  456.5625  821.8125  852.2500  426.1250  456.5625
#> [2703] 1065.3125 1430.5625 1187.0625 1004.4375  700.0625 1430.5625  426.1250
#> [2710] 1126.1875 1004.4375 1765.3750 1156.6250 1278.3750  608.7500 1674.0625
#> [2717] 1339.2500  365.2500  273.9375  700.0625 1065.3125 1004.4375 1582.7500
#> [2724]  426.1250 1613.1875  700.0625 1674.0625 1643.6250 1461.0000 1430.5625
#> [2731] 1217.5000 1643.6250  821.8125 1369.6875 1034.8750 1217.5000 1734.9375
#> [2738]  547.8750 1430.5625  639.1875  487.0000  395.6875  243.5000 1582.7500
#> [2745]  852.2500 1217.5000 1734.9375 1156.6250  791.3750 1552.3125 1217.5000
#> [2752]  304.3750 1461.0000 1187.0625  365.2500 1582.7500  273.9375  700.0625
#> [2759] 1521.8750  700.0625 1491.4375  578.3125 1126.1875  669.6250  700.0625
#> [2766] 1156.6250  426.1250  639.1875  456.5625  426.1250 1126.1875  517.4375
#> [2773] 1004.4375 1278.3750 1613.1875 1156.6250  456.5625 1126.1875 1795.8125
#> [2780] 1034.8750 1521.8750 1765.3750 1400.1250 1461.0000 1065.3125 1521.8750
#> [2787]  882.6875 1095.7500  608.7500  608.7500  487.0000 1217.5000 1795.8125
#> [2794] 1765.3750 1674.0625 1369.6875  243.5000  669.6250 1247.9375 1461.0000
#> [2801] 1034.8750  700.0625 1461.0000  760.9375 1095.7500  730.5000 1765.3750
#> [2808] 1430.5625 1187.0625 1217.5000  821.8125 1156.6250  882.6875  578.3125
#> [2815]  213.0625  243.5000 1582.7500 1156.6250  487.0000 1156.6250 1674.0625
#> [2822]  821.8125 1795.8125 1582.7500 1308.8125  213.0625  669.6250  426.1250
#> [2829] 1400.1250 1704.5000  243.5000  943.5625  943.5625 1126.1875 1613.1875
#> [2836] 1795.8125  760.9375  882.6875  213.0625  487.0000 1034.8750 1552.3125
#> [2843] 1430.5625 1065.3125  760.9375  974.0000 1461.0000  426.1250 1521.8750
#> [2850] 1491.4375 1674.0625 1156.6250 1278.3750 1065.3125 1156.6250  273.9375
#> [2857]  943.5625  304.3750 1156.6250  974.0000  700.0625  304.3750  608.7500
#> [2864]  821.8125 1156.6250  760.9375  243.5000 1065.3125 1369.6875 1430.5625
#> [2871] 1521.8750 1765.3750  700.0625  182.6250 1674.0625  334.8125  395.6875
#> [2878] 1795.8125  913.1250 1095.7500  243.5000  730.5000  760.9375  730.5000
#> [2885] 1156.6250  243.5000  730.5000 1491.4375  547.8750  273.9375  700.0625
#> [2892] 1187.0625  243.5000  243.5000 1095.7500  334.8125 1095.7500  700.0625
#> [2899]  334.8125 1552.3125        NA 1765.3750 1400.1250  639.1875 1034.8750
#> [2906] 1521.8750  608.7500  730.5000 1065.3125 1521.8750  669.6250 1643.6250
#> [2913] 1734.9375  730.5000 1521.8750 1247.9375        NA 1095.7500  608.7500
#> [2920]  974.0000 1278.3750  852.2500 1521.8750  426.1250  487.0000 1369.6875
#> [2927] 1461.0000  730.5000  669.6250  791.3750 1734.9375 1521.8750  882.6875
#> [2934]  791.3750  821.8125  730.5000  730.5000 1217.5000 1247.9375  487.0000
#> [2941]  821.8125  913.1250  608.7500  821.8125  913.1250  243.5000 1247.9375
#> [2948]  700.0625  456.5625  395.6875 1643.6250  426.1250  639.1875 1582.7500
#> [2955]  639.1875  882.6875 1156.6250 1156.6250 1491.4375 1734.9375 1552.3125
#> [2962] 1308.8125 1065.3125  304.3750 1521.8750  791.3750  700.0625  304.3750
#> [2969]  273.9375 1156.6250 1461.0000  334.8125  700.0625  760.9375 1521.8750
#> [2976] 1613.1875 1613.1875  395.6875 1126.1875 1582.7500 1613.1875 1187.0625
#> [2983] 1400.1250  578.3125 1004.4375 1369.6875 1278.3750 1430.5625 1278.3750
#> [2990]  487.0000 1187.0625  608.7500  821.8125 1156.6250 1217.5000 1156.6250
#> [2997] 1521.8750  243.5000 1582.7500  -91.3125 1156.6250  487.0000  639.1875
#> [3004]  974.0000 1582.7500  669.6250 1156.6250  273.9375 1582.7500  243.5000
#> [3011]  456.5625 1430.5625  395.6875  273.9375 1156.6250 1126.1875  365.2500
#> [3018] 1582.7500 1095.7500 1400.1250  760.9375  700.0625  365.2500 1308.8125
#> [3025] 1430.5625 1156.6250 1400.1250  243.5000 1613.1875  304.3750 1187.0625
#> [3032] 1400.1250 1034.8750  395.6875  426.1250 1369.6875 1217.5000  487.0000
#> [3039] 1765.3750 1187.0625        NA 1339.2500  182.6250  578.3125 1734.9375
#> [3046]  669.6250 1278.3750        NA 1461.0000 1126.1875  273.9375        NA
#> [3053]  608.7500  669.6250  426.1250 1187.0625  821.8125        NA 1613.1875
#> [3060]        NA 1430.5625 1582.7500 1521.8750  243.5000  395.6875  213.0625
#> [3067]  943.5625 1004.4375        NA  852.2500  639.1875 1126.1875        NA
#> [3074]  974.0000  365.2500  487.0000 1095.7500 1400.1250  456.5625 1369.6875
#> [3081]  669.6250 1217.5000 1065.3125        NA        NA 1217.5000  365.2500
#> [3088]  517.4375        NA 1095.7500  213.0625  334.8125  943.5625 1521.8750
#> [3095]  821.8125 1247.9375 1430.5625  304.3750 1582.7500  304.3750  974.0000
#> [3102]  487.0000  487.0000  608.7500 1339.2500 1582.7500 1430.5625 1430.5625
#> [3109]  426.1250 1400.1250  882.6875  821.8125 1552.3125 1491.4375 1278.3750
#> [3116]  821.8125  913.1250 1734.9375 1187.0625 1400.1250 1734.9375 1034.8750
#> [3123]  182.6250 1430.5625 1217.5000 1369.6875 1217.5000  243.5000  791.3750
#> [3130]  395.6875 1674.0625  760.9375  882.6875  182.6250 1308.8125  821.8125
#> [3137] 1400.1250 1095.7500 1521.8750  304.3750 1704.5000 1430.5625  487.0000
#> [3144] 1491.4375  913.1250 1095.7500  456.5625 1187.0625 1552.3125  821.8125
#> [3151]  395.6875 1095.7500  852.2500  456.5625  760.9375  882.6875 1095.7500
#> [3158] 1095.7500  426.1250 1461.0000 1126.1875  487.0000 1095.7500 1126.1875
#> [3165]  760.9375 1521.8750 1156.6250 1461.0000  974.0000 1613.1875  517.4375
#> [3172]  791.3750  974.0000 1217.5000  517.4375 1552.3125  730.5000  304.3750
#> [3179] 1217.5000 1582.7500  395.6875  913.1250 1126.1875 1065.3125  517.4375
#> [3186]  547.8750  456.5625  943.5625 1156.6250  730.5000 1552.3125  426.1250
#> [3193]  213.0625 1552.3125 1461.0000  882.6875 1095.7500 1187.0625 1521.8750
#> [3200]  456.5625 1217.5000 1461.0000 1582.7500 1582.7500  730.5000  365.2500
#> [3207] 1065.3125 1491.4375  882.6875 1004.4375 1187.0625  182.6250 1369.6875
#> [3214]  456.5625 1247.9375 1400.1250 1552.3125  456.5625 1156.6250 1613.1875
#> [3221] 1217.5000  852.2500  487.0000 1217.5000 1187.0625  213.0625  243.5000
#> [3228]  456.5625 1582.7500  730.5000 1400.1250  943.5625 1430.5625 1339.2500
#> [3235]  243.5000  730.5000  669.6250  243.5000 1187.0625 1004.4375 1187.0625
#> [3242]  700.0625 1308.8125 1247.9375 1552.3125  700.0625 1217.5000 1278.3750
#> [3249] 1095.7500  456.5625 1065.3125 1491.4375 1126.1875  882.6875 1126.1875
#> [3256] 1582.7500  517.4375 1674.0625 1187.0625 1308.8125 1004.4375  791.3750
#> [3263] 1187.0625  974.0000  243.5000  608.7500  273.9375  821.8125  760.9375
#> [3270] 1156.6250  334.8125 1126.1875  395.6875  639.1875  760.9375 1004.4375
#> [3277]  913.1250  273.9375 1674.0625 1369.6875  821.8125  700.0625 1156.6250
#> [3284] 1065.3125 1552.3125  974.0000  760.9375 1247.9375  760.9375  517.4375
#> [3291]  821.8125 1795.8125 1126.1875 1582.7500 1095.7500 1674.0625 1461.0000
#> [3298] 1187.0625 1095.7500  517.4375  760.9375 1795.8125 1126.1875  791.3750
#> [3305] 1278.3750 1065.3125  608.7500  273.9375 1217.5000  273.9375  304.3750
#> [3312] 1034.8750 1217.5000  760.9375 1461.0000 1491.4375  760.9375 1400.1250
#> [3319]  517.4375 1217.5000 1004.4375 1582.7500  669.6250 1552.3125 1795.8125
#> [3326]  791.3750 1552.3125  395.6875 1004.4375  760.9375 1613.1875 1156.6250
#> [3333]  213.0625 1491.4375  821.8125 1095.7500 1187.0625 1461.0000 1461.0000
#> [3340] 1461.0000  608.7500  547.8750  669.6250  730.5000  608.7500 1552.3125
#> [3347]  760.9375  913.1250  791.3750 1217.5000 1582.7500  456.5625  334.8125
#> [3354]  517.4375 1247.9375 1369.6875 1643.6250 1217.5000  669.6250 1674.0625
#> [3361]  608.7500 1217.5000 1187.0625  608.7500  517.4375 1521.8750  882.6875
#> [3368] 1521.8750  426.1250  882.6875 1004.4375  334.8125 1156.6250 1034.8750
#> [3375]  213.0625 1247.9375 1247.9375 1369.6875 1400.1250  334.8125  760.9375
#> [3382]  426.1250 1156.6250 1643.6250  913.1250  791.3750  365.2500 1034.8750
#> [3389]  821.8125 1369.6875  243.5000 1308.8125 1004.4375 1004.4375 1217.5000
#> [3396] 1552.3125 1034.8750  700.0625 1278.3750  821.8125  974.0000  456.5625
#> [3403]  852.2500  578.3125  273.9375 1674.0625  943.5625  365.2500 1552.3125
#> [3410]  426.1250  791.3750  760.9375  395.6875  791.3750 1034.8750 1065.3125
#> [3417] 1795.8125 1339.2500  791.3750 1369.6875  334.8125  517.4375 1461.0000
#> [3424] 1095.7500        NA  334.8125  760.9375 1461.0000  426.1250  182.6250
#> [3431] 1247.9375 2039.3125 1187.0625  487.0000  395.6875 1156.6250  395.6875
#> [3438] 1521.8750 1765.3750  182.6250  882.6875  273.9375  487.0000  974.0000
#> [3445]  791.3750  273.9375 1613.1875  304.3750 1734.9375  700.0625  273.9375
#> [3452] 1126.1875 1674.0625  913.1250 1156.6250 1643.6250  395.6875 1095.7500
#> [3459] 1095.7500  639.1875  487.0000 1217.5000 1034.8750 1095.7500 1187.0625
#> [3466]  182.6250  517.4375  426.1250  243.5000  334.8125  730.5000 1095.7500
#> [3473] 1461.0000  730.5000  730.5000 1491.4375  730.5000  334.8125  547.8750
#> [3480]  760.9375 1247.9375  426.1250  821.8125  213.0625 1095.7500  852.2500
#> [3487] 1461.0000  791.3750  882.6875 1095.7500 1156.6250 1004.4375  730.5000
#> [3494] 1461.0000  456.5625  608.7500  669.6250  273.9375 1491.4375 1034.8750
#> [3501] 1217.5000 1247.9375  608.7500  852.2500  213.0625  578.3125 1065.3125
#> [3508] 1339.2500 1278.3750  608.7500  700.0625  943.5625  760.9375 1704.5000
#> [3515]  760.9375 1491.4375  517.4375  700.0625 1643.6250 1187.0625 1126.1875
#> [3522]  608.7500 1369.6875  426.1250 1369.6875  273.9375  821.8125 1156.6250
#> [3529]  791.3750 1613.1875 1461.0000 1765.3750 1247.9375  882.6875  639.1875
#> [3536] 1217.5000  547.8750  487.0000 1521.8750  943.5625 1187.0625  182.6250
#> [3543] 1065.3125  821.8125  730.5000 1034.8750  456.5625 1521.8750  487.0000
#> [3550] 1308.8125  700.0625  334.8125 1126.1875 1308.8125 1552.3125 1217.5000
#> [3557]  639.1875  304.3750 1613.1875  760.9375  974.0000 1095.7500  821.8125
#> [3564] 1034.8750  608.7500  365.2500 1461.0000  334.8125 1704.5000  395.6875
#> [3571]  517.4375  852.2500  913.1250  213.0625 1461.0000 1004.4375 1461.0000
#> [3578]  426.1250 1369.6875  608.7500  578.3125 1430.5625  304.3750  730.5000
#> [3585]  395.6875  456.5625