Function to recode age variable and convert to age in days

Function to recode age variable and convert to age in days

recode_age(df)

Arguments

df
A coverage data.frame collected for the Liberia Coverage Survey

Value

A vector of recoded ages based on any re-testing done

Examples

recode_age(df = coverageData.r2)

#>    [1]  700.0625  365.2500  608.7500  426.1250  700.0625 1461.0000  213.0625
#>    [8] 1552.3125 1582.7500 1156.6250 1430.5625  456.5625  760.9375 1095.7500
#>   [15] 1521.8750 1095.7500  304.3750  304.3750  182.6250  426.1250  426.1250
#>   [22]  974.0000  213.0625 1461.0000 1095.7500 1095.7500  760.9375  547.8750
#>   [29]  334.8125  182.6250 1126.1875 1461.0000 1247.9375 1247.9375  487.0000
#>   [36]  182.6250 1004.4375  669.6250  395.6875  547.8750  669.6250 1613.1875
#>   [43] 1461.0000  943.5625  456.5625  608.7500  882.6875  882.6875 1187.0625
#>   [50] 1491.4375  456.5625  669.6250  182.6250  487.0000 1004.4375 1734.9375
#>   [57]  304.3750 1765.3750  182.6250 1461.0000 1187.0625 1521.8750  426.1250
#>   [64]  517.4375  304.3750 1065.3125 1217.5000 1278.3750  882.6875  760.9375
#>   [71] 1095.7500  791.3750 1034.8750 1369.6875 1034.8750  334.8125 1461.0000
#>   [78] 1491.4375  456.5625 1734.9375  639.1875  821.8125 1126.1875 1308.8125
#>   [85]  821.8125  334.8125  365.2500  213.0625 1247.9375 1034.8750 1430.5625
#>   [92]  547.8750 1278.3750  821.8125  273.9375  547.8750  608.7500  791.3750
#>   [99]  456.5625 1278.3750 1734.9375  487.0000  821.8125  608.7500 1034.8750
#>  [106]  730.5000  608.7500 1308.8125  852.2500  791.3750  395.6875 1187.0625
#>  [113]  426.1250 1491.4375  700.0625 1765.3750  669.6250  456.5625 1034.8750
#>  [120]  456.5625 1004.4375        NA  578.3125 1126.1875 1704.5000  578.3125
#>  [127]  700.0625  730.5000  821.8125  578.3125  517.4375 1095.7500  426.1250
#>  [134] 1004.4375 1247.9375  395.6875  974.0000  426.1250 1491.4375  700.0625
#>  [141] 1369.6875 1156.6250 1461.0000 1765.3750 1095.7500 1217.5000 1126.1875
#>  [148]  852.2500 1187.0625  273.9375  517.4375 1643.6250  456.5625  913.1250
#>  [155]  334.8125 1765.3750  730.5000 1643.6250  426.1250 1095.7500  730.5000
#>  [162]  304.3750  182.6250  669.6250  456.5625 1217.5000  213.0625 1247.9375
#>  [169] 1369.6875 1461.0000 1521.8750  730.5000  182.6250 1552.3125  639.1875
#>  [176]  760.9375 1613.1875  487.0000  821.8125  456.5625  913.1250  669.6250
#>  [183] 1552.3125 1156.6250  821.8125  547.8750  882.6875  395.6875 1065.3125
#>  [190] 1308.8125  334.8125 1247.9375  243.5000 1126.1875 1095.7500 1674.0625
#>  [197] 1461.0000  334.8125 1430.5625 1430.5625  943.5625  213.0625 1247.9375
#>  [204]  273.9375 1491.4375  456.5625  395.6875 1156.6250 1156.6250  608.7500
#>  [211] 1552.3125  821.8125  669.6250  547.8750 1339.2500 1004.4375        NA
#>  [218]  243.5000  791.3750 1065.3125  426.1250  821.8125 1156.6250 1461.0000
#>  [225]  487.0000  243.5000 1187.0625  182.6250 1613.1875 1095.7500  182.6250
#>  [232] 1126.1875 1308.8125  669.6250 1034.8750  669.6250 1126.1875  213.0625
#>  [239] 1461.0000  304.3750  821.8125 1795.8125  365.2500 1034.8750  426.1250
#>  [246] 1400.1250  456.5625 1552.3125  547.8750 1765.3750  243.5000  456.5625
#>  [253]  426.1250  760.9375  213.0625 1369.6875 1521.8750  791.3750  639.1875
#>  [260]  517.4375 1095.7500 1704.5000 1095.7500 1369.6875 1552.3125  730.5000
#>  [267] 1491.4375  487.0000 1156.6250 1187.0625 1795.8125  243.5000 1552.3125
#>  [274] 1156.6250 1400.1250  243.5000 1339.2500  273.9375  791.3750  243.5000
#>  [281] 1339.2500  395.6875 1461.0000  730.5000  852.2500  273.9375  273.9375
#>  [288] 1095.7500 1247.9375  913.1250  852.2500 1582.7500  456.5625  334.8125
#>  [295] 1734.9375 1552.3125  243.5000 1521.8750  608.7500 1369.6875  547.8750
#>  [302]  669.6250 1247.9375 1643.6250  334.8125  517.4375  456.5625  334.8125
#>  [309]  882.6875  974.0000  974.0000 1582.7500 1369.6875 1156.6250  608.7500
#>  [316] 1217.5000  487.0000 1491.4375 1613.1875  304.3750 1034.8750 1308.8125
#>  [323]  334.8125  487.0000  273.9375  639.1875  426.1250  882.6875 1004.4375
#>  [330]  395.6875 1217.5000  243.5000  243.5000 1095.7500  669.6250 1491.4375
#>  [337]  882.6875  821.8125 1582.7500 1552.3125  760.9375 1613.1875 1187.0625
#>  [344] 1613.1875 1795.8125  791.3750 1369.6875 1521.8750  669.6250  791.3750
#>  [351] 1126.1875 1552.3125 1339.2500 1278.3750 1552.3125 1156.6250  608.7500
#>  [358] 1734.9375  943.5625 1582.7500        NA  821.8125 1582.7500  730.5000
#>  [365]  273.9375 1126.1875  395.6875  456.5625 1247.9375  821.8125 1217.5000
#>  [372]  365.2500 1339.2500 1034.8750 1674.0625 1217.5000  669.6250 1095.7500
#>  [379] 1461.0000 1674.0625  395.6875  456.5625 1521.8750  243.5000 1339.2500
#>  [386] 1095.7500  700.0625 1187.0625 1613.1875  852.2500  456.5625  426.1250
#>  [393] 1521.8750  821.8125 1430.5625 1095.7500  578.3125 1065.3125  821.8125
#>  [400] 1521.8750  852.2500 1095.7500  852.2500  487.0000  791.3750  852.2500
#>  [407]  852.2500 1552.3125  426.1250 1369.6875 1217.5000  547.8750 1430.5625
#>  [414]  395.6875  913.1250 1065.3125 1095.7500  669.6250 1795.8125  882.6875
#>  [421] 1156.6250 1278.3750  791.3750  365.2500 1704.5000  395.6875 1126.1875
#>  [428]  243.5000 1004.4375  639.1875 1217.5000 1461.0000 1643.6250  913.1250
#>  [435]  608.7500  821.8125 1217.5000 1704.5000 1187.0625 1704.5000  304.3750
#>  [442]  243.5000  730.5000  700.0625 1795.8125 1065.3125 1308.8125  608.7500
#>  [449]  395.6875  608.7500 1339.2500  913.1250 1126.1875 1643.6250 1734.9375
#>  [456] 1004.4375  517.4375 1126.1875  791.3750  517.4375  974.0000  426.1250
#>  [463]  760.9375  913.1250 1582.7500 1552.3125 1095.7500 1095.7500 1582.7500
#>  [470]  730.5000  639.1875 1247.9375  639.1875 1582.7500 1734.9375  821.8125
#>  [477] 1126.1875  760.9375  456.5625  182.6250  395.6875 1187.0625  304.3750
#>  [484] 1156.6250 1430.5625 1643.6250  821.8125  730.5000 1034.8750  882.6875
#>  [491] 1461.0000  517.4375  182.6250 1430.5625  213.0625  456.5625  273.9375
#>  [498] 1339.2500  821.8125  974.0000 1430.5625  517.4375 1187.0625  913.1250
#>  [505] 1217.5000  608.7500 1582.7500 1369.6875 1095.7500 1034.8750 1704.5000
#>  [512]  304.3750 1491.4375 1552.3125 1278.3750  182.6250  913.1250 1187.0625
#>  [519]  821.8125  243.5000        NA 1521.8750  913.1250 1095.7500  365.2500
#>  [526] 1674.0625 1095.7500  730.5000 1126.1875  213.0625  913.1250  882.6875
#>  [533] 1308.8125 1156.6250  213.0625 1734.9375  821.8125 1004.4375        NA
#>  [540] 1430.5625  365.2500  334.8125 1339.2500 1369.6875  395.6875 1217.5000
#>  [547] 1034.8750  304.3750 1461.0000 1095.7500  760.9375  456.5625 1339.2500
#>  [554] 1095.7500 1187.0625  243.5000 1278.3750 1187.0625  700.0625  669.6250
#>  [561]  669.6250 1187.0625 1461.0000 1491.4375 1156.6250 1126.1875  760.9375
#>  [568]  639.1875  760.9375  365.2500  608.7500  852.2500 1339.2500 1095.7500
#>  [575] 1247.9375  365.2500 1156.6250  639.1875 1156.6250  426.1250  182.6250
#>  [582]  913.1250 1217.5000  456.5625 1095.7500  669.6250 1552.3125  517.4375
#>  [589] 1369.6875  760.9375 1795.8125 1126.1875 1765.3750 1521.8750 1004.4375
#>  [596] 1095.7500 1034.8750 1582.7500  608.7500  943.5625  669.6250 1034.8750
#>  [603]  821.8125 1217.5000  578.3125  882.6875 1156.6250 1582.7500  974.0000
#>  [610] 1461.0000  669.6250 1308.8125 1369.6875  730.5000  213.0625  730.5000
#>  [617] 1034.8750  882.6875  517.4375 1065.3125 1187.0625  517.4375 1004.4375
#>  [624] 1521.8750 1095.7500  760.9375  821.8125 1674.0625 1643.6250  913.1250
#>  [631]  456.5625 1521.8750  487.0000  882.6875 1278.3750 1674.0625  760.9375
#>  [638] 1034.8750  487.0000  213.0625 1065.3125  273.9375  760.9375  304.3750
#>  [645] 1187.0625  730.5000 1156.6250 1247.9375 1247.9375  273.9375 1247.9375
#>  [652]  304.3750 1461.0000  821.8125 1765.3750  821.8125 1004.4375 1461.0000
#>  [659]        NA  700.0625 1095.7500  760.9375 1369.6875 1187.0625 1126.1875
#>  [666] 1430.5625  730.5000  700.0625  821.8125  730.5000 1552.3125  700.0625
#>  [673]  730.5000 1704.5000  821.8125 1461.0000  426.1250 1582.7500 1217.5000
#>  [680]  365.2500  578.3125  395.6875 1217.5000 1400.1250 1369.6875  791.3750
#>  [687] 1369.6875 1156.6250 1339.2500  487.0000  700.0625  182.6250  243.5000
#>  [694] 1004.4375  487.0000 1613.1875 1156.6250  882.6875  669.6250  913.1250
#>  [701] 1247.9375 1461.0000  182.6250 1643.6250 1643.6250 1521.8750 1369.6875
#>  [708] 1795.8125  791.3750  487.0000 1308.8125 1217.5000 1217.5000 1369.6875
#>  [715] 1430.5625 1126.1875 1339.2500  700.0625 1156.6250  882.6875  365.2500
#>  [722] 1400.1250 1461.0000 1095.7500  334.8125 1461.0000 1187.0625 1278.3750
#>  [729] 1126.1875  760.9375 1156.6250 1704.5000  547.8750 1126.1875  365.2500
#>  [736] 1004.4375  913.1250 1613.1875  517.4375  913.1250  821.8125  487.0000
#>  [743] 1582.7500  395.6875 1126.1875  943.5625  456.5625 1430.5625  334.8125
#>  [750]  182.6250  517.4375  639.1875 1278.3750 1461.0000 1004.4375 1400.1250
#>  [757] 1339.2500 1308.8125 1187.0625 1430.5625  700.0625 1126.1875  608.7500
#>  [764]  365.2500  974.0000 1247.9375  487.0000  182.6250  547.8750  426.1250
#>  [771] 1582.7500  426.1250 1278.3750  517.4375  882.6875 1491.4375 1613.1875
#>  [778] 1643.6250  547.8750 1674.0625  273.9375 1552.3125  760.9375 1582.7500
#>  [785] 1369.6875  578.3125 1613.1875 1156.6250 1430.5625 1400.1250 1552.3125
#>  [792]  365.2500  182.6250  304.3750 1095.7500  639.1875 1156.6250  304.3750
#>  [799] 1126.1875 1369.6875  700.0625  456.5625  943.5625  365.2500  852.2500
#>  [806] 1674.0625 1643.6250  182.6250  334.8125 1491.4375  517.4375 1461.0000
#>  [813] 1187.0625 1187.0625 1217.5000  791.3750 1461.0000 1400.1250 1156.6250
#>  [820] 1065.3125  547.8750 1734.9375  334.8125  608.7500  913.1250 1126.1875
#>  [827] 1095.7500  730.5000  608.7500  882.6875 1795.8125 1126.1875  304.3750
#>  [834] 1065.3125 1582.7500  730.5000 1765.3750  669.6250  487.0000  791.3750
#>  [841]  426.1250 1369.6875 1461.0000  182.6250  913.1250  821.8125  700.0625
#>  [848]  243.5000  669.6250  517.4375  913.1250 1491.4375  426.1250 1674.0625
#>  [855] 1582.7500 1156.6250 1339.2500  365.2500  608.7500 1613.1875  791.3750
#>  [862] 1278.3750 1582.7500 1217.5000  547.8750 1339.2500 1491.4375 1734.9375
#>  [869]  426.1250  426.1250 1582.7500  730.5000  669.6250  852.2500  760.9375
#>  [876]  852.2500  426.1250 1065.3125  304.3750  791.3750 1521.8750 1552.3125
#>  [883] 1095.7500  821.8125  426.1250  334.8125  334.8125 1095.7500  821.8125
#>  [890]  943.5625 1674.0625 1156.6250  760.9375  213.0625  517.4375  760.9375
#>  [897]  395.6875  760.9375  365.2500  517.4375 1521.8750  608.7500 1795.8125
#>  [904]  760.9375  334.8125  760.9375  913.1250 1430.5625 1004.4375  943.5625
#>  [911]  426.1250 1187.0625  730.5000 1461.0000 1521.8750  578.3125  243.5000
#>  [918]  700.0625  639.1875 1643.6250  608.7500  608.7500  852.2500  395.6875
#>  [925]  730.5000  365.2500 1187.0625  243.5000  273.9375  821.8125 1430.5625
#>  [932] 1613.1875 1339.2500 1461.0000 1643.6250  791.3750 1400.1250  700.0625
#>  [939]  456.5625  213.0625 1643.6250 1369.6875  517.4375 1217.5000  182.6250
#>  [946] 1613.1875 1156.6250 1613.1875 1643.6250  943.5625 1369.6875 1308.8125
#>  [953]  791.3750  791.3750 1704.5000  821.8125  974.0000 1278.3750  547.8750
#>  [960]  821.8125  395.6875  487.0000  821.8125  669.6250 1217.5000  243.5000
#>  [967] 1187.0625  882.6875 1095.7500 1065.3125  182.6250  639.1875  791.3750
#>  [974]  456.5625 1430.5625 1430.5625  395.6875  821.8125 1217.5000 1461.0000
#>  [981]  974.0000 1187.0625  365.2500 1156.6250  578.3125  456.5625 1004.4375
#>  [988]  669.6250 1400.1250 1400.1250  487.0000 1521.8750 1187.0625 1430.5625
#>  [995]  182.6250 1521.8750 1339.2500  669.6250  760.9375 1278.3750        NA
#> [1002]  517.4375 1065.3125  760.9375  669.6250 1613.1875  517.4375 1582.7500
#> [1009] 1278.3750 1156.6250 1613.1875 1704.5000 1765.3750  334.8125 1461.0000
#> [1016]  639.1875  791.3750 1582.7500 1308.8125  213.0625 1521.8750 1156.6250
#> [1023]  669.6250 1521.8750  700.0625 1613.1875 1613.1875 1156.6250  639.1875
#> [1030] 1004.4375  791.3750  334.8125 1765.3750 1217.5000  821.8125  913.1250
#> [1037] 1095.7500  395.6875  791.3750 1582.7500  669.6250  182.6250  213.0625
#> [1044]  791.3750  913.1250  943.5625 1065.3125  273.9375  943.5625 1187.0625
#> [1051]  730.5000  852.2500  456.5625 1156.6250 1278.3750  943.5625 1521.8750
#> [1058] 1339.2500 1704.5000  578.3125 1674.0625 1308.8125  487.0000  456.5625
#> [1065]  821.8125 1521.8750  608.7500 1187.0625  213.0625  700.0625  760.9375
#> [1072] 1521.8750  213.0625 1369.6875  852.2500  426.1250 1582.7500  608.7500
#> [1079] 1461.0000 1156.6250 1491.4375  821.8125 1217.5000  182.6250  365.2500
#> [1086] 1004.4375  791.3750  974.0000  334.8125  791.3750  852.2500 1004.4375
#> [1093]  882.6875  578.3125  578.3125 1461.0000  821.8125 1095.7500 1187.0625
#> [1100]  852.2500 1065.3125  608.7500 1491.4375  730.5000  730.5000  882.6875
#> [1107]  882.6875 1461.0000  487.0000  974.0000 1217.5000 1552.3125  974.0000
#> [1114] 1247.9375  456.5625  304.3750  608.7500 1582.7500 1704.5000  395.6875
#> [1121]  700.0625  304.3750  487.0000  791.3750 1582.7500 1095.7500 1582.7500
#> [1128]  608.7500 1491.4375  943.5625  395.6875  395.6875  426.1250 1765.3750
#> [1135]  700.0625 1308.8125  487.0000  913.1250  578.3125 1765.3750  547.8750
#> [1142]  334.8125 1734.9375  304.3750 1765.3750 1095.7500 1643.6250 1278.3750
#> [1149]  852.2500  882.6875  517.4375 1339.2500  395.6875 1400.1250 1734.9375
#> [1156]  791.3750 1095.7500 1613.1875 1613.1875  182.6250  639.1875  669.6250
#> [1163] 1126.1875 1095.7500  304.3750 1187.0625 1339.2500 1521.8750 1187.0625
#> [1170] 1461.0000  334.8125  395.6875 1004.4375 1034.8750  182.6250  821.8125
#> [1177]  243.5000  669.6250 1582.7500  456.5625 1400.1250 1491.4375  882.6875
#> [1184] 1217.5000  395.6875  243.5000  517.4375 1521.8750  852.2500 1461.0000
#> [1191]  852.2500  639.1875  821.8125  334.8125  608.7500  304.3750  821.8125
#> [1198]  395.6875 1156.6250  730.5000 1552.3125 1400.1250  456.5625  882.6875
#> [1205] 1400.1250 1461.0000 1461.0000 1674.0625 1674.0625 1643.6250  882.6875
#> [1212]  578.3125 1034.8750  243.5000 1247.9375 1461.0000  334.8125 1613.1875
#> [1219]  760.9375  730.5000 1156.6250  791.3750        NA  760.9375  395.6875
#> [1226]  517.4375 1400.1250 1217.5000  821.8125  852.2500  821.8125  426.1250
#> [1233] 1400.1250  182.6250 1369.6875 1247.9375 1247.9375 1734.9375  213.0625
#> [1240]        NA 1643.6250  304.3750  334.8125  700.0625  882.6875 1126.1875
#> [1247] 1552.3125  852.2500 1552.3125  456.5625 1369.6875 1795.8125  760.9375
#> [1254]  760.9375  304.3750 1704.5000 1461.0000 1521.8750 1339.2500 1582.7500
#> [1261]  273.9375  608.7500  760.9375 1095.7500  852.2500 1582.7500 1278.3750
#> [1268]  791.3750  821.8125 1034.8750 1095.7500 1278.3750 1004.4375  273.9375
#> [1275]  791.3750  547.8750 1674.0625 1217.5000  700.0625 1187.0625 1187.0625
#> [1282]  791.3750 1004.4375  334.8125  517.4375  608.7500 1004.4375 1034.8750
#> [1289] 1217.5000  578.3125  791.3750 1065.3125  974.0000  882.6875  669.6250
#> [1296]  517.4375 1339.2500  426.1250 1095.7500 1095.7500  882.6875  821.8125
#> [1303] 1278.3750 1521.8750  304.3750  304.3750 1095.7500  578.3125  517.4375
#> [1310] 1765.3750  243.5000 1187.0625  334.8125  791.3750 1369.6875  243.5000
#> [1317]  547.8750  456.5625 1156.6250 1582.7500  821.8125 1156.6250 1187.0625
#> [1324]  487.0000  487.0000 1643.6250 1004.4375 1156.6250 1521.8750 1734.9375
#> [1331]  821.8125 1521.8750  821.8125  639.1875 1369.6875 1734.9375  791.3750
#> [1338] 1582.7500  365.2500 1095.7500 1521.8750 1065.3125 1491.4375 1278.3750
#> [1345]  456.5625  213.0625 1065.3125 1095.7500 1187.0625 1765.3750 1674.0625
#> [1352] 1065.3125  760.9375 1187.0625 1765.3750  426.1250 1308.8125 1247.9375
#> [1359]  243.5000  608.7500  639.1875 1095.7500  426.1250 1461.0000  334.8125
#> [1366]  243.5000  608.7500  730.5000 1095.7500 1339.2500 1613.1875 1643.6250
#> [1373] 1491.4375  578.3125  943.5625 1521.8750 1795.8125 1400.1250 1491.4375
#> [1380]  852.2500 1430.5625 1552.3125 1004.4375  730.5000  334.8125  821.8125
#> [1387] 1126.1875  517.4375  913.1250 1613.1875  547.8750 1156.6250 1156.6250
#> [1394]  456.5625  913.1250  273.9375 1491.4375 1400.1250 1521.8750  273.9375
#> [1401]  426.1250  913.1250  213.0625 1521.8750 1643.6250 1430.5625 1156.6250
#> [1408] 1126.1875 1521.8750 1187.0625  243.5000  243.5000  974.0000  608.7500
#> [1415]  608.7500 1339.2500 1461.0000  639.1875 1004.4375 1674.0625  791.3750
#> [1422]  182.6250 1734.9375 1004.4375 1187.0625 1034.8750  182.6250 1095.7500
#> [1429] 1461.0000  730.5000  334.8125 1521.8750  395.6875  426.1250  517.4375
#> [1436] 1308.8125  182.6250 1582.7500  730.5000  760.9375 1491.4375 1187.0625
#> [1443]  730.5000 1217.5000  700.0625 1552.3125 1156.6250 1491.4375  304.3750
#> [1450] 1156.6250 1126.1875  365.2500  182.6250  456.5625 1156.6250  365.2500
#> [1457]  882.6875 1187.0625 1430.5625  791.3750  426.1250  213.0625 1521.8750
#> [1464] 1643.6250  395.6875 1430.5625 1247.9375  730.5000 1795.8125 1400.1250
#> [1471] 1156.6250 1156.6250  669.6250 1674.0625  456.5625 1461.0000 1217.5000
#> [1478] 1582.7500 1613.1875  639.1875  547.8750 1095.7500  395.6875  426.1250
#> [1485] 1065.3125  182.6250 1704.5000  821.8125  517.4375 1491.4375  882.6875
#> [1492] 1217.5000 1521.8750  913.1250  487.0000 1704.5000 1765.3750  304.3750
#> [1499] 1034.8750 1491.4375 1765.3750  547.8750  426.1250  487.0000 1674.0625
#> [1506]  426.1250 1095.7500  365.2500  700.0625 1278.3750  395.6875 1339.2500
#> [1513] 1126.1875 1734.9375 1552.3125 1613.1875  426.1250  304.3750  700.0625
#> [1520]  547.8750 1674.0625 1430.5625  913.1250 1217.5000 1126.1875 1095.7500
#> [1527]  913.1250 1461.0000  669.6250  395.6875 1461.0000 1795.8125  639.1875
#> [1534]  608.7500  456.5625  578.3125  395.6875 1065.3125  365.2500 1613.1875
#> [1541]  456.5625  974.0000 1430.5625 1308.8125  974.0000 1461.0000  791.3750
#> [1548] 1004.4375 1582.7500 1217.5000  395.6875 1491.4375 1430.5625 1521.8750
#> [1555]  213.0625 1400.1250  821.8125  608.7500  730.5000 1643.6250  547.8750
#> [1562]  730.5000 1187.0625 1004.4375  395.6875 1278.3750  213.0625 1217.5000
#> [1569]  730.5000  456.5625 1734.9375 1095.7500 1430.5625  943.5625 1126.1875
#> [1576] 1156.6250  821.8125 1613.1875 1308.8125  304.3750  578.3125  334.8125
#> [1583]  913.1250 1369.6875 1095.7500 1369.6875  487.0000  487.0000  669.6250
#> [1590] 1247.9375 1461.0000  578.3125  882.6875 1217.5000  700.0625 1126.1875
#> [1597] 1095.7500  791.3750 1552.3125 1187.0625 1065.3125 1552.3125 1217.5000
#> [1604]  608.7500  882.6875 1126.1875 1339.2500  760.9375  791.3750 1582.7500
#> [1611] 1034.8750 1308.8125  730.5000 1582.7500 1126.1875  456.5625 1156.6250
#> [1618]  821.8125  821.8125 1552.3125 1369.6875  821.8125 1217.5000  487.0000
#> [1625] 1308.8125  273.9375  852.2500 1004.4375  487.0000  334.8125  882.6875
#> [1632]  243.5000  852.2500 1400.1250 1613.1875 1126.1875  426.1250 1461.0000
#> [1639]  852.2500  791.3750  974.0000 1095.7500 1400.1250  426.1250  730.5000
#> [1646]  426.1250 1339.2500 1278.3750  395.6875 1674.0625 1704.5000 1582.7500
#> [1653] 1369.6875 1674.0625 1765.3750  456.5625 1065.3125 1065.3125 1704.5000
#> [1660]  456.5625  913.1250 1643.6250  517.4375  760.9375 1613.1875  578.3125
#> [1667] 1461.0000 1491.4375  334.8125  182.6250  395.6875  456.5625  760.9375
#> [1674]  882.6875  760.9375 1491.4375 1126.1875  974.0000 1278.3750 1217.5000
#> [1681] 1278.3750  426.1250  852.2500  304.3750 1369.6875 1156.6250 1704.5000
#> [1688]  791.3750  334.8125  943.5625 1034.8750  456.5625  791.3750 1339.2500
#> [1695]  395.6875  669.6250  882.6875 1004.4375  487.0000 1552.3125  547.8750
#> [1702]  669.6250  608.7500 1430.5625  -91.3125  487.0000 1034.8750  700.0625
#> [1709] 1004.4375  639.1875 1369.6875  791.3750  760.9375  578.3125 1247.9375
#> [1716]  639.1875  182.6250 1339.2500 1369.6875  517.4375 1521.8750  365.2500
#> [1723] 1521.8750 1430.5625  913.1250  334.8125 1491.4375 1308.8125 1034.8750
#> [1730] 1126.1875 1461.0000 1095.7500  213.0625  974.0000  547.8750 1339.2500
#> [1737]  791.3750  700.0625 1156.6250 1582.7500  821.8125 1187.0625  487.0000
#> [1744]  791.3750  213.0625  517.4375  334.8125  730.5000 1156.6250 1491.4375
#> [1751]  456.5625  669.6250  852.2500 1156.6250 1247.9375  821.8125 1491.4375
#> [1758] 1613.1875  426.1250  487.0000  882.6875 1491.4375 1095.7500  913.1250
#> [1765] 1521.8750  213.0625  974.0000  426.1250 1339.2500 1217.5000 1095.7500
#> [1772] 1095.7500  365.2500 1308.8125  608.7500 1461.0000  882.6875  243.5000
#> [1779] 1065.3125 1643.6250 1004.4375 1613.1875 1247.9375 1034.8750 1521.8750
#> [1786]  821.8125  760.9375  334.8125  426.1250 1095.7500 1613.1875 1521.8750
#> [1793]  395.6875  395.6875 1187.0625 1126.1875  852.2500  913.1250  213.0625
#> [1800] 1004.4375  273.9375  456.5625  243.5000  304.3750 1156.6250 1187.0625
#> [1807] 1187.0625 1004.4375  273.9375  639.1875  821.8125 1004.4375  943.5625
#> [1814] 1126.1875  943.5625  821.8125 1156.6250 1704.5000  365.2500 1765.3750
#> [1821]  882.6875  882.6875  882.6875  426.1250  547.8750 1156.6250 1339.2500
#> [1828]  882.6875 1308.8125  456.5625  243.5000  669.6250  974.0000  943.5625
#> [1835]  943.5625 1187.0625  304.3750  852.2500  943.5625 1734.9375  882.6875
#> [1842] 1034.8750 1461.0000  943.5625  365.2500 1613.1875 1247.9375 1400.1250
#> [1849]  760.9375 1430.5625 1095.7500  487.0000 1187.0625 1156.6250  365.2500
#> [1856]  456.5625  821.8125  791.3750  730.5000 1461.0000 1308.8125 1247.9375
#> [1863]  608.7500 1704.5000 1765.3750  243.5000 1065.3125 1491.4375 1552.3125
#> [1870] 1126.1875 1491.4375 1369.6875  669.6250 1034.8750 1004.4375  943.5625
#> [1877]  243.5000  760.9375  547.8750 1247.9375 1552.3125 1004.4375  304.3750
#> [1884]  852.2500 1278.3750  760.9375 1400.1250 1369.6875  517.4375 1400.1250
#> [1891] 1734.9375  639.1875 1339.2500  487.0000 1156.6250  578.3125 1156.6250
#> [1898] 1552.3125  943.5625  243.5000 1613.1875  273.9375 1400.1250  852.2500
#> [1905] 1156.6250 1461.0000  243.5000  608.7500 1065.3125 1613.1875  730.5000
#> [1912] 1765.3750 1704.5000 1156.6250 1461.0000  395.6875  791.3750 1095.7500
#> [1919]  608.7500 1369.6875  517.4375  304.3750 1156.6250 1126.1875  700.0625
#> [1926] 1430.5625  456.5625  882.6875 1126.1875  273.9375  700.0625  730.5000
#> [1933] 1095.7500 1126.1875  273.9375  821.8125  639.1875 1095.7500  517.4375
#> [1940] 1674.0625  547.8750  760.9375 1247.9375 1339.2500  578.3125 1674.0625
#> [1947] 1095.7500  639.1875  700.0625  365.2500  487.0000 1461.0000 1095.7500
#> [1954]  882.6875 1430.5625  974.0000  974.0000 1308.8125  456.5625 1552.3125
#> [1961]  700.0625 1339.2500  913.1250  852.2500  852.2500  182.6250  852.2500
#> [1968] 1552.3125 1582.7500  639.1875 1034.8750 1065.3125  760.9375 1461.0000
#> [1975] 1187.0625  578.3125  791.3750  426.1250 1217.5000 1765.3750 1521.8750
#> [1982] 1674.0625  426.1250 1247.9375 1308.8125  547.8750  487.0000  304.3750
#> [1989]  821.8125 1369.6875  821.8125 1095.7500 1552.3125  882.6875 1491.4375
#> [1996]  213.0625 1156.6250  669.6250 1369.6875 1156.6250 1613.1875  760.9375
#> [2003] 1095.7500  426.1250 1643.6250  730.5000  882.6875  213.0625  730.5000
#> [2010] 1217.5000 1095.7500 1126.1875  334.8125  213.0625  547.8750  304.3750
#> [2017]  730.5000 1339.2500 1521.8750 1187.0625  821.8125  821.8125  760.9375
#> [2024]  608.7500  669.6250 1521.8750  913.1250 1217.5000 1643.6250  882.6875
#> [2031]  334.8125  487.0000  426.1250 1126.1875 1521.8750 1704.5000 1704.5000
#> [2038]  608.7500 1461.0000  517.4375 1004.4375  243.5000  791.3750  365.2500
#> [2045]  365.2500  821.8125  974.0000  852.2500 1552.3125 1552.3125  974.0000
#> [2052] 1613.1875  760.9375 1430.5625  943.5625  182.6250 1095.7500 1430.5625
#> [2059] 1247.9375 1369.6875 1217.5000 1521.8750 1156.6250 1400.1250  760.9375
#> [2066]  821.8125 1613.1875 1339.2500  608.7500 1156.6250 1400.1250  273.9375
#> [2073]  791.3750  243.5000  821.8125  365.2500 1126.1875  487.0000  273.9375
#> [2080]  395.6875  608.7500 1674.0625  273.9375 1795.8125 1187.0625 1217.5000
#> [2087]  426.1250  730.5000 1126.1875  730.5000  821.8125 1156.6250 1004.4375
#> [2094] 1095.7500  974.0000  517.4375  334.8125  395.6875  578.3125 1278.3750
#> [2101]  730.5000 1461.0000  273.9375  943.5625 1613.1875 1461.0000 1400.1250
#> [2108]  273.9375  578.3125 1643.6250 1704.5000  487.0000  608.7500  791.3750
#> [2115]  760.9375 1156.6250  304.3750 1461.0000  243.5000  700.0625 1004.4375
#> [2122] 1400.1250  487.0000 1674.0625  730.5000 1004.4375  456.5625 1034.8750
#> [2129] 1034.8750  882.6875 1491.4375 1795.8125  426.1250  426.1250  821.8125
#> [2136]  487.0000  578.3125 1217.5000  700.0625  852.2500  608.7500  852.2500
#> [2143] 2039.3125 1065.3125 1582.7500  395.6875  243.5000 1095.7500  426.1250
#> [2150] 1552.3125 1217.5000 1126.1875 1674.0625 1187.0625 1704.5000 1795.8125
#> [2157] 1217.5000 1339.2500 1095.7500 1491.4375 1369.6875  821.8125  395.6875
#> [2164] 1400.1250 1217.5000  821.8125 1704.5000  182.6250  273.9375 1765.3750
#> [2171] 1734.9375  182.6250 1613.1875  365.2500  547.8750  426.1250  182.6250
#> [2178]  791.3750 1034.8750 1521.8750 1491.4375 1217.5000 1400.1250  821.8125
#> [2185]  243.5000  334.8125 1430.5625 1674.0625  791.3750 1217.5000  273.9375
#> [2192] 1156.6250  578.3125 1004.4375  426.1250 1126.1875  304.3750  700.0625
#> [2199] 1217.5000  426.1250 1308.8125  304.3750 1126.1875 1369.6875 1521.8750
#> [2206]  487.0000  852.2500 1491.4375  821.8125  426.1250  334.8125  304.3750
#> [2213] 1247.9375  334.8125  517.4375  395.6875  426.1250 1004.4375  456.5625
#> [2220] 1308.8125 1187.0625 1521.8750  517.4375 1491.4375 1004.4375  517.4375
#> [2227]  487.0000  273.9375 1430.5625  487.0000 1643.6250 1339.2500  273.9375
#> [2234]  974.0000  974.0000 1339.2500 1095.7500  821.8125  487.0000 1734.9375
#> [2241]  273.9375 1004.4375  487.0000 1552.3125  426.1250 1461.0000  334.8125
#> [2248]  943.5625 1461.0000 1552.3125 1674.0625 1095.7500  760.9375  760.9375
#> [2255] 1095.7500 1034.8750  791.3750  517.4375  243.5000 1461.0000 1095.7500
#> [2262]  608.7500 1461.0000  943.5625 1582.7500  608.7500  608.7500  852.2500
#> [2269] 1582.7500 1369.6875 1582.7500  213.0625  760.9375 1156.6250 1095.7500
#> [2276]  882.6875  700.0625 1004.4375  943.5625 1582.7500 1582.7500  273.9375
#> [2283] 1521.8750 1734.9375 1126.1875 1004.4375 1521.8750 1339.2500 1187.0625
#> [2290] 1065.3125 1065.3125 1430.5625  608.7500  700.0625  608.7500 1095.7500
#> [2297]  730.5000 1156.6250  517.4375  700.0625 1308.8125  487.0000 1095.7500
#> [2304]  821.8125 1187.0625  243.5000 1521.8750 1095.7500  182.6250  821.8125
#> [2311]  821.8125  791.3750  273.9375  882.6875 1339.2500  760.9375 1156.6250
#> [2318] 1339.2500  608.7500 1126.1875 1552.3125  791.3750 1369.6875  821.8125
#> [2325] 1582.7500 1552.3125 1247.9375  821.8125 1613.1875  487.0000 1217.5000
#> [2332]  273.9375  304.3750 1247.9375 1369.6875  700.0625  426.1250 1461.0000
#> [2339] 1187.0625 1734.9375  395.6875 1187.0625  456.5625 1430.5625 1734.9375
#> [2346] 1217.5000  182.6250 1795.8125  456.5625  608.7500  882.6875 1613.1875
#> [2353]  517.4375 1461.0000 1765.3750  365.2500 1582.7500 1034.8750  760.9375
#> [2360] 1521.8750  974.0000  273.9375 1034.8750 1369.6875 1065.3125  487.0000
#> [2367] 1430.5625 1369.6875  213.0625 1400.1250 1126.1875 1095.7500 1461.0000
#> [2374]  700.0625  426.1250 1582.7500 1034.8750  456.5625 1491.4375 1004.4375
#> [2381]  487.0000  669.6250 1521.8750  487.0000  456.5625  213.0625  517.4375
#> [2388] 1308.8125 1065.3125 1461.0000  213.0625  182.6250  456.5625  365.2500
#> [2395] 1217.5000  882.6875  578.3125 1034.8750  304.3750  243.5000 1643.6250
#> [2402] 1156.6250  974.0000 1552.3125 1491.4375        NA  456.5625  487.0000
#> [2409]  273.9375 1217.5000 1126.1875  273.9375 1704.5000  365.2500 1095.7500
#> [2416] 1004.4375 1126.1875  639.1875  547.8750 1552.3125  547.8750  456.5625
#> [2423] 1156.6250 1004.4375  700.0625  821.8125 1095.7500  334.8125  304.3750
#> [2430]  395.6875  213.0625  882.6875  182.6250 1765.3750 1126.1875  974.0000
#> [2437] 1613.1875  334.8125  304.3750 1552.3125  760.9375  487.0000  760.9375
#> [2444] 1430.5625  395.6875  456.5625  273.9375  273.9375 1339.2500 1187.0625
#> [2451]  760.9375  882.6875  487.0000 1217.5000  547.8750 1491.4375  608.7500
#> [2458]  791.3750  791.3750 1400.1250 1674.0625  273.9375 1734.9375 1095.7500
#> [2465] 1461.0000 1217.5000 1065.3125  852.2500  730.5000 1582.7500 1461.0000
#> [2472]  456.5625 1734.9375  334.8125 1156.6250 1369.6875  395.6875  487.0000
#> [2479]  426.1250 1430.5625  608.7500  578.3125 1065.3125  304.3750 1095.7500
#> [2486]  882.6875 1004.4375  852.2500 1095.7500 1187.0625  273.9375 1552.3125
#> [2493] 1217.5000  578.3125  913.1250 1247.9375 1521.8750  639.1875  456.5625
#> [2500] 1004.4375  700.0625 1582.7500  365.2500  700.0625 1521.8750 1034.8750
#> [2507] 1065.3125 1065.3125  913.1250 1430.5625 1491.4375 1156.6250  213.0625
#> [2514]  669.6250 1156.6250 1339.2500  760.9375 1065.3125 1278.3750 1095.7500
#> [2521]  365.2500 1582.7500 1795.8125 1217.5000 1734.9375  365.2500  426.1250
#> [2528]  791.3750  913.1250 1491.4375  913.1250 1643.6250  456.5625  243.5000
#> [2535] 1095.7500 1095.7500  852.2500  791.3750  243.5000  273.9375  243.5000
#> [2542]  669.6250 1187.0625 1369.6875 1369.6875  395.6875 1491.4375  700.0625
#> [2549] 1308.8125 1126.1875 1034.8750 1126.1875  913.1250  487.0000 1217.5000
#> [2556]  547.8750  578.3125 1704.5000 1461.0000  456.5625  456.5625 1004.4375
#> [2563] 1004.4375  608.7500  182.6250  821.8125  273.9375  882.6875  608.7500
#> [2570] 1095.7500 1491.4375 1552.3125  243.5000  213.0625 1095.7500 1552.3125
#> [2577]  365.2500 1461.0000 1156.6250  974.0000 1339.2500 1430.5625  487.0000
#> [2584] 1369.6875  852.2500 1461.0000 1095.7500  669.6250 1400.1250  456.5625
#> [2591] 1095.7500        NA  578.3125 1247.9375 1126.1875  487.0000  791.3750
#> [2598]  456.5625  608.7500 1095.7500  365.2500  760.9375  365.2500 1552.3125
#> [2605] 1552.3125 1065.3125  395.6875 1552.3125 1430.5625  426.1250  882.6875
#> [2612] 1369.6875  578.3125  456.5625  639.1875 1369.6875 1247.9375  760.9375
#> [2619] 1552.3125  213.0625  517.4375 1674.0625 1369.6875 1521.8750  487.0000
#> [2626] 1491.4375 1004.4375 1674.0625  730.5000 1004.4375  791.3750  669.6250
#> [2633] 1461.0000 1004.4375  791.3750  974.0000  760.9375  791.3750  213.0625
#> [2640] 1308.8125  365.2500 1339.2500 1065.3125 1765.3750  791.3750  700.0625
#> [2647] 1491.4375  730.5000  974.0000  213.0625 1674.0625 1491.4375  304.3750
#> [2654]  669.6250  669.6250 1217.5000  913.1250 1095.7500 1126.1875  669.6250
#> [2661]  273.9375  426.1250  852.2500 1430.5625 1004.4375 1765.3750  669.6250
#> [2668]  304.3750  426.1250  426.1250  304.3750  760.9375 1065.3125  456.5625
#> [2675] 1491.4375 1095.7500 1643.6250  700.0625  456.5625 1795.8125 1126.1875
#> [2682] 1734.9375  487.0000 1521.8750 1126.1875  639.1875        NA 1674.0625
#> [2689] 1126.1875  882.6875 1217.5000  243.5000  487.0000  487.0000  334.8125
#> [2696]  456.5625 1156.6250  730.5000  456.5625  304.3750 1247.9375  487.0000
#> [2703] 1613.1875  395.6875 1034.8750 1308.8125 1187.0625  547.8750  304.3750
#> [2710]        NA        NA        NA        NA 1065.3125        NA        NA
#> [2717]        NA        NA        NA        NA  700.0625 1126.1875  669.6250
#> [2724] 1674.0625  365.2500  456.5625 1156.6250 1430.5625 1400.1250  700.0625
#> [2731] 1795.8125  639.1875  852.2500 1034.8750 1156.6250  974.0000 1582.7500
#> [2738] 1430.5625  974.0000  273.9375 1461.0000  365.2500 1674.0625 1247.9375
#> [2745] 1095.7500  669.6250 1521.8750 1247.9375  273.9375 1521.8750 1126.1875
#> [2752] 1004.4375  791.3750  334.8125        NA 1339.2500  700.0625  791.3750
#> [2759] 1613.1875 1491.4375 1308.8125  913.1250  273.9375  852.2500 1126.1875
#> [2766]  760.9375 1461.0000 1491.4375  243.5000  791.3750  456.5625 1369.6875
#> [2773]  882.6875  821.8125  639.1875 1217.5000 1217.5000  213.0625  517.4375
#> [2780]  974.0000 1582.7500  639.1875  639.1875  547.8750 1582.7500  760.9375
#> [2787]  669.6250  669.6250  730.5000  487.0000  243.5000        NA 1430.5625
#> [2794] 1613.1875 1217.5000  669.6250 1613.1875 1004.4375  882.6875  882.6875
#> [2801]  487.0000 1034.8750  213.0625  852.2500  821.8125  852.2500 1430.5625
#> [2808]  669.6250  395.6875  760.9375 1521.8750  304.3750  578.3125  334.8125
#> [2815] 1217.5000 1095.7500  456.5625  487.0000 1187.0625 1369.6875  456.5625
#> [2822] 1095.7500 1795.8125  334.8125  517.4375  213.0625  913.1250 1521.8750
#> [2829]  395.6875 1034.8750  395.6875 1582.7500 1156.6250  882.6875 1765.3750
#> [2836] 1156.6250  669.6250  791.3750 1187.0625 1095.7500  974.0000  669.6250
#> [2843] 1278.3750 1461.0000  395.6875  456.5625  760.9375 1643.6250  882.6875
#> [2850] 1461.0000  882.6875  791.3750  273.9375 1156.6250 1491.4375  243.5000
#> [2857]  669.6250 1613.1875  639.1875  852.2500 1552.3125 1643.6250  821.8125
#> [2864] 1126.1875 1521.8750  243.5000  760.9375 1461.0000 1004.4375  852.2500
#> [2871] 1461.0000  456.5625  730.5000 1004.4375 1674.0625 1217.5000  426.1250
#> [2878]  273.9375 1004.4375  243.5000 1674.0625  426.1250  608.7500  334.8125
#> [2885] 1582.7500  760.9375 1400.1250 1795.8125 1491.4375 1126.1875  882.6875
#> [2892] 1491.4375 1582.7500 1369.6875  304.3750 1491.4375  760.9375  700.0625
#> [2899]  608.7500  243.5000 1400.1250 1126.1875 1278.3750  456.5625 1430.5625
#> [2906] 1187.0625  365.2500 1065.3125  456.5625 1217.5000 1187.0625  365.2500
#> [2913]  304.3750  243.5000 1521.8750  578.3125 1704.5000  639.1875 1461.0000
#> [2920] 1004.4375  974.0000 1400.1250 1582.7500  974.0000 1278.3750 1613.1875
#> [2927] 1521.8750 1704.5000 1521.8750 1582.7500 1795.8125 1400.1250 1552.3125
#> [2934] 1308.8125 1461.0000 1065.3125 1461.0000  730.5000 1430.5625  456.5625
#> [2941]  243.5000  669.6250 1400.1250  487.0000 1247.9375 1795.8125 1400.1250
#> [2948] 1004.4375        NA  669.6250 1126.1875 1582.7500 1156.6250  395.6875
#> [2955] 1582.7500  943.5625  578.3125 1187.0625  608.7500  547.8750 1095.7500
#> [2962] 1065.3125  791.3750 1278.3750  760.9375  700.0625  700.0625 1704.5000
#> [2969]  669.6250  304.3750 1795.8125  760.9375  426.1250 1187.0625        NA
#> [2976] 1430.5625  791.3750  882.6875 1552.3125 1004.4375  821.8125 1247.9375
#> [2983] 1430.5625  182.6250 1095.7500 1308.8125  243.5000 1095.7500 1552.3125
#> [2990]  913.1250 1491.4375 1095.7500  913.1250  608.7500  882.6875 1552.3125
#> [2997] 1247.9375  273.9375  517.4375  760.9375 1004.4375 1095.7500 1461.0000
#> [3004]  304.3750  517.4375  487.0000  730.5000 1278.3750 1613.1875        NA
#> [3011]  334.8125 1247.9375 1126.1875 1126.1875 1552.3125  365.2500 1156.6250
#> [3018]  395.6875  243.5000 1065.3125  487.0000 1582.7500  243.5000  243.5000
#> [3025] 1430.5625  974.0000        NA 1430.5625 1400.1250  852.2500  395.6875
#> [3032] 1552.3125 1552.3125 1582.7500 1126.1875  395.6875 1491.4375  821.8125
#> [3039] 1674.0625 1613.1875  456.5625  852.2500 1034.8750 1400.1250  974.0000
#> [3046]  974.0000 1674.0625 1704.5000  882.6875 1400.1250  700.0625 1156.6250
#> [3053] 1126.1875 1339.2500  334.8125  852.2500 1461.0000 1126.1875 1765.3750
#> [3060] 1187.0625  943.5625  760.9375  760.9375 1369.6875 1552.3125  760.9375
#> [3067] 1582.7500 1734.9375  882.6875  791.3750  669.6250  547.8750 1552.3125
#> [3074]  395.6875  913.1250 1461.0000 1095.7500 1034.8750 1004.4375 1400.1250
#> [3081] 1613.1875 1308.8125  365.2500  365.2500 1004.4375 1521.8750 1795.8125
#> [3088] 1582.7500  213.0625 1582.7500 1034.8750  608.7500  578.3125  273.9375
#> [3095]  913.1250  365.2500  365.2500 1217.5000 1704.5000  365.2500 1065.3125
#> [3102] 1613.1875 1095.7500  365.2500 1034.8750 1126.1875  517.4375  487.0000
#> [3109] 1126.1875  426.1250  213.0625 1491.4375  943.5625  578.3125  791.3750
#> [3116] 1308.8125  365.2500 1187.0625 1734.9375  730.5000 1491.4375  760.9375
#> [3123] 1795.8125 1156.6250 1521.8750  791.3750 1674.0625  852.2500  943.5625
#> [3130] 1278.3750  882.6875  730.5000 1126.1875  426.1250  730.5000  243.5000
#> [3137]  608.7500  730.5000  852.2500 1247.9375 1065.3125 1795.8125 1065.3125
#> [3144]  456.5625 1187.0625 1339.2500  791.3750  334.8125  395.6875 1308.8125
#> [3151]  913.1250  334.8125  487.0000  639.1875  426.1250  974.0000 1491.4375
#> [3158] 1187.0625 1004.4375 1065.3125 1582.7500  395.6875  821.8125  730.5000
#> [3165] 1065.3125 1521.8750  791.3750  365.2500 1491.4375 1795.8125 1004.4375
#> [3172]  487.0000  882.6875 1034.8750 1521.8750  852.2500 1308.8125 1765.3750
#> [3179]  882.6875  426.1250  760.9375  791.3750  426.1250  730.5000  395.6875
#> [3186] 1156.6250 1491.4375  426.1250 1247.9375  182.6250 1582.7500 1065.3125
#> [3193] 1552.3125 1156.6250  821.8125 1613.1875  913.1250  365.2500  700.0625
#> [3200] 1521.8750  700.0625  669.6250  700.0625 1734.9375 1582.7500  213.0625
#> [3207]  456.5625 1795.8125  974.0000 1643.6250 1065.3125  852.2500  213.0625
#> [3214]  547.8750 1156.6250  456.5625  821.8125  821.8125 2039.3125 1400.1250
#> [3221] 1308.8125 1247.9375 1126.1875  395.6875 1126.1875  243.5000  547.8750
#> [3228] 1126.1875  487.0000 1126.1875 1065.3125  395.6875  882.6875 1065.3125
#> [3235] 1704.5000  304.3750  821.8125  608.7500  974.0000  913.1250  395.6875
#> [3242] 1400.1250  273.9375 1369.6875  426.1250 1004.4375 1217.5000  517.4375
#> [3249] 1187.0625  700.0625 1521.8750  669.6250 1156.6250  913.1250